refactor(portal): root is now the login, linking to user home
