+++ apps/aurora-portal/src/client/routes/_auth/accounts/index.tsx
@@ -7,7 +7,7 @@ export const Route = createFileRoute("/_auth/accounts/")({
         to: `/projects`,
       })
     } else {
-      throw redirect({ to: "/auth/login" })
+      throw redirect({ to: "/" })
     }
   },
 })
