+++ apps/aurora-portal/src/locales/de/messages.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+0B+ue\":[\"Projekte\"],\"+9CXS9\":[\"Images deaktivieren\"],\"+Jcye3\":[\"Keyname\"],\"+Lt5cp\":[\"Sie sind nicht berechtigt, Projektzugriffe hinzuzufügen. Bitte melden Sie sich erneut an.\"],\"+NwLgN\":[\"Durch die Aktivierung dieses Images kann es wieder zum Starten neuer Instances verwendet werden.\"],\"+Nx1wc\":[\"Floating IPs konnten nicht geladen werden\"],\"+OEi73\":[\"Object Storage (Swift)\"],\"+YQ9qu\":[\"Container: \",[\"containerName\"]],\"+nQTmZ\":[\"Dieses Projekt hat keinen Zugriff auf den Flavor.\"],\"+p6nHr\":[\"Fehler beim Laden der Objekt-Metadaten: \",[\"metadataErrorMessage\"]],\"+zy2Nq\":[\"Typ\"],\"/1MfrG\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht heruntergeladen werden: \",[\"errorMessage\"]],\"/2a/eI\":[\"Flavor wird geladen...\"],\"/9Squ9\":[\"Sie haben keine Berechtigung, die Details dieses Flavors anzuzeigen.\"],\"/BZLRP\":[\"Um diese Aktion zu bestätigen, geben Sie das Wort <0>“detach”</0> in das Feld unten ein.\"],\"/EcdUM\":[\"Ihre Aktion ist erforderlich\"],\"/HgF9q\":[\"Sortieren nach\"],\"/InK0O\":[\"Gesamtspeicher\"],\"/LqWNN\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht gelöscht werden: \",[\"errorMessage\"]],\"/NeNjH\":[\"Eigenschaften des Containers \\\"\",[\"containerName\"],\"\\\" wurden erfolgreich aktualisiert.\"],\"/Nmxy/\":[\"Keine Keypaare verfügbar\"],\"/QIkBY\":[\"<0>Sicher & Zuverlässig:</0> Ihre Daten und Operationen sind mit Sicherheitsstandards auf Unternehmensebene und robuster Zuverlässigkeit geschützt.\"],\"/Qox3b\":[\"Ein Ordner mit diesem Namen existiert bereits\"],\"/Z2leb\":[\"No containers found.\"],\"/Z5n1b\":[\"Ordner erstellen unter:\"],\"/bUiYk\":[\"Router-ID\"],\"/eFtWI\":[\"RBAC-Richtlinien\"],\"/xnbdQ\":[\"Der angegebene Benutzer hat Zugriff. Ein Token für den Benutzer (auf ein beliebiges Projekt bezogen) muss in der Anfrage enthalten sein.\"],\"01/uUD\":[\"Segmente beibehalten (nur Manifest löschen)\"],\"07WXfc\":[\"Der Server hat ein unerwartetes Datenformat für die Metadata zurückgegeben.\"],\"0BSSYj\":[\"Ein Serverfehler ist aufgetreten, während der Projektzugriff entfernt wurde. Bitte versuchen Sie es später noch einmal.\"],\"0Gd0NU\":[\"Geteilt\"],\"0P2gFy\":[\"Die gesuchte Seite existiert nicht.\"],\"0WsqO0\":[\"Container geleert\"],\"0cVgUw\":[\"Filtern nach\"],\"0dZfcJ\":[[\"allContainersCount\"],\" Container\"],\"0eY8Mz\":[\"Für dieses Projekt sind keine Floating IPs verfügbar. Floating IPs ermöglichen es, öffentliche IP-Adressen Instances zuzuordnen.\"],\"0j9NA7\":[\"<0>Sind Sie sicher?</0> Alle Objekte im Container werden gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"0kCt7e\":[\"Die angegebenen Flavor-Daten sind ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"0kc0zi\":[\"Beim Löschen des Metadata ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"0o0OhW\":[\"No objects found.\"],\"0p+s6m\":[\"Typ: \",[\"typeValue\"],\", Code: \",[\"codeValue\"]],\"0u9jhd\":[\"Das Trennen dieser Floating IP hebt die Zuordnung zum aktuellen Port auf. Die Instance ist über diese Adresse nicht mehr erreichbar.\"],\"16085O\":[\"IP-Version\"],\"1H2g6v\":[\"Objekt wird verschoben...\"],\"1NS3nd\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" Container\"],\"other\":[\"#\",\" Container\"]}],\" erfolgreich geleert. \",[\"totalDeleted\",\"plural\",{\"one\":[\"#\",\" Objekt\"],\"other\":[\"#\",\" Objekte\"]}],\" insgesamt gelöscht.\"],\"1RwosK\":[\"Zielprojekt-ID ist erforderlich\"],\"1UzENP\":[\"Nein\"],\"1VDqZj\":[\"<0>Zukunftssicher:</0> Aurora ist darauf ausgelegt, sich mit den neuesten Trends in der Cloud-Technologie weiterzuentwickeln und stellt so sicher, dass Ihre Lösung immer auf dem neuesten Stand ist.\"],\"1iQtS2\":[\"Erste \",[\"actualObjectCount\"],\" von \",[\"total\"],\" Objekten werden angezeigt\"],\"1iUuTT\":[\"Ihre Sitzung ist abgelaufen. Bitte melden Sie sich erneut an.\"],\"1ojTVo\":[\"DNS-Domain auswählen.\"],\"1pGUZa\":[\"Sitzung läuft ab in\"],\"1pdLQw\":[\"Image nicht gefunden\"],\"1rLu3+\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht geleert werden: \",[\"errorMessage\"]],\"1rPB1p\":[\"Der Flavor oder das Projekt konnte nicht gefunden werden. Bitte überprüfen Sie, ob sie existieren.\"],\"1t/NnN\":[\"Ablehnen\"],\"1zZ1IK\":[\"Hallo\"],\"20E+79\":[\"Sie müssen sich anmelden, um auf diese Seite zuzugreifen.\"],\"20Kpaw\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" wurde erfolgreich gelöscht.\"],\"20axE5\":[\"Vom Projekt geteilt\"],\"23wBCX\":[\"Öffentlicher Lesezugriff\"],\"2G6hLq\":[\"Lösche \",[\"specKey\"]],\"2Inn83\":[\"Bulk-Upload von Archivdateien\"],\"2TtIL2\":[\"Gespeichert als X-Object-Meta-*-Header. Schlüssel sind nicht zwischen Groß- und Kleinschreibung unterscheidend.\"],\"2cJIlz\":[\"Floating Network-ID\"],\"2d/OiW\":[\"Geben Sie Ihren Benutzernamen ein\"],\"2dnZwV\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" konnte nicht gelöscht werden: \",[\"errorMessage\"]],\"2gH+i8\":[\"Sie sind nicht berechtigt, Flavors zu löschen. Bitte melden Sie sich erneut an.\"],\"2lq0gq\":[\"<0>Eigenschaften von</0> <1>\",[\"displayName\"],\"</1>\"],\"2mbisJ\":[\"Metadaten \\\"\",[\"trimmedKey\"],\"\\\" wurden erfolgreich hinzugefügt.\"],\"2pnrGl\":[\"Erwartetes Format: JJJJ-MM-TT HH:MM:SS\"],\"2q/Q7x\":[\"Sichtbarkeit\"],\"2ysnjX\":[\"<0>Erhöhte Produktivität:</0> Durch die Reduzierung der betrieblichen Komplexität hilft Aurora Ihrem Team, sich auf das Wesentliche zu konzentrieren – Innovationen voranzutreiben und Geschäftserfolg zu sichern.\"],\"2zceEg\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Das Image wird dauerhaft gelöscht.\"],\"33F2A+\":[\"Geben Sie den Container-Namen zur Bestätigung ein\"],\"3AUpb4\":[\"Alle löschen (\",[\"selectedCount\"],\")\"],\"3Qn0me\":[\"Mitglied hinzufügen\"],\"3n+vCm\":[\"Benutzerdefinierte Dauer (Minuten)\"],\"3nWqQW\":[\"Sie sind nicht berechtigt, Metadata anzusehen. Bitte melden Sie sich erneut an.\"],\"3nh/7E\":[\"Wenn diese Option aktiviert ist, wird dieser Flavor für alle Projekte verfügbar sein. Wenn sie deaktiviert ist, muss der Zugriff explizit für bestimmte Projekt gewährt werden.\"],\"3oChIh\":[\"<0>Vereinheitlichtes Cloud-Management:</0> Konsolidiert alle Ihre Cloud-Ressourcen in einer intuitiven Oberfläche.\"],\"3oc18/\":[\"Private Flavors konnten nicht geladen werden. Die angezeigte Liste ist möglicherweise unvollständig.\"],\"3q1GLx\":[\"Datei-Upload ausstehend...\"],\"3x7Sws\":[\"Security Group-Details werden geladen...\"],\"4/axPe\":[\"Schauen Sie sich unseren Quellcode auf GitHub an!\"],\"47eI0x\":[\"Die Beschreibung muss mindestens 1 Zeichen lang sein.\"],\"4EZrJN\":[\"Regeln\"],\"4O2AH3\":[\"Mitglied \\\"\",[\"memberIdToRemove\"],\"\\\" wurde erfolgreich entfernt.\"],\"4fh0Wj\":[\"Boot-Größe\"],\"4fvDRe\":[\"Zu aktivierende Images:\"],\"4fvcmm\":[\"Objekt wird hochgeladen als: <0>\",[\"selectedObjectName\"],\"</0>\"],\"4h3Eyf\":[\"\\\"\",[\"objectName\"],\"\\\" wurde erfolgreich hochgeladen.\"],\"4kjaAc\":[\"Keine Servergruppen verfügbar\"],\"4mbrAq\":[\"1 Minute\"],\"4opp4r\":[\"Security Groups\"],\"4pOfUd\":[\"Unsere Mission\"],\"4t33sh\":[\"Fehler beim Aktualisieren des Objekts\"],\"4uXhtt\":[\"CIDR\"],\"4utWB4\":[\"Serverrolle:\"],\"5/wyf8\":[\"Floating IP eingeben\"],\"56IxdF\":[\"Fehler beim Laden der Container-Objekte: \",[\"errorMessage\"]],\"5BLR6Q\":[\"IPv4\"],\"5Hdka/\":[[\"allContainersCount\"],\" Container\"],\"5JDSvn\":[\"Maximale Metadaten-Wertlänge\"],\"5M4Te3\":[\"DNS\"],\"5MF8U2\":[\"Fehler beim Aktualisieren des Containers\"],\"5W2i10\":[\"Verwalten Sie OpenStack mit \"],\"5Yrl6N\":[\"Lädt Servergruppen\"],\"5aNQ3F\":[\"\\\"\",[\"objectName\"],\"\\\" wurde erfolgreich nach \",[\"destination\"],\" kopiert.\"],\"5g7owI\":[\"Floating IP wird aktualisiert...\"],\"5y3O+A\":[\"Deaktivieren\"],\"6+7EwD\":[\"Objekte als Index bereitstellen, wenn Dateiname ist:\"],\"6+OdGi\":[\"Protokoll\"],\"6/xipy\":[\"Container-Format\"],\"644xgx\":[\"Geschützt\"],\"6BDqha\":[\"Limits\"],\"6CDYXS\":[\"Statisches Website-Serving\"],\"6GBt0m\":[\"Metadata\"],\"6H/Lg1\":[\"Dies ist ein öffentliches Image. Alle Benutzer haben Zugriff darauf. Eine explizite Freigabe ist nicht erforderlich.\"],\"6KRclz\":[\"Ordner erstellt\"],\"6Kjltl\":[\"Zugriffskontrolle für Container:\"],\"6OopEX\":[\"Container geleert\"],\"6Rnrsz\":[\"Zugriff verwalten - \",[\"flavorName\"]],\"6X/9Di\":[\"\\\"\",[\"objectName\"],\"\\\" wurde erfolgreich nach \",[\"destination\"],\" verschoben.\"],\"6YtxFj\":[\"Name\"],\"6jAi8c\":[\"Bereich\"],\"6luZQA\":[\"Objekt verschoben\"],\"6oolxV\":[\"Dieser Metadata-Key existiert bereits. Bitte verwenden Sie einen anderen Key.\"],\"6qzsuS\":[\"Schreib-ACLs\"],\"6sxz+g\":[\"Port-Name\"],\"6w+VnM\":[\"Container erstellt\"],\"6z9W13\":[\"Neustart\"],\"76RKuS\":[\"ICMP-Code\"],\"78+riR\":[\"Sie sind nicht berechtigt, Projektzugriffe zu entfernen. Bitte melden Sie sich erneut an.\"],\"7AfIPZ\":[\"Floating Network\"],\"7BpykL\":[\"Fehler beim Erstellen des Extra-Specs. Bitte versuchen Sie es erneut.\"],\"7L01XJ\":[\"Aktionen\"],\"7NC3vm\":[\"Subnet\"],\"7NSdfG\":[\"Container \",[\"progressCurrent\"],\" von \",[\"progressTotal\"],\" wird geleert, bitte warten...\"],\"7Q24LN\":[\"Richtlinie\"],\"7T1fHv\":[\"Fehler beim Entfernen des Mitglieds \\\"\",[\"memberIdToRemove\"],\"\\\"\"],\"7UlHhT\":[\"Metadaten \\\"\",[\"keyToDelete\"],\"\\\" wurden erfolgreich gelöscht.\"],\"7XQ3QJ\":[\"Abgelehnter Referrer: \",[\"host\"]],\"7ZnTL8\":[\"Fehler beim Aktualisieren des Objekts: \",[\"mutationErrorMessage\"]],\"7a4DvD\":[\"Keine Server verfügbar\"],\"7d1a0d\":[\"Öffentlich\"],\"7flw0l\":[\"Der Projektzugriff für \\\"\",[\"trimmedTenantId\"],\"\\\" wurde erfolgreich hinzugefügt.\"],\"7sMeHQ\":[\"Key\"],\"88kg0+\":[\"Erstellt am\"],\"8AriEH\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde erstellt\"],\"8TSI9h\":[\"Durch die Deaktivierung dieses Images kann es nicht mehr zum Starten neuer Instances verwendet werden. Bestehende Instances sind nicht betroffen.\"],\"8Tg/JR\":[\"Benutzerdefiniert\"],\"8ZOb7O\":[[\"numberDeleted\"],\" Objekt wurde dauerhaft gelöscht.\"],\"8ZsakT\":[\"Passwort\"],\"8c3/77\":[\"Maximale Metadaten-Namenslänge\"],\"8jLXs3\":[\"Versionierte Schreibvorgänge\"],\"8s0tOH\":[\"Sie haben keine Berechtigung, diesem Flavor Projektzugriffe hinzuzufügen.\"],\"8t1+HU\":[[\"successCount\"],\" Image(s) deaktiviert, aber \",[\"failedCount\"],\" Image(s) konnten nicht deaktiviert werden.\"],\"8wdCNd\":[\"tcp, udp, icmp oder Protokollnummer\"],\"8zAn1f\":[\"Fehler beim Löschen des Flavors. Bitte versuchen Sie es erneut.\"],\"98Fs4G\":[\"Image wird erstellt...\"],\"9J93Xr\":[\"Container-Name darf keine Schrägstriche enthalten\"],\"9SX0bO\":[\"Das Image \\\"\",[\"imageName\"],\"\\\" konnte nicht aktualisiert werden: \"],\"9X8lAk\":[\"Zuweisen\"],\"9doWrf\":[\"Fehler beim Hinzufügen des Mitglieds\"],\"9dsDHD\":[\"Das Image \\\"\",[\"imageId\"],\"\\\" konnte nicht reaktiviert werden: \",[\"message\"]],\"9iz2XW\":[\"Image kann nicht aktualisiert werden\"],\"9njIiV\":[\"Fehler beim Aktivieren der Images\"],\"9rz81C\":[\"Geräte-ID\"],\"9v5VLp\":[\"Keine benutzerdefinierten Eigenschaften definiert\"],\"9vSW3U\":[\"Rekursiv löschen\"],\"9x6EkK\":[\"Dies ist ein öffentlicher Flavor. Alle Projekte haben Zugriff darauf.\"],\"A7CVME\":[\"Wählen Sie zuerst das Festplattenformat\"],\"AB4Tnl\":[\"Bitte wählen Sie eine Datei zum Hochladen aus\"],\"AGXLLY\":[\"Image-Datei kann nicht hochgeladen werden\"],\"AJRhSM\":[\"Root Disk muss eine ganze Zahl ≥ 0 sein.\"],\"AN0DBJ\":[\"Enter drücken zum Hinzufügen\"],\"AX9Juz\":[\"Die ID darf nur alphanumerische Zeichen, Bindestriche, Unterstriche und Punkte enthalten.\"],\"AZyHwC\":[\"Muss eine gültige IPv4- oder IPv6-Adresse sein (z. B.: 172.24.4.228 oder 2001:db8::1).\"],\"Ac6dy9\":[\"Name eingeben\"],\"AdtLNV\":[\"Stellen Sie sicher, dass ACL-Einträge gültig sind — korrekte Projekt-IDs, Benutzer-IDs und Formate liegen in Ihrer Verantwortung. Ungültige Einträge können stillschweigend unbeabsichtigten Zugriff gewähren oder verweigern.\"],\"AeXO77\":[\"Account\"],\"Afh/Lb\":[\"Zielordner auswählen\"],\"AlbUVn\":[\"<0>Optimierte Skalierbarkeit:</0> Aurora ist für Unternehmen jeder Größe konzipiert und wächst mit Ihnen mit, unterstützt einfache Umgebungen und komplexe Multi-Cloud-Setups gleichermaßen.\"],\"Alx2/L\":[\"In neuem Tab öffnen\"],\"AuQtzx\":[\"Muss eine nicht-negative Ganzzahl sein\"],\"AxZkIr\":[\"Disk (GiB)\"],\"B2Czeb\":[\"Min. RAM\"],\"B2i9cQ\":[\"Zu löschende Objekte (\",[\"totalCount\"],\")\"],\"B3toQF\":[\"Objekte\"],\"B4Jzm7\":[\"Ceph\"],\"BCJPTn\":[\"Zugriff für alle Benutzer aus diesem Projekt gewähren.\"],\"BCXapL\":[\"Fehler beim Laden der Container-Eigenschaften: \",[\"errorMessage\"]],\"BJt+PJ\":[\"Fehler beim Löschen des Containers\"],\"BMTd81\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Das Zielprojekt verliert sofort den Zugriff auf diese Sicherheitsgruppe.\"],\"BOQYRn\":[\"Lädt Key Pairs...\"],\"BP4Fwj\":[\"Fehler beim Laden der Objekte: \",[\"errorMessage\"]],\"BSaBkZ\":[\"Objekte — \",[\"containerName\"]],\"BYH/2L\":[\"Image kann nicht deaktiviert werden\"],\"BZpsYm\":[\"Failed to load containers: \",[\"errorMessage\"]],\"BgMp/T\":[\"Ungültige Formatkombination für das ausgewählte Festplattenformat\"],\"BoIAP6\":[\"Die ID des Netzwerks, das der Floating IP zugeordnet ist.\"],\"BoPocW\":[\"MD5-Prüfsumme\"],\"BrrIs8\":[\"Storage\"],\"CA8ZeT\":[\"Sichtbarkeit des Images \\\"\",[\"imageName\"],\"\\\" auf \",[\"visibility\"],\" aktualisiert\"],\"CBFSfX\":[\"Bitte korrigieren Sie die Validierungsfehler unten.\"],\"CFMxC8\":[\"Images gelöscht\"],\"CMVP7y\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Die Regel wird dauerhaft gelöscht.\"],\"CgZxr7\":[\"Min. RAM (MB)\"],\"ChOuUj\":[\"Floating IP nicht gefunden\"],\"Cj2Gtd\":[\"Größe\"],\"ClGcRq\":[\"Container\"],\"Cu6xuZ\":[\"Dies ist ein <0>dynamisches großes Objekt (DLO)</0>-Manifest. Metadaten-Änderungen gelten nur für das Manifest — Segment-Objekte sind nicht betroffen.\"],\"CunRry\":[\"Ungültiges Projekt-ID-Format. Muss 32 hexadezimale Zeichen sein (z. B. b90f9c4bc76140e18540b2cec1299e2a) oder UUID-Format (z. B. 12345678-1234-1234-1234-123456789abc)\"],\"Cxgv2U\":[\"Min. Disk\"],\"D+Se/Q\":[\"Container leeren:\"],\"D/8vkD\":[\"Es wird in Ihrer Image-Liste angezeigt.\"],\"D1F789\":[\"<0>Sind Sie sicher?</0> Der Container wird gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"D3IRXw\":[\"Floating IP wird getrennt...\"],\"D7qT9F\":[\"Warum Aurora wählen?\"],\"DDRhQm\":[\"Ihre Sitzung ist abgelaufen.\"],\"DJT9tB\":[\"Account-Quotas\"],\"DKkOPx\":[\"Zusätzliche Spezifikationen\"],\"DNVql8\":[\"Vollständiges Lifecycle-Management von Floating IPs, einschließlich Zuweisung, Port-Zuordnung/-Aufhebung, DNS-Einstellungen und Löschung\"],\"DcMIiu\":[\"ACLs für Container \\\"\",[\"containerName\"],\"\\\" konnten nicht aktualisiert werden: \",[\"errorMessage\"]],\"Df0YHr\":[\"Security Group aktualisieren\"],\"Dh1qvV\":[\"Sie sind dabei, \",[\"deletableCount\"],\" Image(s) zu löschen. Diese Aktion kann nicht rückgängig gemacht werden.\"],\"Dia2Ue\":[\"Für diese Security Group sind keine RBAC-Richtlinien vorhanden\"],\"Do5/uH\":[\"Der Flavor oder das Projekt konnte nicht gefunden werden. Möglicherweise wurde es bereits entfernt.\"],\"Dqnh7K\":[\"Spezifischer Referrer: \",[\"host\"]],\"Dt5W9T\":[\"RBAC-Richtlinie entfernen\"],\"DvB4XF\":[\"Legen Sie Ihre Datei hier ab\"],\"E/QGRL\":[\"Deaktiviert\"],\"E4QYe7\":[\"Empfohlene Images\"],\"E6nRW7\":[\"URL kopieren\"],\"EF2EU9\":[\"Wird gelöscht...\"],\"EPMHs9\":[\"Sie haben keine Berechtigung, Flavors in diesem Projekt zu löschen.\"],\"EQnVgi\":[\"Der Flavor-Service ist für dieses Projekt nicht verfügbar.\"],\"EdQY6l\":[\"Keine\"],\"Ef7StM\":[\"Unknown\"],\"Enpdmy\":[\"Geben Sie <0>entfernen</0> zur Bestätigung ein:\"],\"EoKe5U\":[\"Domain\"],\"Eq5PsT\":[\"\\\"detach\\\" zur Bestätigung eingeben\"],\"EqSPkP\":[\"Lädt Flavors...\"],\"Erlvqg\":[\"Objektname darf keine führenden oder nachfolgenden Leerzeichen haben\"],\"ExLULX\":[\"Image-Name\"],\"EztMB8\":[\"Fehler beim Abrufen der Flavors vom Server.\"],\"F02e8I\":[\"Keine benutzerdefinierten Metadaten. Klicken Sie auf \\\"Eigenschaft hinzufügen\\\", um eine zu erstellen.\"],\"F6YIQe\":[\"Effiziente Massenlöschung\"],\"FKL6Jv\":[\"z. B. .r:*,.rlistings\"],\"FNcMGM\":[\"Creation Date\"],\"FORsv1\":[\"Finden Sie detaillierte Anleitungen und Referenzen in unserer Dokumentation.\"],\"FOcBn3\":[\"Trennen\"],\"FPsvA8\":[\"Verstanden!\"],\"FQBaXG\":[\"Aktivieren\"],\"FRtmJJ\":[\"Storage-Container nicht gefunden\"],\"FSbpS7\":[\"CPU\"],\"FjONW3\":[\"Fehler beim Laden des Flavors\"],\"FjPnAE\":[\"Fehler beim Laden der Security Group\"],\"Flugry\":[[\"progressPct\"],\"%\"],\"FwSyEp\":[\"Das angegebene Projekt existiert nicht oder Sie haben keine Berechtigung, es zu teilen.\"],\"Fzrzfe\":[\"Ordnername ist erforderlich\"],\"G6AP+o\":[\"Geteilt:\"],\"GEgjm+\":[\"Objekte werden geladen...\"],\"GPuCEo\":[\"Leer lassen für alle Typen\"],\"GSIPwA\":[\"Temporäre URL\"],\"GbKqnI\":[[\"successCount\"],\" Image(s) aktiviert, aber \",[\"failedCount\"],\" Image(s) konnten nicht aktiviert werden.\"],\"Gfx1qQ\":[\"Inhalt kann nicht geladen werden\"],\"GxkJXS\":[\"Wird hochgeladen...\"],\"Gyd3No\":[\"Kein spezifischer Projektzugriff für diesen privaten Flavor konfiguriert. Klicken Sie auf \\\"Projektzugriff hinzufügen\\\", um Zugriff zu gewähren.\"],\"H+a5j6\":[\"Freigeben\"],\"H0MuMF\":[\"Aurora bietet Ihnen robuste Tools, um Ihre Cloud-Verwaltung nahtlos zu optimieren. Erleben Sie eine mühelose Möglichkeit, Ihre Ressourcen in OpenStack-Umgebungen bereitzustellen, zu konfigurieren und zu skalieren.\"],\"H4Qwmp\":[\"Keine Objekte entsprechen Ihrer Suche. Versuchen Sie, Ihren Suchbegriff anzupassen.\"],\"H7u085\":[\"Noch kein Projekt hat Zugriff auf dieses Image. Klicken Sie auf \\\"Projektzugriff hinzufügen\\\", um Zugriff zu gewähren.\"],\"HAkrpK\":[\"Bei Aurora ist es unsere Mission, eine zentrale Plattform bereitzustellen, die das Cloud-Management vereinheitlicht. Wir streben danach, die Komplexitäten der Bereitstellung, Konfiguration und Skalierung von Ressourcen über verschiedene Cloud-Umgebungen hinweg zu vereinfachen und gleichzeitig ein nahtloses Wachstum für Ihr Unternehmen zu ermöglichen.\"],\"HBI6nY\":[\"Benötigen Sie Hilfe?\"],\"HBpi4q\":[\"Lädt Images...\"],\"HM56Bx\":[\"Creating...\"],\"HNlEFZ\":[\"löschen\"],\"HQH8HM\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht aktualisiert werden: \",[\"errorMessage\"]],\"HVdrr1\":[\"Beliebiger Referrer\"],\"HivZR9\":[\"Create Credential\"],\"Hivb/4\":[\"Der Server hat Probleme. Bitte versuchen Sie es später erneut.\"],\"Hiw1Ha\":[\"Keine Container gefunden\"],\"HlwgQN\":[\"Objekt \\\"\",[\"objectName\"],\"\\\" wurde dauerhaft gelöscht.\"],\"HuA8iQ\":[\"Floating IP wird zugewiesen...\"],\"HxTYrE\":[\"Der Flavor konnte nicht gefunden werden. Möglicherweise wurde er gelöscht.\"],\"I5kZVK\":[\"Entfernte Quelle\"],\"INUP6f\":[\"<0>Mühelose Ressourcenbereitstellung:</0> Stellen Sie Ressourcen wie Server, Netzwerke und Volumes schnell bereit, konfigurieren Sie diese und setzen Sie sie mit nur wenigen Klicks ein.\"],\"IOkHLC\":[\"Fehler beim Kopieren des Objekts: \",[\"errorMessage\"]],\"IUwGEM\":[\"Änderungen speichern\"],\"IWF68U\":[\"Speicherübersicht\"],\"IZ6Mh2\":[\"Geben Sie den Domain ein\"],\"IbYr/u\":[\"Content type\"],\"J4DKSM\":[\"Container-Format ist erforderlich\"],\"J6EOll\":[\"Objekt verschieben/umbenennen:\"],\"J7+bZb\":[\"Ordner gelöscht\"],\"J9QcnV\":[[\"successCount\"],\" von \",[\"totalCount\"],\" Image(s) erfolgreich aktiviert\"],\"J9cmxx\":[\"Fehler beim Aktualisieren der Sichtbarkeit auf \",[\"newVisibility\"]],\"JB0bhm\":[\"Mitmachen\"],\"JNGYAW\":[\"Container-Name ist erforderlich\"],\"JT3I1g\":[\"Flavor löschen\"],\"JeRXll\":[\"Dieser Schlüssel ist reserviert und wird separat verwaltet\"],\"JfWCsP\":[\"Teilweise erfolgreich deaktiviert\"],\"Jh4rAZ\":[\"Fehler beim Laden des Images\"],\"Jim5X9\":[\"Stateful\"],\"Jj0eys\":[\"Repository ansehen\"],\"JoECY1\":[\"Die bereitgestellten Metadata-Daten sind ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"JpZn1L\":[\"Bereits deaktiviert (wird übersprungen)\"],\"JrmKyf\":[\"Fehlgeschlagen: \",[\"errorDetails\"]],\"JtHgVz\":[\"Images löschen\"],\"K+e/0e\":[\"RAM (MiB)\"],\"K3bUTE\":[\"Minimale Festplatte muss 0 oder größer sein\"],\"K8Qnlj\":[\"Wird verschoben...\"],\"K9eC8x\":[\"Dies kann auf unzureichende Berechtigungen oder ein vorübergehendes Serviceproblem zurückzuführen sein. Bitte überprüfen Sie Ihre Zugriffsrechte oder aktualisieren Sie die Seite.\"],\"KDw4GX\":[\"Erneut versuchen\"],\"KJC+M7\":[\"Serverfehler beim Abrufen der Flavor-Details. Bitte versuchen Sie es später erneut.\"],\"KOpPMt\":[\"Gesamtspeicher-Quota\"],\"KSW/GC\":[\"Keine Flavors verfügbar. Filtern Sie neu oder erstellen Sie einen neuen Flavor.\"],\"KZN4Lc\":[\"Alle löschen\"],\"Km4AGG\":[\"Sicherheitsgruppe wird erstellt...\"],\"KoQP4F\":[\"Ein Serverfehler ist aufgetreten, während der Projektzugriff hinzugefügt wurde. Bitte versuchen Sie es später noch einmal.\"],\"KsIM0b\":[\"Boot-RAM\"],\"KsnZ3m\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" wurde erfolgreich erstellt.\"],\"KzUd7m\":[\"new-folder-name\"],\"LI8Z2I\":[[\"rowDisplayName\"],\" herunterladen\"],\"LK0pQN\":[\"Festplattenformat ist erforderlich\"],\"LMdsuJ\":[\"Port (von)\"],\"LQQCas\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" und \",[\"deletedCount\"],\" Objekt wurden dauerhaft gelöscht.\"],\"LiyMGv\":[\"Die Cloud betreten\"],\"Llcakz\":[\"Aktualisiert am\"],\"LqMb+g\":[\"Um diese Aktion zu bestätigen, geben Sie das Wort <0>\\\"release\\\"</0> in das Feld unten ein.\"],\"LtI9AS\":[\"Eigentümer\"],\"Lylr9Z\":[\"Objekt kopiert\"],\"M470oJ\":[\"Der Flavor konnte nicht gefunden werden oder hat keine Metadata.\"],\"M5Epeo\":[\"Image-Details bearbeiten\"],\"M5RhXF\":[\"Wird entfernt...\"],\"M5rEN5\":[\"Sitzung abgelaufen\"],\"M9H+/G\":[\"Projekte\"],\"MEIAzV\":[\"Unbenannt\"],\"MJtNLd\":[\"Zu löschende Images:\"],\"MOug+V\":[\"Geben Sie ein Tag ein und drücken Sie Enter oder klicken Sie auf Hinzufügen\"],\"MRB7nI\":[\"Richtung\"],\"MXoA/6\":[\"Objekt hochladen\"],\"MXw7Fr\":[\"Servername\"],\"MZGbkp\":[\"VCPUs\"],\"MbKJNP\":[\"Sie haben keine Berechtigung, auf die Zugriffsinformationen für diesen Flavor zuzugreifen.\"],\"MgZyuJ\":[\"Sie sind dabei, <0>\",[\"deactivatedCount\"],\" Image(s)</0> zu aktivieren. Aktivierte Images stehen zum Starten neuer Instances zur Verfügung.\"],\"MmtQVF\":[\"Ungültiger Wert für die Einstellung des öffentlichen Flavors.\"],\"Mt6sRo\":[\"Sie sind nicht berechtigt, auf die Zugriffsinformationen des Flavors zuzugreifen. Bitte melden Sie sich erneut an.\"],\"MtzSbv\":[\" Objektname ist erforderlich\"],\"MuKU9V\":[\"Failed to load objects: \",[\"errorMessage\"]],\"N2S1rs\":[\"Leeren\"],\"N5I2RJ\":[\"\\\"release\\\" zur Bestätigung eingeben\"],\"N5vGcw\":[\"Geben Sie Ihre Anmeldedaten ein, um auf Ihr Konto zuzugreifen.\"],\"NH2fsP\":[\"Bereits deaktiviert (wird übersprungen):\"],\"NOdFZR\":[\"Wird generiert...\"],\"NQU1Nn\":[\"Container-Name kopieren\"],\"NRMm0E\":[\"Dieses Projekt hat bereits Zugriff auf den Flavor.\"],\"NRP2uq\":[\"Objekt teilen:\"],\"NRVSdy\":[\"Mitglieds-ID\"],\"NW4PIb\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" konnte nicht erstellt werden: \",[\"errorMessage\"]],\"NZJhro\":[\"Objektname darf keine Schrägstriche enthalten\"],\"Nc7QKU\":[\"Feste IP-Adresse\"],\"NeUjqc\":[\"Dateilisting aktivieren\"],\"NixRmA\":[\"Min. Disk (GB)\"],\"NlcF/v\":[\"Kein Flavor für die Löschung ausgewählt.\"],\"NopYGU\":[\"Festplattenformat\"],\"Np28ib\":[\"oder per Drag & Drop\"],\"Nu4oKW\":[\"Beschreibung\"],\"Nvfd2b\":[\"Versionierung ist aktiviert\"],\"O80bQY\":[\"Objekt-Eigenschaften werden geladen...\"],\"O8tK4v\":[\"Regel hinzufügen\"],\"ONWvwQ\":[\"Hochladen\"],\"OR475H\":[\"Netzwerk\"],\"OSlLnz\":[\"Image-Sichtbarkeit\"],\"OYHzN1\":[\"Tags\"],\"OZImTR\":[\"Container-Listing-Limit\"],\"OaSktR\":[\"Geräteeigentümer\"],\"OlmKCg\":[\"Ein Flavor mit dieser ID oder diesem Namen existiert bereits. Bitte verwenden Sie andere Werte.\"],\"OvEjsP\":[\"Wird kopiert...\"],\"Ovofy+\":[\"Floating IP \",[\"floating_ip_address\"],\" freigeben\"],\"OxDN2m\":[\"Fehler beim Erstellen des Flavors. Bitte versuchen Sie es erneut.\"],\"OxaeYj\":[\"Wir entwickeln das Aurora-Dashboard, um Ihnen einen besseren Service zu bieten. Ihr Feedback ist von unschätzbarem Wert, um ein Werkzeug zu gestalten, das den einzigartigen Bedürfnissen von Unternehmen wie Ihrem gerecht wird. Bleiben Sie in Verbindung und begleiten Sie uns, während wir das Cloud-Management neu definieren.\"],\"Oxl1UN\":[\"Wenn keine Index-Datei vorhanden ist, zeigt die URL eine Liste der Objekte im Container an.\"],\"PAKSdy\":[\"Floating IP eingeben oder leer lassen für automatische Zuweisung\"],\"PEGvy+\":[\"Minimaler RAM muss 0 oder größer sein\"],\"PHt+EV\":[\"<0>delete</0> zur Bestätigung eingeben:\"],\"PIbPRX\":[\"RX/TX Faktor muss eine ganze Zahl ≥ 1 sein.\"],\"PLwzWR\":[\"Alle Container\"],\"PYQUjU\":[\"Metadaten-Konfiguration konnte nicht geladen werden.\"],\"PZnUbs\":[\"Bitte melden Sie sich erneut an, um fortzufahren.\"],\"PgNNGl\":[\"Weitere Aktionen\"],\"PiH3UR\":[\"Kopiert!\"],\"PiyQJ/\":[\"Keine Flavors gefunden\"],\"PkfPsB\":[\"Geben Sie die ID des Projekts ein, mit dem Sie diese Sicherheitsgruppe teilen möchten. Sie finden Projekt-IDs im Konto-/Projekt-Umschalter oder im Identity-Service.\"],\"Pkw7J9\":[\"Dieser Ordner ist leer.\"],\"PsEGri\":[\"Ubuntu 22.04 LTS\"],\"PtjzS+\":[\"Wird dem ausgewählten Port zugeordnet. Wenn der Port mehrere IPs hat, wählen Sie die gewünschte feste IP-Adresse.\"],\"PzgYM9\":[\"Prüfsumme\"],\"Q1W//7\":[\"Keine Dienste für dieses Projekt verfügbar.\"],\"Q2xmVl\":[\"Symlinks\"],\"Q9f2QF\":[[\"numberDeleted\"],\" Objekte wurden erfolgreich gelöscht, aber einige Löschvorgänge sind fehlgeschlagen.\"],\"QAUa4B\":[\"Geben Sie einen einzelnen Port ein oder definieren Sie einen Bereich, indem Sie auch \\\"Port (bis)\\\" ausfüllen. \\\"Port (bis)\\\" ist optional.\"],\"QEtDlS\":[\"Objekt wird kopiert...\"],\"QNHur0\":[\"Fehler beim Laden der Container-ACLs: \",[\"errorMessage\"]],\"QQ8wUG\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Der Flavor wird dauerhaft gelöscht.\"],\"QV1ZPO\":[\"Schlüssel ist erforderlich\"],\"QWdKwH\":[\"Verschieben\"],\"QYiqYb\":[\"Fehler beim Aktualisieren des Zugriffsstatus\"],\"Qb+14I\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Die Security Group wird dauerhaft gelöscht.\"],\"QetsXP\":[\"Upload fehlgeschlagen: \",[\"uploadError\"]],\"Qg4EG6\":[\"Die bereitgestellten Flavor-Daten sind ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"QuJSSl\":[\"Fehler beim Erstellen des Flavors. Bitte versuchen Sie es erneut.\"],\"QvqBQa\":[\"Ziel-Container\"],\"Qx7DM7\":[\"Capabilities\"],\"QxBGbh\":[\"Geschützt (wird übersprungen):\"],\"R0xWJL\":[\"Löschen nicht möglich. Container enthält Objekte. Bitte leeren Sie ihn zuerst.\"],\"R6kcsL\":[\"Muss eine gültige PQDN oder FQDN sein (nur alphanumerische Zeichen und Bindestriche, darf nicht mit einem Bindestrich beginnen oder enden).\"],\"R6u5CR\":[[\"failedCount\"],\" von \",[\"totalCount\"],\" Image(s) konnten nicht aktiviert werden. Einige Images sind möglicherweise bereits aktiv oder in einem ungültigen Zustand.\"],\"RByeNR\":[\"Ihre Sitzung ist abgelaufen. Bitte melden Sie sich erneut an.\"],\"RCr0yv\":[\"Flavor-Details konnten nicht geladen werden. Bitte versuchen Sie es erneut.\"],\"RFDYCD\":[\"Minimale Festplattengröße, die zum Booten dieses Images erforderlich ist\"],\"RGhYAo\":[\"RAM\"],\"RGrgxg\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht gelöscht werden: \",[\"errorMessage\"]],\"RGwfoL\":[\"Maximale Metadaten-Anzahl\"],\"RNBvdl\":[\"Maximale SLO-Segmente\"],\"RSFkXF\":[\"Image aktivieren\"],\"RSMPjT\":[\"Sie befinden sich derzeit auf der Dashboard-Route.\"],\"RSg/pq\":[\"Fehler beim Löschen des Objekts\"],\"RTQFAw\":[\"Sie sind nicht berechtigt, Metadata zu erstellen. Bitte melden Sie sich erneut an.\"],\"Rih53k\":[\"Maximale Container-Namenslänge\"],\"RkXlPZ\":[\"GitHub\"],\"Rlp5zj\":[\"Flavor erstellen\"],\"S0kLOH\":[\"ID\"],\"S1iTXO\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde gelöscht\"],\"S3olSf\":[\"Keine Metadata gefunden. Klicken Sie auf \\\"Metadata hinzufügen\\\", um eine zu erstellen.\"],\"S5CUKP\":[\"Mitglieds-ID (Projekt-UUID) ist erforderlich.\"],\"S63NbU\":[\"Das Image \\\"\",[\"imageId\"],\"\\\" konnte nicht deaktiviert werden: \",[\"message\"]],\"S8/j2h\":[\"Fehler beim Leeren des Containers\"],\"SBGiGm\":[\"Lese-ACLs\"],\"SCY5an\":[\"Fehler beim Verschieben des Objekts\"],\"SFo0kK\":[\"Alle Images\"],\"SIfYq6\":[\"Metadaten bearbeiten\"],\"SLEH7X\":[\"DNS-Name eingeben\"],\"STc+7E\":[\"Maximale Container pro Extraktion\"],\"SU0uxT\":[\"Objekt hochladen nach:\"],\"SUSS9i\":[\"Container-Name\"],\"SVLToM\":[\"\\\"remove\\\" zur Bestätigung eingeben\"],\"SXCuo2\":[\"Dokumentation ansehen\"],\"SZw9tS\":[\"Details ansehen\"],\"Sb/VT5\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich geleert. \",[\"deletedCount\"],\" Objekte gelöscht.\"],\"SfW/3r\":[\"Keine Gruppen vorhanden\"],\"Sgz1vJ\":[\"Mitglied \\\"\",[\"trimmedMemberId\"],\"\\\" wurde erfolgreich hinzugefügt.\"],\"Smk7M2\":[\"Fehler beim Laden der Floating-IP\"],\"SuX2Ca\":[\"Grundlegende Informationen\"],\"SysqAR\":[\"Flavor-Details\"],\"T6Gm5y\":[\"Externes Netzwerk auswählen\"],\"T7mgdd\":[[\"successCount\"],\" von \",[\"totalCount\"],\" Image(s) erfolgreich gelöscht\"],\"T8N6oi\":[\"Eigenschaftsschlüssel\"],\"T9Mtpi\":[\"Projekt ID\"],\"TM93nK\":[\"Sicherheitsgruppenregel löschen\"],\"TPMaxo\":[\"“release” zur Bestätigung eingeben\"],\"TQn3hH\":[\"Image konnte nicht erstellt werden. Bitte versuchen Sie es erneut.\"],\"TZJiVf\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich geleert. \",[\"deletedCount\"],\" Objekt gelöscht.\"],\"TfC9O+\":[\"Zuletzt geändert (UTC)\"],\"TfdeUd\":[\"Fehler beim Löschen des Extra-Specs. Bitte versuchen Sie es erneut.\"],\"TpGxnq\":[\"Mitglieds-ID eingeben\"],\"TvY/XA\":[\"Dokumentation\"],\"Tx4Ym+\":[\"Gültige PQDN oder FQDN (max. 63 Zeichen) eingeben, um sie der Floating IP zuzuordnen. A- und PTR-Einträge werden automatisch erstellt.\"],\"TyODHt\":[\"Metadata speichern\"],\"U/oahm\":[\"URL kopiert\"],\"U2wTy/\":[\"Hinweis: Das Attribut 'stateful' kann nicht geändert werden, wenn diese Security Group derzeit von einem oder mehreren Ports verwendet wird.\"],\"U4fmHG\":[\"Der Text muss “detach” in Kleinbuchstaben entsprechen.\"],\"U6L+P/\":[\"Inaktivitäts-Timeout\"],\"U9q4M7\":[\"Zurück zu Security Groups\"],\"UGhVPl\":[\"Objekttyp\"],\"UJVf0u\":[\"Image wird geladen...\"],\"UK2mpr\":[\"Temporäre URL wird generiert...\"],\"UKwOYH\":[\"Image-Datei\"],\"UO3hJ2\":[\"Temporäre URLs\"],\"UQ7Wyv\":[\"Zugriff für Image verwalten - \",[\"imageName\"]],\"URmyfc\":[\"Details\"],\"USiuNX\":[\"Container-Quotas\"],\"UVFHGY\":[\"z. B. PROJECT_ID:USER_ID\"],\"UVSFVV\":[\"Geteiltes Image ablehnen\"],\"UYSopm\":[\"Minimaler RAM (MB)\"],\"UbRKMZ\":[\"Ausstehend\"],\"UdcGJu\":[\"Images aktivieren\"],\"UiNv/G\":[\"S3 Object Storage requires EC2 credentials (access key + secret key) to authenticate your requests. You need to create credentials before accessing S3 resources.\"],\"Uj+n/2\":[\"Fehler beim Löschen des Ordners\"],\"UkVkoq\":[\"Leer lassen für alle Codes\"],\"UmQ3/m\":[\"Ausgewählte deaktivieren\"],\"Uwo8Xw\":[\"Dieses Image wurde am \",[\"sharedAt\"],\" von <0>\",[\"ownerProject\"],\"</0> mit Ihnen geteilt.\"],\"UztfYZ\":[\"Port zur Zuordnung auswählen\"],\"V/8B9A\":[\"Ich bestätige, dass alle vorhandenen Versionen ebenfalls gelöscht werden\"],\"V/SINY\":[\"Objekt aktualisieren\"],\"V1TzeS\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich gelöscht.\"],\"V66Jih\":[\"Zugriffsstatus\"],\"V7fN5X\":[\"Objekt kopieren:\"],\"V804LY\":[\"Security Group wird aktualisiert...\"],\"VCM3KS\":[\"Projektzugriff hinzufügen\"],\"VKmlZ+\":[\"Zu leerende Container (\",[\"totalCount\"],\")\"],\"VLI9eO\":[\"Floating IP-Details werden geladen...\"],\"VV1fdg\":[\"Jeder Benutzer hat Lesezugriff auf Objekte. Es ist kein Token in der Anfrage erforderlich.\"],\"VaA9mu\":[\"24 Stunden\"],\"VakxP/\":[\"Fehler beim Hochladen des Objekts\"],\"Vg0k6h\":[\"Zeige \",[\"filteredCount\"],\" von \",[\"totalCount\"],\" \",[\"itemName\"]],\"Vh/Uj5\":[\"Zielpfad\"],\"Vj8XFg\":[\"Fehler beim Erstellen des Containers\"],\"Vl4XTj\":[\"Ordnername darf keine Schrägstriche enthalten\"],\"Vmojta\":[\" Zugriffsstatus auf \\\"\",[\"newStatus\"],\"\\\" aktualisiert.\"],\"VoxR3s\":[\"Objekt wurde kopiert, konnte aber nicht von der Quelle gelöscht werden: \",[\"deleteErrorMessage\"]],\"Vz+7ZA\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht erstellt werden: \",[\"errorMessage\"]],\"Vzlopx\":[\"Container löschen:\"],\"W0MCSG\":[\"Zugriff auf Image akzeptieren\"],\"W5FkH9\":[\"Container-Name eingeben\"],\"W9PZE0\":[\"Objekte gelöscht\"],\"W9kfjU\":[\"QoS-Policy-ID\"],\"WCKEqI\":[\"Dies ist ein <0>statisches großes Objekt (SLO)</0>-Manifest. Metadaten-Änderungen gelten nur für das Manifest — Segment-Objekte sind nicht betroffen.\"],\"WCLyHI\":[\"Keine Floating IPs gefunden\"],\"WErCZy\":[\"Minimaler RAM, der zum Booten dieses Images erforderlich ist\"],\"WRZ3Mt\":[\"Container-Eigenschaften werden geladen...\"],\"WYb0Td\":[\"<0>Sind Sie sicher?</0> Alle Objekte in den ausgewählten Containern werden dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"WYiUDa\":[\"Container werden geladen...\"],\"Wbg1jv\":[[\"text\"],\" in die Zwischenablage kopieren\"],\"Wca9WC\":[\"Security Groups konnten nicht geladen werden\"],\"WlpcJv\":[\"DNS-Domain\"],\"WoSkGY\":[\"Entfernte IP\"],\"WrUky8\":[\"Teilen (Temporäre URL)\"],\"WyKwnD\":[\"Alte Objekt-Versionen speichern\"],\"WzVwU0\":[\"Zielprojekt-ID\"],\"X2OnDx\":[\"Ephemeral Disk (GiB)\"],\"X70LXS\":[[\"numberDeleted\"],\" Objekte wurden dauerhaft gelöscht.\"],\"XLk16/\":[\"Gemeinsam können wir das volle Potenzial Ihrer Cloud-Infrastruktur erschließen.\"],\"XYZLy9\":[\"Schlüssel enthält ungültige Zeichen\"],\"XvjC4F\":[\"Wird gespeichert...\"],\"XwxJJB\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich erstellt.\"],\"XxjLdW\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" Container war bereits leer.\"],\"other\":[\"#\",\" Container waren bereits leer.\"]}]],\"Y+2SDm\":[\"Security Group \\\"\",[\"securityGroupName\"],\"\\\" löschen\"],\"Y1YKad\":[\"Details bearbeiten\"],\"YIix5Y\":[\"Suchen...\"],\"YNgcgc\":[\"Flavor-Details werden geladen...\"],\"YRexkb\":[\"Objekt aktualisiert\"],\"YUU0QW\":[\"Flavor-ID ist erforderlich und darf nicht leer sein\"],\"YZmsaT\":[\"Teilweise erfolgreich aktiviert\"],\"YiMCKk\":[\"ACLs-Vorschau anzeigen\"],\"Yin3uB\":[\"Floating IP wird freigegeben...\"],\"YjAOtb\":[\"Security Group erstellen\"],\"YrAy/S\":[\"Sie haben keine Berechtigung, Metadata für diesen Flavor zu löschen.\"],\"YsOJlj\":[\"Ein Serverfehler ist aufgetreten, während die Zugriffsinformationen für den Flavor abgerufen wurden. Bitte versuchen Sie es später noch einmal.\"],\"YsrbQh\":[\"Eigentümer-Projekt-ID\"],\"YuC9dj\":[\"Zuordnen\"],\"YuGQWb\":[\"Regeltyp\"],\"Z/eWPC\":[\"Das Objekt wird zu diesem Pfad kopiert. Navigieren Sie oben durch die Ordner, um das Ziel zu ändern.\"],\"Z2fZGD\":[\"Kein Projekt ausgewählt\"],\"Z3FXyt\":[\"Lädt...\"],\"Z42tfY\":[\"Ordner im Object Storage sind virtuell — sie werden als Null-Byte-Platzhalterobjekte mit einem abschließenden Schrägstrich erstellt. Der Ordner wird nach der Erstellung angezeigt.\"],\"Z5r9vC\":[\"Teilweise erfolgreich gelöscht\"],\"Z8lGw6\":[\"Teilen\"],\"ZAx+d1\":[\"Maximale Gesamt-Metadatengröße\"],\"ZAy0zp\":[[\"successCount\"],\" von \",[\"totalCount\"],\" Image(s) erfolgreich deaktiviert\"],\"ZUmOzn\":[\"Der Server hat ein unerwartetes Datenformat für Flavor-Details zurückgegeben.\"],\"ZcWMT1\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde deaktiviert\"],\"Zgp2Sm\":[\"Keine Projekte verfügbar.\"],\"ZhVSpK\":[\"Fehler beim Entfernen des Mandantenzugriffs vom Flavor. Bitte versuchen Sie es erneut.\"],\"Zq6Y5u\":[\"Der DNS-Name darf höchstens 63 Zeichen lang sein.\"],\"ZvIpwi\":[\"Security Group auswählen...\"],\"Zw49f9\":[\"folder-name\"],\"Zw8Q49\":[\"Security Group nicht gefunden\"],\"a/nTb8\":[\"Image erstellen\"],\"a12lSo\":[\"Port (bis)\"],\"a13wDR\":[\"Zum Suchen von Containern eingeben...\"],\"a3LDKx\":[\"Sicherheit\"],\"a4A2uB\":[\"Account-Listing-Limit\"],\"a4N/Bg\":[\"Mehr laden\"],\"a7C4YS\":[\"Container aktualisiert\"],\"a88X3d\":[\"<0>Sind Sie sicher?</0> Objekt <1>\\\"\",[\"displayName\"],\"\\\"</1> wird dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"aG9OiI\":[\"Freigabe-Details\"],\"aI8Tgp\":[\"Eigentümer-Projekt-ID\"],\"aL1w5Z\":[\"Belegt\"],\"aOeFR+\":[\"Container leeren\"],\"aSsVD3\":[\"Öffentlicher Lesezugriff ist nicht aktiviert. Bevor Sie statisches Website-Serving konfigurieren, gehen Sie zu <0>Zugriff verwalten</0> und aktivieren Sie den öffentlichen Lesezugriff.\"],\"aTqCTq\":[\"Image-Datei ist erforderlich\"],\"aV6KPH\":[\"Versehentliches Löschen verhindern\"],\"aiqFbS\":[\"<0>Sind Sie sicher?</0> Die ausgewählten Objekte werden dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"an5hVd\":[\"Images\"],\"ao/ZJi\":[\"Ordner und alle Inhalte werden gelöscht...\"],\"aqagJH\":[\"Image-Sichtbarkeit kann nicht aktualisiert werden\"],\"arel2K\":[\"Keine Objekte gefunden\"],\"azXlY+\":[\"Zugriffsstatus:\"],\"b0uU1G\":[\"Alte Objekt-Versionen in Container speichern:\"],\"b2BLBa\":[\"Sicherheitsgruppenregel hinzufügen\"],\"bISG26\":[\"Fehler beim Abrufen der Flavor-Zugriffsinformationen. Bitte versuchen Sie es erneut.\"],\"bM1O3m\":[\"Image-Instanz\"],\"bQBMTH\":[\"Dies ist ein <0>statisches großes Objekt</0>. Standardmäßig werden auch alle zugehörigen Segment-Objekte dauerhaft gelöscht.\"],\"bRgFkJ\":[\"Fehler beim Hochladen der Datei \\\"\",[\"fileName\"],\"\\\": \"],\"bYRFNi\":[\"Fehler beim Löschen der Objekte\"],\"bc67JN\":[\"Benutzerdefinierte Eigenschaften / Metadaten\"],\"bmQLn5\":[\"Regel hinzufügen\"],\"bnql/K\":[\"Zurück zu Images\"],\"boJ+Y1\":[\"Ordner erstellen\"],\"boJlGf\":[\"Seite nicht gefunden\"],\"bpme7e\":[\"Flavor nicht gefunden\"],\"bwRvnp\":[\"Aktion\"],\"bwhBhT\":[\"Security Group\"],\"byKna+\":[\"Ein unerwarteter Fehler ist aufgetreten. Bitte versuchen Sie es erneut.\"],\"bzMKg7\":[\"Akzeptiert\"],\"bzSI52\":[\"Verwerfen\"],\"c+fUtV\":[\"Beginnen Sie mit der Eingabe, um nach einem Container zu suchen\"],\"c+xCSz\":[\"True\"],\"c1uL4p\":[\"Das Image \\\"\",[\"imageId\"],\"\\\" konnte nicht gelöscht werden: \",[\"message\"]],\"c6b6fz\":[\"Ausgewählte löschen\"],\"cCfxH1\":[\"Wird heruntergeladen...\"],\"cJDQIO\":[\"Root Disk\"],\"cPKL6O\":[\"Sie sind nicht berechtigt, Flavors zu erstellen. Bitte melden Sie sich erneut an.\"],\"cWbW6w\":[\"Zugriff verwalten\"],\"cXuXkb\":[\"Benutzer \",[\"userId\"],\" aus Projekt \",[\"projectId\"]],\"chL5IG\":[\"Community\"],\"cj17eo\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde aktiviert\"],\"cjEOmc\":[\"Teilen Sie diese Sicherheitsgruppe mit einem anderen Projekt. Das Zielprojekt kann diese Sicherheitsgruppe ansehen und verwenden, aber nicht ändern oder löschen.\"],\"cnGeoo\":[\"Löschen\"],\"cpw++p\":[\"Unterstützung für statische große Objekte\"],\"cqQyPB\":[\"Ordnername\"],\"d+F6q9\":[\"Erstellt\"],\"d+Ugpw\":[\"<0>Sind Sie sicher?</0> Ordner <1>\\\"\",[\"folderDisplayName\"],\"\\\"</1> und alle darin enthaltenen Objekte werden dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"d/I0J3\":[\"Ausgewählte aktivieren\"],\"d0pLfy\":[\"Security Group konnte nicht gelöscht werden\"],\"dEgA5A\":[\"Abbrechen\"],\"dFb5Nt\":[\"Id\"],\"dLFiER\":[\"Fehler beim Laden der Container: \",[\"errorMessage\"]],\"dOevLB\":[\"Läuft ab in \",[\"selectedPresetLabel\"],\" — um \",[\"expiresAtFormatted\"]],\"dPBJAJ\":[\"Alle leeren (\",[\"selectedCount\"],\")\"],\"dPj4yB\":[\"Bei Ihrem Konto anmelden\"],\"dPoCVe\":[\"“detach” zur Bestätigung eingeben\"],\"dTNzBI\":[\"Schlüssel muss mindestens ein alphanumerisches Zeichen enthalten\"],\"dVdc7N\":[\"Sie sind nicht berechtigt, Metadata zu löschen. Bitte melden Sie sich erneut an.\"],\"dd2ndz\":[\"Einträge in ACLs werden durch Kommas getrennt. Beispiele:\"],\"diFNkW\":[\"Fehler beim Laden der Komponente\"],\"e0NrBM\":[\"Projekt\"],\"eChIh7\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" wurde erfolgreich erstellt.\"],\"eGEHJE\":[\"DNS-Name\"],\"eKC+EC\":[\"-\"],\"ePK91l\":[\"Bearbeiten\"],\"eYlnXt\":[\"Keine Images gefunden\"],\"eh/k36\":[\"Regeltyp auswählen...\"],\"ekCRTP\":[\"Abgelehnt\"],\"eks7oA\":[\"Port-ID\"],\"eu70nA\":[\"Läuft ab um (UTC)\"],\"eyRsaH\":[\"Root\"],\"ezT9KW\":[\"Container-Synchronisierung\"],\"f+Uq1E\":[\"Neuer Objektname\"],\"f0cwjH\":[\"Objekt-Eigenschaften\"],\"fCPhho\":[\"Ein oder mehrere Objekte konnten nicht gelöscht werden:\"],\"fIvd7X\":[\"Fehler beim Löschen der Images\"],\"fJpv9x\":[\"Fehler beim Deaktivieren der Images\"],\"fj5byd\":[\"keine Angabe\"],\"fxnDd7\":[\"Fehler beim Generieren der temporären URL: \",[\"generalError\"]],\"fzfAAa\":[\"Ingress\"],\"g+Jead\":[\"IPv6\"],\"g1IxCo\":[\"RAM muss eine ganze Zahl ≥ 128 MB sein.\"],\"g3BSCe\":[\"Swap-Disk muss eine ganze Zahl ≥ 0 sein.\"],\"g3UF2V\":[\"Akzeptieren\"],\"g8Yxlg\":[\"Temporäre URL für \\\"\",[\"objectName\"],\"\\\" wurde in die Zwischenablage kopiert.\"],\"g9m7gK\":[\"ACL-Einträge steuern, wer von diesem Container lesen oder in ihn schreiben kann. Mehrere Einträge werden durch Kommas getrennt. Änderungen werden sofort nach dem Speichern wirksam.\"],\"gFKJBP\":[\"Ordnername darf keine führenden oder nachfolgenden Leerzeichen haben\"],\"gGdfWx\":[\"Der Compute-Dienst ist für dieses Projekt derzeit nicht verfügbar. Bitte versuchen Sie es später erneut.\"],\"gHTJc/\":[\"Object Storage\"],\"gMYsdZ\":[\"Auf \\\"Geteilt\\\" setzen\"],\"gU7JFm\":[\"Sicherheitsgruppenregel wird erstellt...\"],\"gYe+hC\":[\"Zum Hochladen klicken\"],\"go0J2x\":[\"Fehler beim Kopieren der temporären URL in die Zwischenablage\"],\"grs4+e\":[\"Compute-Übersicht\"],\"gztCjq\":[\"Die angegebene Projekt-ID ist ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"h3P8z+\":[\"Das Löschen des Flavor ist fehlgeschlagen. Bitte versuchen Sie es erneut\"],\"h47p9L\":[\"—\"],\"h8h6oz\":[\"Images deaktiviert\"],\"h99+4y\":[\"Floating IP zuweisen\"],\"hH3kDo\":[\"Image-Details werden geladen...\"],\"hHL/wm\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde aktualisiert\"],\"hLp49h\":[\"<0>\",[\"deleteWord\"],\"</0> zur Bestätigung eingeben:\"],\"hPz54a\":[\"Fehler beim Herunterladen\"],\"hQr1Cr\":[\"Image deaktivieren\"],\"hXUWyd\":[\"Wert ist erforderlich\"],\"hYgDIe\":[\"Erstellen\"],\"he3ygx\":[\"Kopieren\"],\"he4q+i\":[\"z. B. b90f9c4bc76140e18540b2cec1299e2a\"],\"hgpMHD\":[\"Gesamtspeicher\"],\"hkjZ7P\":[\"Fehler beim Aktualisieren der Sichtbarkeit für \\\"\",[\"imageName\"],\"\\\":\"],\"hrBow7\":[\"Netzwerk-ID\"],\"i0qMbr\":[\"Startseite\"],\"i30J2U\":[\"Keine Projekte gefunden\"],\"i41Xuw\":[\"Feste IP-Adresse auswählen\"],\"i5MEDc\":[\"Fehler beim Verschieben des Objekts: \",[\"copyErrorMessage\"]],\"i6/ygf\":[\"Eine Eigenschaft mit diesem Schlüssel existiert bereits\"],\"i9TIyi\":[\"Entfernte Sicherheitsgruppe\"],\"i9qiyR\":[\"Läuft ab in\"],\"iH8pgl\":[\"Zurück\"],\"igVDFt\":[\"Anhängen\"],\"iqUvrS\":[\"Benutzer C/D/I\"],\"izMhIO\":[\"Benutzer \",[\"userId\"],\" (beliebiges Projekt)\"],\"j9hkgJ\":[\"Regeln\"],\"jBIkmi\":[\"QCOW2, Raw, VMDK, VHD, VHDX, VDI, AMI, ARI, AKI, ISO, PLOOP\"],\"jIPNJG\":[\"Grundlegende Informationen\"],\"jK6wqe\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" und \",[\"deletedCount\"],\" Objekt wurden dauerhaft gelöscht.\"],\"jKopCP\":[\"Netzwerk & Routing\"],\"jMc/mo\":[\"Beim Erstellen von Metadata ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"jNzyQo\":[\"Objekt-Versionierung\"],\"jPxavx\":[\"Security Group konnte nicht aktualisiert werden\"],\"jS4B2+\":[\"Container-Name stimmt nicht überein\"],\"jSG7wx\":[\"Bitte geben Sie eine gültige Anzahl von Minuten größer als 0 ein\"],\"jhU93c\":[\"Das Image \\\"\",[\"imageName\"],\"\\\" konnte nicht erstellt werden: \"],\"js24f6\":[\"Ordner löschen:\"],\"jtnAf8\":[\"Geschützte Images (können nicht gelöscht werden)\"],\"jyqLKs\":[\"Dieses Mitglied hat bereits Zugriff auf dieses Image.\"],\"k0vAWv\":[\"Objekt-Anzahl-Quota\"],\"k5nYwm\":[\"vCPU\"],\"k7ENJG\":[[\"rowDisplayName\"],\" vorschauen\"],\"k99j0U\":[\"Upload abbrechen\"],\"kA2lMP\":[\"Externes Netzwerk\"],\"kCLnJG\":[\"Alle leeren\"],\"kGmM/p\":[\"Sie haben keine Berechtigung, Metadata für diesen Flavor zu erstellen.\"],\"kIuDMT\":[\"Konfigurieren Sie die Eingangs- und Ausgangsregeln, die steuern, welcher Datenverkehr für diese Sicherheitsgruppe erlaubt ist.\"],\"kKK8AH\":[\"Verfügbare Quota:\"],\"kQYfgO\":[\"Ein oder mehrere Container konnten nicht geleert werden: \",[\"errorMessage\"]],\"kRfwU+\":[\"Nichts zu tun. Container ist bereits leer.\"],\"kqJVBO\":[\"Sie haben keine Berechtigung, diese Sicherheitsgruppe zu teilen.\"],\"kuYWaD\":[\"Reservierte Schlüssel: web-index, web-listings, quota-count, quota-bytes\"],\"kzkYE6\":[\"Checking S3 credentials...\"],\"l75CjT\":[\"Ja\"],\"lAsm87\":[\"Dieses Image vor dem Löschen schützen\"],\"lBVhQs\":[\"Etwas ist schiefgelaufen.\"],\"lN/Z9n\":[\"Security Group-Name ist erforderlich\"],\"lN3xvy\":[\"Regel löschen\"],\"lQ3EIe\":[\"Maximale Löschvorgänge pro Anfrage\"],\"lWTy+Y\":[\"Image kann nicht erstellt werden\"],\"lWxDDh\":[\"Flavor Name\"],\"lZvIXd\":[\"Die Beschreibung darf höchstens 255 Zeichen lang sein.\"],\"lhIa6x\":[\"Fehler beim Laden der Extra-Specs. Bitte versuchen Sie es erneut.\"],\"lq/mBZ\":[\"Objekt-Informationen werden geladen...\"],\"lw1412\":[\"Sie wurden aufgrund von Inaktivität abgemeldet.\"],\"lxentK\":[\"Ein unerwarteter Fehler ist aufgetreten\"],\"m16xKo\":[\"Hinzufügen\"],\"m6X3ro\":[\"Gruppenname\"],\"mQSO1Y\":[\"Port Forwarding\"],\"mSLePW\":[\"Sie haben keine Berechtigung, auf Flavors für dieses Projekt zuzugreifen.\"],\"mSfwLL\":[\"Projekt-ID\"],\"mYnJeY\":[\"Der Text muss “release” in Kleinbuchstaben entsprechen.\"],\"mpby9d\":[\"Support kontaktieren\"],\"mqljvE\":[\"Metadaten kopieren\"],\"mvz5Eo\":[\"URL für öffentlichen Zugriff\"],\"mxPfpY\":[\"Neuen Ordner hier erstellen\"],\"mzI/c+\":[\"Herunterladen\"],\"n0ZttO\":[\"Root Disk (GiB)\"],\"n1ekoW\":[\"Einloggen\"],\"n1gB0L\":[\"Floating IP \",[\"floating_ip_address\"],\" bearbeiten\"],\"n22YIM\":[\"Beschreibung bearbeiten\"],\"n2IuBI\":[\"Eine temporäre URL gewährt zeitlich begrenzten Lesezugriff auf dieses Objekt ohne Authentifizierung. Jeder mit dem Link kann es herunterladen, bis es abläuft.\"],\"n3eQzA\":[\"Diese Eigenschaft ist reserviert und kann nicht geändert werden\"],\"n46oLW\":[\"Fehler beim Entfernen des Mitglieds\"],\"n9jJG6\":[\"Mitgliedszugriff entfernen\"],\"nETBrc\":[\"Egress\"],\"nLvo6K\":[\"Details der RBAC-Richtlinie:\"],\"nUuaq8\":[\"Fehler beim Aktualisieren des Containers: \",[\"errorMessage\"]],\"nW/hX9\":[\"Allgemeine Image-Daten\"],\"nZbdB+\":[\"Upload abgebrochen\"],\"ne/GWZ\":[\"Innerhalb eines Projekts werden Objekte in Containern gespeichert. Container sind der Ort, an dem Sie Zugriffsberechtigungen und Quotas definieren.\"],\"neiJm0\":[\"Flavors\"],\"nkpZyk\":[\"Container \\\"\",[\"containerName\"],\"\\\" war bereits leer.\"],\"nnxwBn\":[\"Es gibt keine Regeln für diese Sicherheitsgruppe\"],\"ntNlXu\":[\"Zugriff auflisten\"],\"o/VDOG\":[\"Image kann nicht gelöscht werden\"],\"o6M6l0\":[\"Security Group konnte nicht erstellt werden\"],\"oDkgME\":[\"Sie sind nicht berechtigt, Flavors zu erstellen. Bitte melden Sie sich erneut an.\"],\"oEGiW3\":[\"Wird hochgeladen... \",[\"progressPct\"],\"%\"],\"ocUvR+\":[\"False\"],\"odVI9Y\":[\"Container gelöscht\"],\"olfSYj\":[\"Zugriffskontrolle aktualisiert\"],\"onHi/J\":[\"Es wird aus Ihrer Image-Liste entfernt.\"],\"p4nMut\":[\"Swap (MiB)\"],\"p6CSHM\":[\"Objekte löschen\"],\"p7DzCB\":[\"Fehler beim Aktualisieren der Zugriffskontrolle\"],\"pFg+7w\":[\"Aktualisiert:\"],\"pOPvlj\":[\"Bereits aktiv (wird übersprungen)\"],\"pU25+T\":[\"Upload von \\\"\",[\"objectName\"],\"\\\" wurde abgebrochen.\"],\"pbzA+s\":[\"Optionale Beschreibung\"],\"pebLmQ\":[\"Zugriff für \",[\"memberIdDisplay\"],\" entfernen\"],\"plnnns\":[[\"successCount\"],\" Image(s) gelöscht, aber \",[\"failedCount\"],\" Image(s) konnten nicht gelöscht werden.\"],\"poCbZw\":[\"ACLs werden geladen...\"],\"podzPY\":[\"Projekt-ID\"],\"psPHye\":[\"Geteiltes Image akzeptieren\"],\"pubQie\":[\"Value eingeben\"],\"q0Rla3\":[\"Projektzugriff hinzufügen\"],\"q44uUq\":[\"Container teilweise geleert\"],\"q5sTNZ\":[\"<0>Kein Temp-URL-Schlüssel konfiguriert.</0> Ein temporärer URL-Schlüssel muss auf Account- oder Container-Ebene gesetzt werden, bevor temporäre URLs generiert werden können. Wenden Sie sich an Ihren Administrator, um <1>X-Account-Meta-Temp-URL-Key</1> oder <2>X-Container-Meta-Temp-URL-Key</2> zu konfigurieren.\"],\"q6K46F\":[\"Schlüssel existiert bereits\"],\"q88/6A\":[\"Fehler beim Erstellen des Ordners\"],\"qAkkjP\":[\"Maximale Objekt-Namenslänge\"],\"qEDO1j\":[\"Dies ist ein <0>dynamisches großes Objekt</0>. Nur das Manifest wird gelöscht — die zugehörigen Segment-Objekte (gespeichert unter dem Manifest-Präfix) werden <1>nicht</1> automatisch entfernt und müssen separat gelöscht werden.\"],\"qFDA8L\":[\"Zugriff auf Image ablehnen\"],\"qJb6G2\":[\"Erneut versuchen\"],\"qQ1QBh\":[\"Hardware-Spezifikationen\"],\"qST5TS\":[\"Fehler – Image-Details\"],\"qUlxA+\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" wurde dauerhaft gelöscht.\"],\"qaAo9Y\":[\"Es ist ein Serverfehler beim Erstellen des Flavors aufgetreten. Bitte versuchen Sie es später erneut.\"],\"qh5W8q\":[\"Richtlinie entfernen\"],\"qs+BrU\":[\"Sie haben keine Berechtigung, Projektzugriffe von diesem Flavor zu entfernen.\"],\"qtoOYG\":[\"Kein Limit\"],\"quU9wK\":[[\"failedCount\"],\" von \",[\"totalCount\"],\" Image(s) konnten nicht deaktiviert werden. Einige Images sind möglicherweise bereits deaktiviert oder in einem ungültigen Zustand.\"],\"qvF2D8\":[\"Keine Images verfügbar. Filtern Sie neu oder erstellen Sie ein neues Image.\"],\"qxxo7y\":[\"Keine Richtlinien entsprechen Ihrer Suche\"],\"qyNaF7\":[\"Geben Sie einen Zeitstempel wie \\\"YYYY-MM-DD HH:mm:ss\\\" ein, um die automatische Löschung zu planen\"],\"qzIZOL\":[\"Ungültiges Dateiformat. Unterstützte Formate: \",[\"supportedFileFormats\"]],\"qzhUb9\":[\"Erste \",[\"maxOptions\"],\" von \",[\"totalCount\"],\" werden angezeigt — verfeinern Sie Ihre Suche, um die Ergebnisse einzugrenzen\"],\"r5SQFW\":[\"Container-Name muss \",[\"maxContainerNameLength\"],\" Zeichen oder weniger haben\"],\"r9Aac8\":[\"Ephemeral Disk\"],\"rAtQcX\":[\"Sie können das Objekt umbenennen, indem Sie den Namen hier ändern.\"],\"rD9yV1\":[\"Zu deaktivierende Images:\"],\"rIe0oV\":[\"Fehler beim Hinzufügen des Zugriffs für das Project zum Flavor. Bitte versuchen Sie es erneut.\"],\"rIi6x4\":[\"Der Flavor konnte nicht gefunden werden. Möglicherweise wurde er bereits gelöscht.\"],\"rJe6vw\":[\"7 Tage\"],\"rbuO5A\":[\"Diese Sicherheitsgruppe ist bereits mit dem angegebenen Projekt geteilt.\"],\"rcBt6T\":[\"Failed to create credential: \",[\"errorMessage\"]],\"rdUucN\":[\"Vorschau\"],\"rhaNn7\":[\"Container werden geladen...\"],\"riR9oD\":[\"Hinweis: Für <0>statische und dynamische große Objekte</0> werden nur die Manifeste gelöscht — ihre Segmente außerhalb dieses Ordnerpräfixes sind nicht betroffen.\"],\"rlgAtt\":[\"Das Objekt wird zu diesem Pfad verschoben. Navigieren Sie oben durch die Ordner, um das Ziel zu ändern.\"],\"rp0Bd0\":[\"Compute\"],\"rrjuul\":[\"Weitere Details finden Sie in der <0>Dokumentation</0>.\"],\"rvXsSb\":[\"Der Projektzugriff für \\\"\",[\"tenantIdToRemove\"],\"\\\" wurde erfolgreich entfernt.\"],\"rwBVXS\":[\"Zu löschende Images (\",[\"deletableCount\"],\")\"],\"ryf/ee\":[\"Images aktiviert\"],\"ryxYVo\":[\"Zu deaktivierende Images (\",[\"activeCount\"],\")\"],\"s/s1lz\":[\"Jeder Benutzer kann eine HEAD- oder GET-Operation auf dem Container ausführen, sofern der Benutzer auch Lesezugriff auf Objekte hat. Es ist kein Token erforderlich.\"],\"s2ubkU\":[\"Flavor ID\"],\"s4Vnq2\":[\"Wird geleert...\"],\"sNVNmf\":[\"MAC-Adresse\"],\"sPFHpI\":[\"Disk\"],\"sSNyf3\":[\"Willkommen beim <0>Aurora-Dashboard</0>, Ihrer Cloud-Management-Lösung der nächsten Generation. Wir sind bestrebt, die Art und Weise, wie Sie mit Ihrer Cloud-Infrastruktur interagieren und diese verwalten, zu vereinfachen. Mit Effizienz, Skalierbarkeit und Benutzerfreundlichkeit im Kern konzipiert, ermöglicht Aurora Ihnen, Prozesse zu optimieren und das volle Potenzial Ihrer Cloud-Ressourcen auszuschöpfen.\"],\"sWBLli\":[\"Eigenschaft hinzufügen\"],\"sXd+qS\":[\"Eigenschaften von \\\"\",[\"objectName\"],\"\\\" wurden erfolgreich aktualisiert.\"],\"sa4CV6\":[\"Alle Benutzer aus Projekt \",[\"projectId\"]],\"shKIZu\":[\"Zu aktivierende Images (\",[\"deactivatedCount\"],\")\"],\"sheDTJ\":[\"Bitte beachten Sie: Für <0>dynamische</0> und <1>statische große Objekte</1> werden nur die Manifeste gelöscht. Die zugehörigen Segmente werden nicht gelöscht.\"],\"sihD20\":[\"Images werden geladen...\"],\"sjMCOP\":[\"Zuletzt geändert\"],\"slWh5C\":[\"Floating IP \",[\"floating_ip_address\"],\" einem Port zuordnen\"],\"sxbP3b\":[\"Objekt-Anzahl\"],\"t/YqKh\":[\"Entfernen\"],\"t0X9+8\":[\"Container-Name\"],\"t1POAD\":[\"Keine benutzerdefinierten Metadaten-Eigenschaften gefunden. Klicken Sie auf \\\"Eigenschaft hinzufügen\\\", um eine zu erstellen.\"],\"t1fq6V\":[\"Der Server hat ein unerwartetes Datenformat zurückgegeben.\"],\"t7ff15\":[\"gültiges Token erforderlich: false\"],\"t95VRV\":[\"Über das Aurora-Dashboard\"],\"tASa/P\":[\"Beim Löschen des Flavors ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"tIrNgH\":[\"Beim Abrufen der Metadata ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"tLerHy\":[\"Ephemeral Disk muss eine ganze Zahl ≥ 0 sein.\"],\"tM5SEI\":[\"ACLs für Container \\\"\",[\"containerName\"],\"\\\" wurden erfolgreich aktualisiert.\"],\"tOkmLM\":[\"Fehler beim Kopieren des Objekts\"],\"tV/Ozb\":[\"Port-Bereich\"],\"tVSmFT\":[\"Weitere werden geladen...\"],\"tX5yOZ\":[\"Neuer Ordner\"],\"tasfos\":[\"entfernen\"],\"tbwGSx\":[\"Minimale Festplatte (GB)\"],\"tejJLY\":[\"Floating IP wird zugeordnet...\"],\"tfAKBU\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht hochgeladen werden: \",[\"errorMessage\"]],\"tfDRzk\":[\"Speichern\"],\"tfxu04\":[\"Zugriff entfernen für \",[\"tenantId\"]],\"thHAVL\":[\"Akzeptierte Images\"],\"tiflqy\":[\"Image kann nicht reaktiviert werden\"],\"tlfxPP\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht kopiert werden: \",[\"errorMessage\"]],\"tmpGvt\":[\"production, linux\"],\"u+VWhB\":[\"In die Zwischenablage kopiert!\"],\"u2xIeO\":[\"Fehler beim Aktualisieren der ACLs: \",[\"errorMessage\"]],\"u5HztT\":[\"RX/TX Factor\"],\"u77/s4\":[\"Floating IPs\"],\"u7En0V\":[\"Metadata hinzufügen\"],\"uAI0yI\":[\"Objekt löschen:\"],\"uAQUqI\":[\"Status\"],\"uLtFAr\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht aktualisiert werden: \",[\"errorMessage\"]],\"uSdnuQ\":[\"VCPUs müssen eine ganze Zahl ≥ 1 sein.\"],\"ujK/QN\":[\"Objekte werden geladen...\"],\"uly9ET\":[\"Regeldetails:\"],\"up0ZSW\":[\"Fingerprint\"],\"uuKb0T\":[\"Die Beschreibung muss weniger als 65535 Zeichen haben.\"],\"v0hPHE\":[\"Details anzeigen\"],\"v3djpU\":[\"Verschieben/Umbenennen\"],\"v9Dn8m\":[\"Das Aurora-Dashboard ist mehr als nur ein Werkzeug – es ist Ihr Partner bei der Navigation in der Cloud. Egal, ob Sie ein kleines Startup oder ein globales Unternehmen sind, Aurora bietet Ihnen die Flexibilität, Leistung und Einfachheit, die Sie benötigen, um Ihre Ziele zu erreichen.\"],\"vBUQNE\":[\"Das Metadata konnte nicht gefunden werden. Möglicherweise wurde es bereits gelöscht.\"],\"vEkTR9\":[\"Quota\"],\"vH2C/2\":[\"Swap\"],\"vLYu4Z\":[\"Nichts zu tun. Container scheint leer zu sein — die Objekt-Anzahl ist möglicherweise aufgrund einer kürzlichen Operation noch nicht synchronisiert.\"],\"vR4HmN\":[\"Lädt Instanzen...\"],\"vTh35P\":[\"Container erstellen\"],\"vXmL4D\":[\"Legen Sie Ihre Image-Datei hier ab\"],\"vZUKSz\":[\"Floating IP \",[\"floating_ip_address\"],\" trennen\"],\"vbajgL\":[\"Öffentlicher Flavor\"],\"vcQSZh\":[\"Dieser Ordner ist leer — verwenden Sie \\\"Neuer Ordner\\\", um einen zu erstellen.\"],\"vcXmqy\":[\"Netzwerk-Übersicht\"],\"vcvCXq\":[\"Fehler – Flavor-Details\"],\"vg84cD\":[[\"allCount\"],\" Elemente\"],\"vmRPFm\":[\"Sicherheitsgruppe teilen\"],\"vmYyLY\":[\"Entferntes IP-Präfix\"],\"vp5vfW\":[\"1 Stunde\"],\"vpt8cE\":[\"URL generieren\"],\"vrPCbw\":[\"Image-ID\"],\"w3bAcf\":[\"Diese Aktion ist dauerhaft. Die Adresse wird aus Ihrem Projekt entfernt und dem öffentlichen Pool zurückgegeben. Dies kann nicht rückgängig gemacht werden.\"],\"w9+8d7\":[\"Projektzugriff entfernen\"],\"wEfZld\":[\"Neuen Flavor erstellen\"],\"wFaT8w\":[\"Fehler beim Leeren der Container\"],\"wMHvYH\":[\"Value\"],\"wPrtGF\":[\"Key eingeben\"],\"wTg+FY\":[\"Maximale Dateigröße\"],\"wXxPjv\":[\"S3 Object Storage — Setup Required\"],\"wa1Bcq\":[\"Project ID eingeben\"],\"wbqM4L\":[[\"customMinutes\"],\" Minuten\"],\"wcUecy\":[\"Sie haben keine Berechtigung, Metadata für diesen Flavor anzusehen.\"],\"we28Pq\":[\"ACLs-Vorschau ausblenden\"],\"wlQNTg\":[\"Members\"],\"wlUDbB\":[\"Zuletzt aktualisiert: \",[\"formattedDate\"]],\"wrXcuy\":[\"Objektname\"],\"wrk/xj\":[\"Image-Details\"],\"wyIOMP\":[\"Image-Name ist erforderlich\"],\"wzqqS+\":[\"Hauptmerkmale\"],\"x/XQrD\":[\"Beliebiger Dateityp\"],\"x1bK0h\":[\"Mit den aktuellen Suchkriterien sind keine Container verfügbar. Versuchen Sie, Ihren Suchbegriff anzupassen.\"],\"x3T4pq\":[\"Die Container-Metadaten melden Objekte, aber keine wurden aufgelistet. Dies kann eine vorübergehende Synchronisierungsverzögerung sein — bitte warten Sie einen Moment und versuchen Sie es erneut.\"],\"x5l/TK\":[\"Bereits aktiv (wird übersprungen):\"],\"x9AdZ8\":[\"property_key\"],\"xNG/3n\":[\"Floating IP-Adresse\"],\"xNZKYy\":[[\"failedCount\"],\" von \",[\"totalCount\"],\" Image(s) konnten nicht gelöscht werden. Einige Images sind möglicherweise geschützt oder werden verwendet.\"],\"xqhyRT\":[\"Objekt hochgeladen\"],\"xw2UtT\":[\"Neues Image erstellen\"],\"y+KBOY\":[\"z. B. production, linux, ubuntu\"],\"y02Bu1\":[\"Container:\"],\"y0u86k\":[\"Der angeforderte Flavor konnte nicht gefunden werden. Er wurde möglicherweise gelöscht oder Sie haben keinen Zugriff darauf.\"],\"y1GYnY\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht verschoben werden: \",[\"errorMessage\"]],\"yPWFWy\":[\"ICMP-Typ\"],\"yTtJTy\":[\"Image-Metadaten bearbeiten\"],\"yYxB17\":[\"Alle Filter löschen\"],\"ylfbpz\":[\"Der Key für die Extra-Spec ist erforderlich und darf nicht leer sein.\"],\"yp0UjB\":[\"Ethertype\"],\"yqPflB\":[\"... und \",[\"hiddenCount\"],\" weitere\"],\"yu9G3x\":[\"Security Group bearbeiten\"],\"yz7wBu\":[\"Schließen\"],\"z+zpLP\":[\"gültiges Token erforderlich: true\"],\"z1JceR\":[\"Zurück zu Floating IPs\"],\"z45o5B\":[\"Objekt-Anzahl\"],\"z9NAjZ\":[\"Objekt gelöscht\"],\"zCD96i\":[\"Sie sind nicht berechtigt, Flavor-Details anzuzeigen. Bitte melden Sie sich erneut an.\"],\"zDS0JC\":[\"Der Name muss 2-50 Zeichen lang sein.\"],\"zWb/Nn\":[\"Maximale Header-Größe\"],\"zc5dcw\":[\"Anmeldung fehlgeschlagen. Bitte überprüfen Sie Ihre Anmeldeinformationen und versuchen Sie es erneut.\"],\"zga9sT\":[\"OK\"],\"zhM8FP\":[\"Zugriff für einen Benutzer aus einem anderen Projekt gewähren.\"],\"zm7+/D\":[\"Sie sind dabei, <0>\",[\"activeCount\"],\" Image(s)</0> zu deaktivieren. Deaktivierte Images können nicht zum Starten neuer Instances verwendet werden.\"],\"zwBp5t\":[\"Privat\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+0B+ue\":[\"Projekte\"],\"+9CXS9\":[\"Images deaktivieren\"],\"+Jcye3\":[\"Keyname\"],\"+Lt5cp\":[\"Sie sind nicht berechtigt, Projektzugriffe hinzuzufügen. Bitte melden Sie sich erneut an.\"],\"+NwLgN\":[\"Durch die Aktivierung dieses Images kann es wieder zum Starten neuer Instances verwendet werden.\"],\"+Nx1wc\":[\"Floating IPs konnten nicht geladen werden\"],\"+OEi73\":[\"Object Storage (Swift)\"],\"+YQ9qu\":[\"Container: \",[\"containerName\"]],\"+nQTmZ\":[\"Dieses Projekt hat keinen Zugriff auf den Flavor.\"],\"+p6nHr\":[\"Fehler beim Laden der Objekt-Metadaten: \",[\"metadataErrorMessage\"]],\"+zy2Nq\":[\"Typ\"],\"/1MfrG\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht heruntergeladen werden: \",[\"errorMessage\"]],\"/2a/eI\":[\"Flavor wird geladen...\"],\"/9Squ9\":[\"Sie haben keine Berechtigung, die Details dieses Flavors anzuzeigen.\"],\"/BZLRP\":[\"Um diese Aktion zu bestätigen, geben Sie das Wort <0>“detach”</0> in das Feld unten ein.\"],\"/EcdUM\":[\"Ihre Aktion ist erforderlich\"],\"/HgF9q\":[\"Sortieren nach\"],\"/InK0O\":[\"Gesamtspeicher\"],\"/LqWNN\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht gelöscht werden: \",[\"errorMessage\"]],\"/NeNjH\":[\"Eigenschaften des Containers \\\"\",[\"containerName\"],\"\\\" wurden erfolgreich aktualisiert.\"],\"/Nmxy/\":[\"Keine Keypaare verfügbar\"],\"/QIkBY\":[\"<0>Sicher & Zuverlässig:</0> Ihre Daten und Operationen sind mit Sicherheitsstandards auf Unternehmensebene und robuster Zuverlässigkeit geschützt.\"],\"/Qox3b\":[\"Ein Ordner mit diesem Namen existiert bereits\"],\"/Z2leb\":[\"No containers found.\"],\"/Z5n1b\":[\"Ordner erstellen unter:\"],\"/bUiYk\":[\"Router-ID\"],\"/eFtWI\":[\"RBAC-Richtlinien\"],\"/xnbdQ\":[\"Der angegebene Benutzer hat Zugriff. Ein Token für den Benutzer (auf ein beliebiges Projekt bezogen) muss in der Anfrage enthalten sein.\"],\"01/uUD\":[\"Segmente beibehalten (nur Manifest löschen)\"],\"07WXfc\":[\"Der Server hat ein unerwartetes Datenformat für die Metadata zurückgegeben.\"],\"0BSSYj\":[\"Ein Serverfehler ist aufgetreten, während der Projektzugriff entfernt wurde. Bitte versuchen Sie es später noch einmal.\"],\"0Gd0NU\":[\"Geteilt\"],\"0P2gFy\":[\"Die gesuchte Seite existiert nicht.\"],\"0WsqO0\":[\"Container geleert\"],\"0cVgUw\":[\"Filtern nach\"],\"0dZfcJ\":[[\"allContainersCount\"],\" Container\"],\"0eY8Mz\":[\"Für dieses Projekt sind keine Floating IPs verfügbar. Floating IPs ermöglichen es, öffentliche IP-Adressen Instances zuzuordnen.\"],\"0j9NA7\":[\"<0>Sind Sie sicher?</0> Alle Objekte im Container werden gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"0kCt7e\":[\"Die angegebenen Flavor-Daten sind ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"0kc0zi\":[\"Beim Löschen des Metadata ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"0o0OhW\":[\"No objects found.\"],\"0p+s6m\":[\"Typ: \",[\"typeValue\"],\", Code: \",[\"codeValue\"]],\"0u9jhd\":[\"Das Trennen dieser Floating IP hebt die Zuordnung zum aktuellen Port auf. Die Instance ist über diese Adresse nicht mehr erreichbar.\"],\"16085O\":[\"IP-Version\"],\"1H2g6v\":[\"Objekt wird verschoben...\"],\"1NS3nd\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" Container\"],\"other\":[\"#\",\" Container\"]}],\" erfolgreich geleert. \",[\"totalDeleted\",\"plural\",{\"one\":[\"#\",\" Objekt\"],\"other\":[\"#\",\" Objekte\"]}],\" insgesamt gelöscht.\"],\"1RwosK\":[\"Zielprojekt-ID ist erforderlich\"],\"1UzENP\":[\"Nein\"],\"1VDqZj\":[\"<0>Zukunftssicher:</0> Aurora ist darauf ausgelegt, sich mit den neuesten Trends in der Cloud-Technologie weiterzuentwickeln und stellt so sicher, dass Ihre Lösung immer auf dem neuesten Stand ist.\"],\"1iQtS2\":[\"Erste \",[\"actualObjectCount\"],\" von \",[\"total\"],\" Objekten werden angezeigt\"],\"1iUuTT\":[\"Ihre Sitzung ist abgelaufen. Bitte melden Sie sich erneut an.\"],\"1ojTVo\":[\"DNS-Domain auswählen.\"],\"1pGUZa\":[\"Sitzung läuft ab in\"],\"1pdLQw\":[\"Image nicht gefunden\"],\"1rLu3+\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht geleert werden: \",[\"errorMessage\"]],\"1rPB1p\":[\"Der Flavor oder das Projekt konnte nicht gefunden werden. Bitte überprüfen Sie, ob sie existieren.\"],\"1t/NnN\":[\"Ablehnen\"],\"1zZ1IK\":[\"Hallo\"],\"20E+79\":[\"Sie müssen sich anmelden, um auf diese Seite zuzugreifen.\"],\"20Kpaw\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" wurde erfolgreich gelöscht.\"],\"20axE5\":[\"Vom Projekt geteilt\"],\"23wBCX\":[\"Öffentlicher Lesezugriff\"],\"2G6hLq\":[\"Lösche \",[\"specKey\"]],\"2Inn83\":[\"Bulk-Upload von Archivdateien\"],\"2TtIL2\":[\"Gespeichert als X-Object-Meta-*-Header. Schlüssel sind nicht zwischen Groß- und Kleinschreibung unterscheidend.\"],\"2cJIlz\":[\"Floating Network-ID\"],\"2d/OiW\":[\"Geben Sie Ihren Benutzernamen ein\"],\"2dnZwV\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" konnte nicht gelöscht werden: \",[\"errorMessage\"]],\"2gH+i8\":[\"Sie sind nicht berechtigt, Flavors zu löschen. Bitte melden Sie sich erneut an.\"],\"2lq0gq\":[\"<0>Eigenschaften von</0> <1>\",[\"displayName\"],\"</1>\"],\"2mbisJ\":[\"Metadaten \\\"\",[\"trimmedKey\"],\"\\\" wurden erfolgreich hinzugefügt.\"],\"2pnrGl\":[\"Erwartetes Format: JJJJ-MM-TT HH:MM:SS\"],\"2q/Q7x\":[\"Sichtbarkeit\"],\"2ysnjX\":[\"<0>Erhöhte Produktivität:</0> Durch die Reduzierung der betrieblichen Komplexität hilft Aurora Ihrem Team, sich auf das Wesentliche zu konzentrieren – Innovationen voranzutreiben und Geschäftserfolg zu sichern.\"],\"2zceEg\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Das Image wird dauerhaft gelöscht.\"],\"33F2A+\":[\"Geben Sie den Container-Namen zur Bestätigung ein\"],\"3AUpb4\":[\"Alle löschen (\",[\"selectedCount\"],\")\"],\"3Qn0me\":[\"Mitglied hinzufügen\"],\"3n+vCm\":[\"Benutzerdefinierte Dauer (Minuten)\"],\"3nWqQW\":[\"Sie sind nicht berechtigt, Metadata anzusehen. Bitte melden Sie sich erneut an.\"],\"3nh/7E\":[\"Wenn diese Option aktiviert ist, wird dieser Flavor für alle Projekte verfügbar sein. Wenn sie deaktiviert ist, muss der Zugriff explizit für bestimmte Projekt gewährt werden.\"],\"3oChIh\":[\"<0>Vereinheitlichtes Cloud-Management:</0> Konsolidiert alle Ihre Cloud-Ressourcen in einer intuitiven Oberfläche.\"],\"3oc18/\":[\"Private Flavors konnten nicht geladen werden. Die angezeigte Liste ist möglicherweise unvollständig.\"],\"3q1GLx\":[\"Datei-Upload ausstehend...\"],\"3x7Sws\":[\"Security Group-Details werden geladen...\"],\"47eI0x\":[\"Die Beschreibung muss mindestens 1 Zeichen lang sein.\"],\"4EZrJN\":[\"Regeln\"],\"4O2AH3\":[\"Mitglied \\\"\",[\"memberIdToRemove\"],\"\\\" wurde erfolgreich entfernt.\"],\"4fh0Wj\":[\"Boot-Größe\"],\"4fvDRe\":[\"Zu aktivierende Images:\"],\"4fvcmm\":[\"Objekt wird hochgeladen als: <0>\",[\"selectedObjectName\"],\"</0>\"],\"4h3Eyf\":[\"\\\"\",[\"objectName\"],\"\\\" wurde erfolgreich hochgeladen.\"],\"4kjaAc\":[\"Keine Servergruppen verfügbar\"],\"4mbrAq\":[\"1 Minute\"],\"4opp4r\":[\"Security Groups\"],\"4pOfUd\":[\"Unsere Mission\"],\"4t33sh\":[\"Fehler beim Aktualisieren des Objekts\"],\"4uXhtt\":[\"CIDR\"],\"4utWB4\":[\"Serverrolle:\"],\"5/wyf8\":[\"Floating IP eingeben\"],\"56IxdF\":[\"Fehler beim Laden der Container-Objekte: \",[\"errorMessage\"]],\"5BLR6Q\":[\"IPv4\"],\"5Hdka/\":[[\"allContainersCount\"],\" Container\"],\"5JDSvn\":[\"Maximale Metadaten-Wertlänge\"],\"5M4Te3\":[\"DNS\"],\"5MF8U2\":[\"Fehler beim Aktualisieren des Containers\"],\"5Yrl6N\":[\"Lädt Servergruppen\"],\"5aNQ3F\":[\"\\\"\",[\"objectName\"],\"\\\" wurde erfolgreich nach \",[\"destination\"],\" kopiert.\"],\"5g7owI\":[\"Floating IP wird aktualisiert...\"],\"5y3O+A\":[\"Deaktivieren\"],\"6+7EwD\":[\"Objekte als Index bereitstellen, wenn Dateiname ist:\"],\"6+OdGi\":[\"Protokoll\"],\"6/xipy\":[\"Container-Format\"],\"644xgx\":[\"Geschützt\"],\"6BDqha\":[\"Limits\"],\"6CDYXS\":[\"Statisches Website-Serving\"],\"6GBt0m\":[\"Metadata\"],\"6H/Lg1\":[\"Dies ist ein öffentliches Image. Alle Benutzer haben Zugriff darauf. Eine explizite Freigabe ist nicht erforderlich.\"],\"6KRclz\":[\"Ordner erstellt\"],\"6Kjltl\":[\"Zugriffskontrolle für Container:\"],\"6OopEX\":[\"Container geleert\"],\"6Rnrsz\":[\"Zugriff verwalten - \",[\"flavorName\"]],\"6X/9Di\":[\"\\\"\",[\"objectName\"],\"\\\" wurde erfolgreich nach \",[\"destination\"],\" verschoben.\"],\"6YtxFj\":[\"Name\"],\"6jAi8c\":[\"Bereich\"],\"6luZQA\":[\"Objekt verschoben\"],\"6oolxV\":[\"Dieser Metadata-Key existiert bereits. Bitte verwenden Sie einen anderen Key.\"],\"6qzsuS\":[\"Schreib-ACLs\"],\"6sxz+g\":[\"Port-Name\"],\"6w+VnM\":[\"Container erstellt\"],\"6z9W13\":[\"Neustart\"],\"76RKuS\":[\"ICMP-Code\"],\"78+riR\":[\"Sie sind nicht berechtigt, Projektzugriffe zu entfernen. Bitte melden Sie sich erneut an.\"],\"7AfIPZ\":[\"Floating Network\"],\"7BpykL\":[\"Fehler beim Erstellen des Extra-Specs. Bitte versuchen Sie es erneut.\"],\"7L01XJ\":[\"Aktionen\"],\"7NC3vm\":[\"Subnet\"],\"7NSdfG\":[\"Container \",[\"progressCurrent\"],\" von \",[\"progressTotal\"],\" wird geleert, bitte warten...\"],\"7Q24LN\":[\"Richtlinie\"],\"7T1fHv\":[\"Fehler beim Entfernen des Mitglieds \\\"\",[\"memberIdToRemove\"],\"\\\"\"],\"7UlHhT\":[\"Metadaten \\\"\",[\"keyToDelete\"],\"\\\" wurden erfolgreich gelöscht.\"],\"7XQ3QJ\":[\"Abgelehnter Referrer: \",[\"host\"]],\"7ZnTL8\":[\"Fehler beim Aktualisieren des Objekts: \",[\"mutationErrorMessage\"]],\"7a4DvD\":[\"Keine Server verfügbar\"],\"7d1a0d\":[\"Öffentlich\"],\"7flw0l\":[\"Der Projektzugriff für \\\"\",[\"trimmedTenantId\"],\"\\\" wurde erfolgreich hinzugefügt.\"],\"7sMeHQ\":[\"Key\"],\"88kg0+\":[\"Erstellt am\"],\"8AriEH\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde erstellt\"],\"8TSI9h\":[\"Durch die Deaktivierung dieses Images kann es nicht mehr zum Starten neuer Instances verwendet werden. Bestehende Instances sind nicht betroffen.\"],\"8Tg/JR\":[\"Benutzerdefiniert\"],\"8ZOb7O\":[[\"numberDeleted\"],\" Objekt wurde dauerhaft gelöscht.\"],\"8ZsakT\":[\"Passwort\"],\"8c3/77\":[\"Maximale Metadaten-Namenslänge\"],\"8jLXs3\":[\"Versionierte Schreibvorgänge\"],\"8s0tOH\":[\"Sie haben keine Berechtigung, diesem Flavor Projektzugriffe hinzuzufügen.\"],\"8t1+HU\":[[\"successCount\"],\" Image(s) deaktiviert, aber \",[\"failedCount\"],\" Image(s) konnten nicht deaktiviert werden.\"],\"8wdCNd\":[\"tcp, udp, icmp oder Protokollnummer\"],\"8zAn1f\":[\"Fehler beim Löschen des Flavors. Bitte versuchen Sie es erneut.\"],\"98Fs4G\":[\"Image wird erstellt...\"],\"9J93Xr\":[\"Container-Name darf keine Schrägstriche enthalten\"],\"9SX0bO\":[\"Das Image \\\"\",[\"imageName\"],\"\\\" konnte nicht aktualisiert werden: \"],\"9X8lAk\":[\"Zuweisen\"],\"9doWrf\":[\"Fehler beim Hinzufügen des Mitglieds\"],\"9dsDHD\":[\"Das Image \\\"\",[\"imageId\"],\"\\\" konnte nicht reaktiviert werden: \",[\"message\"]],\"9iz2XW\":[\"Image kann nicht aktualisiert werden\"],\"9njIiV\":[\"Fehler beim Aktivieren der Images\"],\"9rz81C\":[\"Geräte-ID\"],\"9v5VLp\":[\"Keine benutzerdefinierten Eigenschaften definiert\"],\"9vSW3U\":[\"Rekursiv löschen\"],\"9x6EkK\":[\"Dies ist ein öffentlicher Flavor. Alle Projekte haben Zugriff darauf.\"],\"A7CVME\":[\"Wählen Sie zuerst das Festplattenformat\"],\"AB4Tnl\":[\"Bitte wählen Sie eine Datei zum Hochladen aus\"],\"AGXLLY\":[\"Image-Datei kann nicht hochgeladen werden\"],\"AJRhSM\":[\"Root Disk muss eine ganze Zahl ≥ 0 sein.\"],\"AN0DBJ\":[\"Enter drücken zum Hinzufügen\"],\"AX9Juz\":[\"Die ID darf nur alphanumerische Zeichen, Bindestriche, Unterstriche und Punkte enthalten.\"],\"AZyHwC\":[\"Muss eine gültige IPv4- oder IPv6-Adresse sein (z. B.: 172.24.4.228 oder 2001:db8::1).\"],\"Ac6dy9\":[\"Name eingeben\"],\"AdtLNV\":[\"Stellen Sie sicher, dass ACL-Einträge gültig sind — korrekte Projekt-IDs, Benutzer-IDs und Formate liegen in Ihrer Verantwortung. Ungültige Einträge können stillschweigend unbeabsichtigten Zugriff gewähren oder verweigern.\"],\"AeXO77\":[\"Account\"],\"Afh/Lb\":[\"Zielordner auswählen\"],\"AlbUVn\":[\"<0>Optimierte Skalierbarkeit:</0> Aurora ist für Unternehmen jeder Größe konzipiert und wächst mit Ihnen mit, unterstützt einfache Umgebungen und komplexe Multi-Cloud-Setups gleichermaßen.\"],\"Alx2/L\":[\"In neuem Tab öffnen\"],\"AuQtzx\":[\"Muss eine nicht-negative Ganzzahl sein\"],\"AxZkIr\":[\"Disk (GiB)\"],\"B2Czeb\":[\"Min. RAM\"],\"B2i9cQ\":[\"Zu löschende Objekte (\",[\"totalCount\"],\")\"],\"B3toQF\":[\"Objekte\"],\"B4Jzm7\":[\"Ceph\"],\"BCJPTn\":[\"Zugriff für alle Benutzer aus diesem Projekt gewähren.\"],\"BCXapL\":[\"Fehler beim Laden der Container-Eigenschaften: \",[\"errorMessage\"]],\"BJt+PJ\":[\"Fehler beim Löschen des Containers\"],\"BMTd81\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Das Zielprojekt verliert sofort den Zugriff auf diese Sicherheitsgruppe.\"],\"BOQYRn\":[\"Lädt Key Pairs...\"],\"BP4Fwj\":[\"Fehler beim Laden der Objekte: \",[\"errorMessage\"]],\"BSaBkZ\":[\"Objekte — \",[\"containerName\"]],\"BYH/2L\":[\"Image kann nicht deaktiviert werden\"],\"BZpsYm\":[\"Failed to load containers: \",[\"errorMessage\"]],\"BgMp/T\":[\"Ungültige Formatkombination für das ausgewählte Festplattenformat\"],\"BoIAP6\":[\"Die ID des Netzwerks, das der Floating IP zugeordnet ist.\"],\"BoPocW\":[\"MD5-Prüfsumme\"],\"BrrIs8\":[\"Storage\"],\"CA8ZeT\":[\"Sichtbarkeit des Images \\\"\",[\"imageName\"],\"\\\" auf \",[\"visibility\"],\" aktualisiert\"],\"CBFSfX\":[\"Bitte korrigieren Sie die Validierungsfehler unten.\"],\"CFMxC8\":[\"Images gelöscht\"],\"CMVP7y\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Die Regel wird dauerhaft gelöscht.\"],\"CgZxr7\":[\"Min. RAM (MB)\"],\"ChOuUj\":[\"Floating IP nicht gefunden\"],\"Cj2Gtd\":[\"Größe\"],\"ClGcRq\":[\"Container\"],\"Cu6xuZ\":[\"Dies ist ein <0>dynamisches großes Objekt (DLO)</0>-Manifest. Metadaten-Änderungen gelten nur für das Manifest — Segment-Objekte sind nicht betroffen.\"],\"CunRry\":[\"Ungültiges Projekt-ID-Format. Muss 32 hexadezimale Zeichen sein (z. B. b90f9c4bc76140e18540b2cec1299e2a) oder UUID-Format (z. B. 12345678-1234-1234-1234-123456789abc)\"],\"Cxgv2U\":[\"Min. Disk\"],\"D+Se/Q\":[\"Container leeren:\"],\"D/8vkD\":[\"Es wird in Ihrer Image-Liste angezeigt.\"],\"D1F789\":[\"<0>Sind Sie sicher?</0> Der Container wird gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"D3IRXw\":[\"Floating IP wird getrennt...\"],\"D7qT9F\":[\"Warum Aurora wählen?\"],\"DDRhQm\":[\"Ihre Sitzung ist abgelaufen.\"],\"DJT9tB\":[\"Account-Quotas\"],\"DKkOPx\":[\"Zusätzliche Spezifikationen\"],\"DNVql8\":[\"Vollständiges Lifecycle-Management von Floating IPs, einschließlich Zuweisung, Port-Zuordnung/-Aufhebung, DNS-Einstellungen und Löschung\"],\"DcMIiu\":[\"ACLs für Container \\\"\",[\"containerName\"],\"\\\" konnten nicht aktualisiert werden: \",[\"errorMessage\"]],\"Df0YHr\":[\"Security Group aktualisieren\"],\"Dh1qvV\":[\"Sie sind dabei, \",[\"deletableCount\"],\" Image(s) zu löschen. Diese Aktion kann nicht rückgängig gemacht werden.\"],\"Dia2Ue\":[\"Für diese Security Group sind keine RBAC-Richtlinien vorhanden\"],\"Do5/uH\":[\"Der Flavor oder das Projekt konnte nicht gefunden werden. Möglicherweise wurde es bereits entfernt.\"],\"Dqnh7K\":[\"Spezifischer Referrer: \",[\"host\"]],\"Dt5W9T\":[\"RBAC-Richtlinie entfernen\"],\"DvB4XF\":[\"Legen Sie Ihre Datei hier ab\"],\"E/QGRL\":[\"Deaktiviert\"],\"E4QYe7\":[\"Empfohlene Images\"],\"E6nRW7\":[\"URL kopieren\"],\"EF2EU9\":[\"Wird gelöscht...\"],\"EPMHs9\":[\"Sie haben keine Berechtigung, Flavors in diesem Projekt zu löschen.\"],\"EQnVgi\":[\"Der Flavor-Service ist für dieses Projekt nicht verfügbar.\"],\"EdQY6l\":[\"Keine\"],\"Ef7StM\":[\"Unknown\"],\"Enpdmy\":[\"Geben Sie <0>entfernen</0> zur Bestätigung ein:\"],\"EoKe5U\":[\"Domain\"],\"Eq5PsT\":[\"\\\"detach\\\" zur Bestätigung eingeben\"],\"EqSPkP\":[\"Lädt Flavors...\"],\"Erlvqg\":[\"Objektname darf keine führenden oder nachfolgenden Leerzeichen haben\"],\"ExLULX\":[\"Image-Name\"],\"EztMB8\":[\"Fehler beim Abrufen der Flavors vom Server.\"],\"F02e8I\":[\"Keine benutzerdefinierten Metadaten. Klicken Sie auf \\\"Eigenschaft hinzufügen\\\", um eine zu erstellen.\"],\"F6YIQe\":[\"Effiziente Massenlöschung\"],\"FKL6Jv\":[\"z. B. .r:*,.rlistings\"],\"FNcMGM\":[\"Creation Date\"],\"FOcBn3\":[\"Trennen\"],\"FPsvA8\":[\"Verstanden!\"],\"FQBaXG\":[\"Aktivieren\"],\"FRtmJJ\":[\"Storage-Container nicht gefunden\"],\"FSbpS7\":[\"CPU\"],\"FjONW3\":[\"Fehler beim Laden des Flavors\"],\"FjPnAE\":[\"Fehler beim Laden der Security Group\"],\"Flugry\":[[\"progressPct\"],\"%\"],\"FwSyEp\":[\"Das angegebene Projekt existiert nicht oder Sie haben keine Berechtigung, es zu teilen.\"],\"Fzrzfe\":[\"Ordnername ist erforderlich\"],\"G6AP+o\":[\"Geteilt:\"],\"GEgjm+\":[\"Objekte werden geladen...\"],\"GPuCEo\":[\"Leer lassen für alle Typen\"],\"GSIPwA\":[\"Temporäre URL\"],\"GbKqnI\":[[\"successCount\"],\" Image(s) aktiviert, aber \",[\"failedCount\"],\" Image(s) konnten nicht aktiviert werden.\"],\"Gfx1qQ\":[\"Inhalt kann nicht geladen werden\"],\"GxkJXS\":[\"Wird hochgeladen...\"],\"Gyd3No\":[\"Kein spezifischer Projektzugriff für diesen privaten Flavor konfiguriert. Klicken Sie auf \\\"Projektzugriff hinzufügen\\\", um Zugriff zu gewähren.\"],\"H+a5j6\":[\"Freigeben\"],\"H4Qwmp\":[\"Keine Objekte entsprechen Ihrer Suche. Versuchen Sie, Ihren Suchbegriff anzupassen.\"],\"H7u085\":[\"Noch kein Projekt hat Zugriff auf dieses Image. Klicken Sie auf \\\"Projektzugriff hinzufügen\\\", um Zugriff zu gewähren.\"],\"HAkrpK\":[\"Bei Aurora ist es unsere Mission, eine zentrale Plattform bereitzustellen, die das Cloud-Management vereinheitlicht. Wir streben danach, die Komplexitäten der Bereitstellung, Konfiguration und Skalierung von Ressourcen über verschiedene Cloud-Umgebungen hinweg zu vereinfachen und gleichzeitig ein nahtloses Wachstum für Ihr Unternehmen zu ermöglichen.\"],\"HBI6nY\":[\"Benötigen Sie Hilfe?\"],\"HBpi4q\":[\"Lädt Images...\"],\"HM56Bx\":[\"Creating...\"],\"HNlEFZ\":[\"löschen\"],\"HQH8HM\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht aktualisiert werden: \",[\"errorMessage\"]],\"HVdrr1\":[\"Beliebiger Referrer\"],\"HivZR9\":[\"Create Credential\"],\"Hivb/4\":[\"Der Server hat Probleme. Bitte versuchen Sie es später erneut.\"],\"Hiw1Ha\":[\"Keine Container gefunden\"],\"HlwgQN\":[\"Objekt \\\"\",[\"objectName\"],\"\\\" wurde dauerhaft gelöscht.\"],\"HuA8iQ\":[\"Floating IP wird zugewiesen...\"],\"HxTYrE\":[\"Der Flavor konnte nicht gefunden werden. Möglicherweise wurde er gelöscht.\"],\"I5kZVK\":[\"Entfernte Quelle\"],\"INUP6f\":[\"<0>Mühelose Ressourcenbereitstellung:</0> Stellen Sie Ressourcen wie Server, Netzwerke und Volumes schnell bereit, konfigurieren Sie diese und setzen Sie sie mit nur wenigen Klicks ein.\"],\"IOkHLC\":[\"Fehler beim Kopieren des Objekts: \",[\"errorMessage\"]],\"IUwGEM\":[\"Änderungen speichern\"],\"IWF68U\":[\"Speicherübersicht\"],\"IZ6Mh2\":[\"Geben Sie den Domain ein\"],\"IbYr/u\":[\"Content type\"],\"J4DKSM\":[\"Container-Format ist erforderlich\"],\"J6EOll\":[\"Objekt verschieben/umbenennen:\"],\"J7+bZb\":[\"Ordner gelöscht\"],\"J9QcnV\":[[\"successCount\"],\" von \",[\"totalCount\"],\" Image(s) erfolgreich aktiviert\"],\"J9cmxx\":[\"Fehler beim Aktualisieren der Sichtbarkeit auf \",[\"newVisibility\"]],\"JB0bhm\":[\"Mitmachen\"],\"JNGYAW\":[\"Container-Name ist erforderlich\"],\"JT3I1g\":[\"Flavor löschen\"],\"JeRXll\":[\"Dieser Schlüssel ist reserviert und wird separat verwaltet\"],\"JfWCsP\":[\"Teilweise erfolgreich deaktiviert\"],\"Jh4rAZ\":[\"Fehler beim Laden des Images\"],\"Jim5X9\":[\"Stateful\"],\"JoECY1\":[\"Die bereitgestellten Metadata-Daten sind ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"JpZn1L\":[\"Bereits deaktiviert (wird übersprungen)\"],\"JrmKyf\":[\"Fehlgeschlagen: \",[\"errorDetails\"]],\"JtHgVz\":[\"Images löschen\"],\"K+e/0e\":[\"RAM (MiB)\"],\"K3bUTE\":[\"Minimale Festplatte muss 0 oder größer sein\"],\"K8Qnlj\":[\"Wird verschoben...\"],\"K9eC8x\":[\"Dies kann auf unzureichende Berechtigungen oder ein vorübergehendes Serviceproblem zurückzuführen sein. Bitte überprüfen Sie Ihre Zugriffsrechte oder aktualisieren Sie die Seite.\"],\"KDw4GX\":[\"Erneut versuchen\"],\"KJC+M7\":[\"Serverfehler beim Abrufen der Flavor-Details. Bitte versuchen Sie es später erneut.\"],\"KOpPMt\":[\"Gesamtspeicher-Quota\"],\"KSW/GC\":[\"Keine Flavors verfügbar. Filtern Sie neu oder erstellen Sie einen neuen Flavor.\"],\"KZN4Lc\":[\"Alle löschen\"],\"Km4AGG\":[\"Sicherheitsgruppe wird erstellt...\"],\"KoQP4F\":[\"Ein Serverfehler ist aufgetreten, während der Projektzugriff hinzugefügt wurde. Bitte versuchen Sie es später noch einmal.\"],\"KsIM0b\":[\"Boot-RAM\"],\"KsnZ3m\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" wurde erfolgreich erstellt.\"],\"KzUd7m\":[\"new-folder-name\"],\"LI8Z2I\":[[\"rowDisplayName\"],\" herunterladen\"],\"LK0pQN\":[\"Festplattenformat ist erforderlich\"],\"LMdsuJ\":[\"Port (von)\"],\"LQQCas\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" und \",[\"deletedCount\"],\" Objekt wurden dauerhaft gelöscht.\"],\"Llcakz\":[\"Aktualisiert am\"],\"LqMb+g\":[\"Um diese Aktion zu bestätigen, geben Sie das Wort <0>\\\"release\\\"</0> in das Feld unten ein.\"],\"LtI9AS\":[\"Eigentümer\"],\"Lylr9Z\":[\"Objekt kopiert\"],\"M470oJ\":[\"Der Flavor konnte nicht gefunden werden oder hat keine Metadata.\"],\"M5Epeo\":[\"Image-Details bearbeiten\"],\"M5RhXF\":[\"Wird entfernt...\"],\"M5rEN5\":[\"Sitzung abgelaufen\"],\"M9H+/G\":[\"Projekte\"],\"MEIAzV\":[\"Unbenannt\"],\"MJtNLd\":[\"Zu löschende Images:\"],\"MOug+V\":[\"Geben Sie ein Tag ein und drücken Sie Enter oder klicken Sie auf Hinzufügen\"],\"MRB7nI\":[\"Richtung\"],\"MXoA/6\":[\"Objekt hochladen\"],\"MXw7Fr\":[\"Servername\"],\"MZGbkp\":[\"VCPUs\"],\"MbKJNP\":[\"Sie haben keine Berechtigung, auf die Zugriffsinformationen für diesen Flavor zuzugreifen.\"],\"MgZyuJ\":[\"Sie sind dabei, <0>\",[\"deactivatedCount\"],\" Image(s)</0> zu aktivieren. Aktivierte Images stehen zum Starten neuer Instances zur Verfügung.\"],\"MmtQVF\":[\"Ungültiger Wert für die Einstellung des öffentlichen Flavors.\"],\"Mt6sRo\":[\"Sie sind nicht berechtigt, auf die Zugriffsinformationen des Flavors zuzugreifen. Bitte melden Sie sich erneut an.\"],\"MtzSbv\":[\" Objektname ist erforderlich\"],\"MuKU9V\":[\"Failed to load objects: \",[\"errorMessage\"]],\"N2S1rs\":[\"Leeren\"],\"N5I2RJ\":[\"\\\"release\\\" zur Bestätigung eingeben\"],\"N5vGcw\":[\"Geben Sie Ihre Anmeldedaten ein, um auf Ihr Konto zuzugreifen.\"],\"NH2fsP\":[\"Bereits deaktiviert (wird übersprungen):\"],\"NOdFZR\":[\"Wird generiert...\"],\"NQU1Nn\":[\"Container-Name kopieren\"],\"NRMm0E\":[\"Dieses Projekt hat bereits Zugriff auf den Flavor.\"],\"NRP2uq\":[\"Objekt teilen:\"],\"NRVSdy\":[\"Mitglieds-ID\"],\"NW4PIb\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" konnte nicht erstellt werden: \",[\"errorMessage\"]],\"NZJhro\":[\"Objektname darf keine Schrägstriche enthalten\"],\"Nc7QKU\":[\"Feste IP-Adresse\"],\"NeUjqc\":[\"Dateilisting aktivieren\"],\"NixRmA\":[\"Min. Disk (GB)\"],\"NlcF/v\":[\"Kein Flavor für die Löschung ausgewählt.\"],\"NopYGU\":[\"Festplattenformat\"],\"Np28ib\":[\"oder per Drag & Drop\"],\"Nu4oKW\":[\"Beschreibung\"],\"Nvfd2b\":[\"Versionierung ist aktiviert\"],\"O80bQY\":[\"Objekt-Eigenschaften werden geladen...\"],\"O8tK4v\":[\"Regel hinzufügen\"],\"ONWvwQ\":[\"Hochladen\"],\"OR475H\":[\"Netzwerk\"],\"OSlLnz\":[\"Image-Sichtbarkeit\"],\"OYHzN1\":[\"Tags\"],\"OZImTR\":[\"Container-Listing-Limit\"],\"OaSktR\":[\"Geräteeigentümer\"],\"OlmKCg\":[\"Ein Flavor mit dieser ID oder diesem Namen existiert bereits. Bitte verwenden Sie andere Werte.\"],\"OvEjsP\":[\"Wird kopiert...\"],\"Ovofy+\":[\"Floating IP \",[\"floating_ip_address\"],\" freigeben\"],\"OxDN2m\":[\"Fehler beim Erstellen des Flavors. Bitte versuchen Sie es erneut.\"],\"OxaeYj\":[\"Wir entwickeln das Aurora-Dashboard, um Ihnen einen besseren Service zu bieten. Ihr Feedback ist von unschätzbarem Wert, um ein Werkzeug zu gestalten, das den einzigartigen Bedürfnissen von Unternehmen wie Ihrem gerecht wird. Bleiben Sie in Verbindung und begleiten Sie uns, während wir das Cloud-Management neu definieren.\"],\"Oxl1UN\":[\"Wenn keine Index-Datei vorhanden ist, zeigt die URL eine Liste der Objekte im Container an.\"],\"PAKSdy\":[\"Floating IP eingeben oder leer lassen für automatische Zuweisung\"],\"PEGvy+\":[\"Minimaler RAM muss 0 oder größer sein\"],\"PHt+EV\":[\"<0>delete</0> zur Bestätigung eingeben:\"],\"PIbPRX\":[\"RX/TX Faktor muss eine ganze Zahl ≥ 1 sein.\"],\"PLwzWR\":[\"Alle Container\"],\"PYQUjU\":[\"Metadaten-Konfiguration konnte nicht geladen werden.\"],\"PZnUbs\":[\"Bitte melden Sie sich erneut an, um fortzufahren.\"],\"PgNNGl\":[\"Weitere Aktionen\"],\"PiH3UR\":[\"Kopiert!\"],\"PiyQJ/\":[\"Keine Flavors gefunden\"],\"PkfPsB\":[\"Geben Sie die ID des Projekts ein, mit dem Sie diese Sicherheitsgruppe teilen möchten. Sie finden Projekt-IDs im Konto-/Projekt-Umschalter oder im Identity-Service.\"],\"Pkw7J9\":[\"Dieser Ordner ist leer.\"],\"PsEGri\":[\"Ubuntu 22.04 LTS\"],\"PtjzS+\":[\"Wird dem ausgewählten Port zugeordnet. Wenn der Port mehrere IPs hat, wählen Sie die gewünschte feste IP-Adresse.\"],\"PzgYM9\":[\"Prüfsumme\"],\"Q1W//7\":[\"Keine Dienste für dieses Projekt verfügbar.\"],\"Q2xmVl\":[\"Symlinks\"],\"Q9f2QF\":[[\"numberDeleted\"],\" Objekte wurden erfolgreich gelöscht, aber einige Löschvorgänge sind fehlgeschlagen.\"],\"QAUa4B\":[\"Geben Sie einen einzelnen Port ein oder definieren Sie einen Bereich, indem Sie auch \\\"Port (bis)\\\" ausfüllen. \\\"Port (bis)\\\" ist optional.\"],\"QEtDlS\":[\"Objekt wird kopiert...\"],\"QNHur0\":[\"Fehler beim Laden der Container-ACLs: \",[\"errorMessage\"]],\"QQ8wUG\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Der Flavor wird dauerhaft gelöscht.\"],\"QV1ZPO\":[\"Schlüssel ist erforderlich\"],\"QWdKwH\":[\"Verschieben\"],\"QYiqYb\":[\"Fehler beim Aktualisieren des Zugriffsstatus\"],\"Qb+14I\":[\"Diese Aktion kann nicht rückgängig gemacht werden. Die Security Group wird dauerhaft gelöscht.\"],\"QetsXP\":[\"Upload fehlgeschlagen: \",[\"uploadError\"]],\"Qg4EG6\":[\"Die bereitgestellten Flavor-Daten sind ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"QuJSSl\":[\"Fehler beim Erstellen des Flavors. Bitte versuchen Sie es erneut.\"],\"QvqBQa\":[\"Ziel-Container\"],\"Qx7DM7\":[\"Capabilities\"],\"QxBGbh\":[\"Geschützt (wird übersprungen):\"],\"R0xWJL\":[\"Löschen nicht möglich. Container enthält Objekte. Bitte leeren Sie ihn zuerst.\"],\"R6kcsL\":[\"Muss eine gültige PQDN oder FQDN sein (nur alphanumerische Zeichen und Bindestriche, darf nicht mit einem Bindestrich beginnen oder enden).\"],\"R6u5CR\":[[\"failedCount\"],\" von \",[\"totalCount\"],\" Image(s) konnten nicht aktiviert werden. Einige Images sind möglicherweise bereits aktiv oder in einem ungültigen Zustand.\"],\"RByeNR\":[\"Ihre Sitzung ist abgelaufen. Bitte melden Sie sich erneut an.\"],\"RCr0yv\":[\"Flavor-Details konnten nicht geladen werden. Bitte versuchen Sie es erneut.\"],\"RFDYCD\":[\"Minimale Festplattengröße, die zum Booten dieses Images erforderlich ist\"],\"RGhYAo\":[\"RAM\"],\"RGrgxg\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht gelöscht werden: \",[\"errorMessage\"]],\"RGwfoL\":[\"Maximale Metadaten-Anzahl\"],\"RNBvdl\":[\"Maximale SLO-Segmente\"],\"RSFkXF\":[\"Image aktivieren\"],\"RSMPjT\":[\"Sie befinden sich derzeit auf der Dashboard-Route.\"],\"RSg/pq\":[\"Fehler beim Löschen des Objekts\"],\"RTQFAw\":[\"Sie sind nicht berechtigt, Metadata zu erstellen. Bitte melden Sie sich erneut an.\"],\"Rih53k\":[\"Maximale Container-Namenslänge\"],\"Rlp5zj\":[\"Flavor erstellen\"],\"S0kLOH\":[\"ID\"],\"S1iTXO\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde gelöscht\"],\"S3olSf\":[\"Keine Metadata gefunden. Klicken Sie auf \\\"Metadata hinzufügen\\\", um eine zu erstellen.\"],\"S5CUKP\":[\"Mitglieds-ID (Projekt-UUID) ist erforderlich.\"],\"S63NbU\":[\"Das Image \\\"\",[\"imageId\"],\"\\\" konnte nicht deaktiviert werden: \",[\"message\"]],\"S8/j2h\":[\"Fehler beim Leeren des Containers\"],\"SBGiGm\":[\"Lese-ACLs\"],\"SCY5an\":[\"Fehler beim Verschieben des Objekts\"],\"SFo0kK\":[\"Alle Images\"],\"SIfYq6\":[\"Metadaten bearbeiten\"],\"SLEH7X\":[\"DNS-Name eingeben\"],\"STc+7E\":[\"Maximale Container pro Extraktion\"],\"SU0uxT\":[\"Objekt hochladen nach:\"],\"SUSS9i\":[\"Container-Name\"],\"SVLToM\":[\"\\\"remove\\\" zur Bestätigung eingeben\"],\"SZw9tS\":[\"Details ansehen\"],\"Sb/VT5\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich geleert. \",[\"deletedCount\"],\" Objekte gelöscht.\"],\"SfW/3r\":[\"Keine Gruppen vorhanden\"],\"Sgz1vJ\":[\"Mitglied \\\"\",[\"trimmedMemberId\"],\"\\\" wurde erfolgreich hinzugefügt.\"],\"Smk7M2\":[\"Fehler beim Laden der Floating-IP\"],\"SuX2Ca\":[\"Grundlegende Informationen\"],\"SysqAR\":[\"Flavor-Details\"],\"T6Gm5y\":[\"Externes Netzwerk auswählen\"],\"T7mgdd\":[[\"successCount\"],\" von \",[\"totalCount\"],\" Image(s) erfolgreich gelöscht\"],\"T8N6oi\":[\"Eigenschaftsschlüssel\"],\"T9Mtpi\":[\"Projekt ID\"],\"TM93nK\":[\"Sicherheitsgruppenregel löschen\"],\"TPMaxo\":[\"“release” zur Bestätigung eingeben\"],\"TQn3hH\":[\"Image konnte nicht erstellt werden. Bitte versuchen Sie es erneut.\"],\"TZJiVf\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich geleert. \",[\"deletedCount\"],\" Objekt gelöscht.\"],\"TfC9O+\":[\"Zuletzt geändert (UTC)\"],\"TfdeUd\":[\"Fehler beim Löschen des Extra-Specs. Bitte versuchen Sie es erneut.\"],\"TpGxnq\":[\"Mitglieds-ID eingeben\"],\"Tx4Ym+\":[\"Gültige PQDN oder FQDN (max. 63 Zeichen) eingeben, um sie der Floating IP zuzuordnen. A- und PTR-Einträge werden automatisch erstellt.\"],\"TyODHt\":[\"Metadata speichern\"],\"U/oahm\":[\"URL kopiert\"],\"U2wTy/\":[\"Hinweis: Das Attribut 'stateful' kann nicht geändert werden, wenn diese Security Group derzeit von einem oder mehreren Ports verwendet wird.\"],\"U4fmHG\":[\"Der Text muss “detach” in Kleinbuchstaben entsprechen.\"],\"U6L+P/\":[\"Inaktivitäts-Timeout\"],\"U9q4M7\":[\"Zurück zu Security Groups\"],\"UGhVPl\":[\"Objekttyp\"],\"UJVf0u\":[\"Image wird geladen...\"],\"UK2mpr\":[\"Temporäre URL wird generiert...\"],\"UKwOYH\":[\"Image-Datei\"],\"UO3hJ2\":[\"Temporäre URLs\"],\"UQ7Wyv\":[\"Zugriff für Image verwalten - \",[\"imageName\"]],\"URmyfc\":[\"Details\"],\"USiuNX\":[\"Container-Quotas\"],\"UVFHGY\":[\"z. B. PROJECT_ID:USER_ID\"],\"UVSFVV\":[\"Geteiltes Image ablehnen\"],\"UYSopm\":[\"Minimaler RAM (MB)\"],\"UbRKMZ\":[\"Ausstehend\"],\"UdcGJu\":[\"Images aktivieren\"],\"UiNv/G\":[\"S3 Object Storage requires EC2 credentials (access key + secret key) to authenticate your requests. You need to create credentials before accessing S3 resources.\"],\"Uj+n/2\":[\"Fehler beim Löschen des Ordners\"],\"UkVkoq\":[\"Leer lassen für alle Codes\"],\"UmQ3/m\":[\"Ausgewählte deaktivieren\"],\"Uwo8Xw\":[\"Dieses Image wurde am \",[\"sharedAt\"],\" von <0>\",[\"ownerProject\"],\"</0> mit Ihnen geteilt.\"],\"UztfYZ\":[\"Port zur Zuordnung auswählen\"],\"V/8B9A\":[\"Ich bestätige, dass alle vorhandenen Versionen ebenfalls gelöscht werden\"],\"V/SINY\":[\"Objekt aktualisieren\"],\"V1TzeS\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich gelöscht.\"],\"V66Jih\":[\"Zugriffsstatus\"],\"V7fN5X\":[\"Objekt kopieren:\"],\"V804LY\":[\"Security Group wird aktualisiert...\"],\"VCM3KS\":[\"Projektzugriff hinzufügen\"],\"VKmlZ+\":[\"Zu leerende Container (\",[\"totalCount\"],\")\"],\"VLI9eO\":[\"Floating IP-Details werden geladen...\"],\"VV1fdg\":[\"Jeder Benutzer hat Lesezugriff auf Objekte. Es ist kein Token in der Anfrage erforderlich.\"],\"VaA9mu\":[\"24 Stunden\"],\"VakxP/\":[\"Fehler beim Hochladen des Objekts\"],\"Vg0k6h\":[\"Zeige \",[\"filteredCount\"],\" von \",[\"totalCount\"],\" \",[\"itemName\"]],\"Vh/Uj5\":[\"Zielpfad\"],\"Vj8XFg\":[\"Fehler beim Erstellen des Containers\"],\"Vl4XTj\":[\"Ordnername darf keine Schrägstriche enthalten\"],\"Vmojta\":[\" Zugriffsstatus auf \\\"\",[\"newStatus\"],\"\\\" aktualisiert.\"],\"VoxR3s\":[\"Objekt wurde kopiert, konnte aber nicht von der Quelle gelöscht werden: \",[\"deleteErrorMessage\"]],\"Vz+7ZA\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht erstellt werden: \",[\"errorMessage\"]],\"Vzlopx\":[\"Container löschen:\"],\"W0MCSG\":[\"Zugriff auf Image akzeptieren\"],\"W5FkH9\":[\"Container-Name eingeben\"],\"W9PZE0\":[\"Objekte gelöscht\"],\"W9kfjU\":[\"QoS-Policy-ID\"],\"WCKEqI\":[\"Dies ist ein <0>statisches großes Objekt (SLO)</0>-Manifest. Metadaten-Änderungen gelten nur für das Manifest — Segment-Objekte sind nicht betroffen.\"],\"WCLyHI\":[\"Keine Floating IPs gefunden\"],\"WErCZy\":[\"Minimaler RAM, der zum Booten dieses Images erforderlich ist\"],\"WRZ3Mt\":[\"Container-Eigenschaften werden geladen...\"],\"WYb0Td\":[\"<0>Sind Sie sicher?</0> Alle Objekte in den ausgewählten Containern werden dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"WYiUDa\":[\"Container werden geladen...\"],\"Wbg1jv\":[[\"text\"],\" in die Zwischenablage kopieren\"],\"Wca9WC\":[\"Security Groups konnten nicht geladen werden\"],\"WlpcJv\":[\"DNS-Domain\"],\"WoSkGY\":[\"Entfernte IP\"],\"WrUky8\":[\"Teilen (Temporäre URL)\"],\"WyKwnD\":[\"Alte Objekt-Versionen speichern\"],\"WzVwU0\":[\"Zielprojekt-ID\"],\"X2OnDx\":[\"Ephemeral Disk (GiB)\"],\"X70LXS\":[[\"numberDeleted\"],\" Objekte wurden dauerhaft gelöscht.\"],\"XLk16/\":[\"Gemeinsam können wir das volle Potenzial Ihrer Cloud-Infrastruktur erschließen.\"],\"XYZLy9\":[\"Schlüssel enthält ungültige Zeichen\"],\"XvjC4F\":[\"Wird gespeichert...\"],\"XwxJJB\":[\"Container \\\"\",[\"containerName\"],\"\\\" wurde erfolgreich erstellt.\"],\"XxjLdW\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" Container war bereits leer.\"],\"other\":[\"#\",\" Container waren bereits leer.\"]}]],\"Y+2SDm\":[\"Security Group \\\"\",[\"securityGroupName\"],\"\\\" löschen\"],\"Y1YKad\":[\"Details bearbeiten\"],\"YIix5Y\":[\"Suchen...\"],\"YNgcgc\":[\"Flavor-Details werden geladen...\"],\"YRexkb\":[\"Objekt aktualisiert\"],\"YUU0QW\":[\"Flavor-ID ist erforderlich und darf nicht leer sein\"],\"YZmsaT\":[\"Teilweise erfolgreich aktiviert\"],\"YiMCKk\":[\"ACLs-Vorschau anzeigen\"],\"Yin3uB\":[\"Floating IP wird freigegeben...\"],\"YjAOtb\":[\"Security Group erstellen\"],\"YrAy/S\":[\"Sie haben keine Berechtigung, Metadata für diesen Flavor zu löschen.\"],\"YsOJlj\":[\"Ein Serverfehler ist aufgetreten, während die Zugriffsinformationen für den Flavor abgerufen wurden. Bitte versuchen Sie es später noch einmal.\"],\"YsrbQh\":[\"Eigentümer-Projekt-ID\"],\"YuC9dj\":[\"Zuordnen\"],\"YuGQWb\":[\"Regeltyp\"],\"Z/eWPC\":[\"Das Objekt wird zu diesem Pfad kopiert. Navigieren Sie oben durch die Ordner, um das Ziel zu ändern.\"],\"Z2fZGD\":[\"Kein Projekt ausgewählt\"],\"Z3FXyt\":[\"Lädt...\"],\"Z42tfY\":[\"Ordner im Object Storage sind virtuell — sie werden als Null-Byte-Platzhalterobjekte mit einem abschließenden Schrägstrich erstellt. Der Ordner wird nach der Erstellung angezeigt.\"],\"Z5r9vC\":[\"Teilweise erfolgreich gelöscht\"],\"Z8lGw6\":[\"Teilen\"],\"ZAx+d1\":[\"Maximale Gesamt-Metadatengröße\"],\"ZAy0zp\":[[\"successCount\"],\" von \",[\"totalCount\"],\" Image(s) erfolgreich deaktiviert\"],\"ZUmOzn\":[\"Der Server hat ein unerwartetes Datenformat für Flavor-Details zurückgegeben.\"],\"ZcWMT1\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde deaktiviert\"],\"Zgp2Sm\":[\"Keine Projekte verfügbar.\"],\"ZhVSpK\":[\"Fehler beim Entfernen des Mandantenzugriffs vom Flavor. Bitte versuchen Sie es erneut.\"],\"Zq6Y5u\":[\"Der DNS-Name darf höchstens 63 Zeichen lang sein.\"],\"ZvIpwi\":[\"Security Group auswählen...\"],\"Zw49f9\":[\"folder-name\"],\"Zw8Q49\":[\"Security Group nicht gefunden\"],\"a/nTb8\":[\"Image erstellen\"],\"a12lSo\":[\"Port (bis)\"],\"a13wDR\":[\"Zum Suchen von Containern eingeben...\"],\"a3LDKx\":[\"Sicherheit\"],\"a4A2uB\":[\"Account-Listing-Limit\"],\"a4N/Bg\":[\"Mehr laden\"],\"a7C4YS\":[\"Container aktualisiert\"],\"a88X3d\":[\"<0>Sind Sie sicher?</0> Objekt <1>\\\"\",[\"displayName\"],\"\\\"</1> wird dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"aG9OiI\":[\"Freigabe-Details\"],\"aI8Tgp\":[\"Eigentümer-Projekt-ID\"],\"aL1w5Z\":[\"Belegt\"],\"aOeFR+\":[\"Container leeren\"],\"aSsVD3\":[\"Öffentlicher Lesezugriff ist nicht aktiviert. Bevor Sie statisches Website-Serving konfigurieren, gehen Sie zu <0>Zugriff verwalten</0> und aktivieren Sie den öffentlichen Lesezugriff.\"],\"aTqCTq\":[\"Image-Datei ist erforderlich\"],\"aV6KPH\":[\"Versehentliches Löschen verhindern\"],\"aiqFbS\":[\"<0>Sind Sie sicher?</0> Die ausgewählten Objekte werden dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"an5hVd\":[\"Images\"],\"ao/ZJi\":[\"Ordner und alle Inhalte werden gelöscht...\"],\"aqagJH\":[\"Image-Sichtbarkeit kann nicht aktualisiert werden\"],\"arel2K\":[\"Keine Objekte gefunden\"],\"azXlY+\":[\"Zugriffsstatus:\"],\"b0uU1G\":[\"Alte Objekt-Versionen in Container speichern:\"],\"b2BLBa\":[\"Sicherheitsgruppenregel hinzufügen\"],\"bISG26\":[\"Fehler beim Abrufen der Flavor-Zugriffsinformationen. Bitte versuchen Sie es erneut.\"],\"bM1O3m\":[\"Image-Instanz\"],\"bQBMTH\":[\"Dies ist ein <0>statisches großes Objekt</0>. Standardmäßig werden auch alle zugehörigen Segment-Objekte dauerhaft gelöscht.\"],\"bRgFkJ\":[\"Fehler beim Hochladen der Datei \\\"\",[\"fileName\"],\"\\\": \"],\"bYRFNi\":[\"Fehler beim Löschen der Objekte\"],\"bc67JN\":[\"Benutzerdefinierte Eigenschaften / Metadaten\"],\"bmQLn5\":[\"Regel hinzufügen\"],\"bnql/K\":[\"Zurück zu Images\"],\"boJ+Y1\":[\"Ordner erstellen\"],\"boJlGf\":[\"Seite nicht gefunden\"],\"bpme7e\":[\"Flavor nicht gefunden\"],\"bwRvnp\":[\"Aktion\"],\"bwhBhT\":[\"Security Group\"],\"byKna+\":[\"Ein unerwarteter Fehler ist aufgetreten. Bitte versuchen Sie es erneut.\"],\"bzMKg7\":[\"Akzeptiert\"],\"bzSI52\":[\"Verwerfen\"],\"c+fUtV\":[\"Beginnen Sie mit der Eingabe, um nach einem Container zu suchen\"],\"c+xCSz\":[\"True\"],\"c1uL4p\":[\"Das Image \\\"\",[\"imageId\"],\"\\\" konnte nicht gelöscht werden: \",[\"message\"]],\"c6b6fz\":[\"Ausgewählte löschen\"],\"cCfxH1\":[\"Wird heruntergeladen...\"],\"cJDQIO\":[\"Root Disk\"],\"cPKL6O\":[\"Sie sind nicht berechtigt, Flavors zu erstellen. Bitte melden Sie sich erneut an.\"],\"cWbW6w\":[\"Zugriff verwalten\"],\"cXuXkb\":[\"Benutzer \",[\"userId\"],\" aus Projekt \",[\"projectId\"]],\"chL5IG\":[\"Community\"],\"cj17eo\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde aktiviert\"],\"cjEOmc\":[\"Teilen Sie diese Sicherheitsgruppe mit einem anderen Projekt. Das Zielprojekt kann diese Sicherheitsgruppe ansehen und verwenden, aber nicht ändern oder löschen.\"],\"cnGeoo\":[\"Löschen\"],\"cpw++p\":[\"Unterstützung für statische große Objekte\"],\"cqQyPB\":[\"Ordnername\"],\"d+F6q9\":[\"Erstellt\"],\"d+Ugpw\":[\"<0>Sind Sie sicher?</0> Ordner <1>\\\"\",[\"folderDisplayName\"],\"\\\"</1> und alle darin enthaltenen Objekte werden dauerhaft gelöscht. Dies kann nicht rückgängig gemacht werden.\"],\"d/I0J3\":[\"Ausgewählte aktivieren\"],\"d0pLfy\":[\"Security Group konnte nicht gelöscht werden\"],\"dEgA5A\":[\"Abbrechen\"],\"dFb5Nt\":[\"Id\"],\"dLFiER\":[\"Fehler beim Laden der Container: \",[\"errorMessage\"]],\"dOevLB\":[\"Läuft ab in \",[\"selectedPresetLabel\"],\" — um \",[\"expiresAtFormatted\"]],\"dPBJAJ\":[\"Alle leeren (\",[\"selectedCount\"],\")\"],\"dPj4yB\":[\"Bei Ihrem Konto anmelden\"],\"dPoCVe\":[\"“detach” zur Bestätigung eingeben\"],\"dTNzBI\":[\"Schlüssel muss mindestens ein alphanumerisches Zeichen enthalten\"],\"dVdc7N\":[\"Sie sind nicht berechtigt, Metadata zu löschen. Bitte melden Sie sich erneut an.\"],\"dd2ndz\":[\"Einträge in ACLs werden durch Kommas getrennt. Beispiele:\"],\"diFNkW\":[\"Fehler beim Laden der Komponente\"],\"e0NrBM\":[\"Projekt\"],\"eChIh7\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" wurde erfolgreich erstellt.\"],\"eGEHJE\":[\"DNS-Name\"],\"eKC+EC\":[\"-\"],\"ePK91l\":[\"Bearbeiten\"],\"eYlnXt\":[\"Keine Images gefunden\"],\"eh/k36\":[\"Regeltyp auswählen...\"],\"ekCRTP\":[\"Abgelehnt\"],\"eks7oA\":[\"Port-ID\"],\"eu70nA\":[\"Läuft ab um (UTC)\"],\"eyRsaH\":[\"Root\"],\"ezT9KW\":[\"Container-Synchronisierung\"],\"f+Uq1E\":[\"Neuer Objektname\"],\"f0cwjH\":[\"Objekt-Eigenschaften\"],\"fCPhho\":[\"Ein oder mehrere Objekte konnten nicht gelöscht werden:\"],\"fIvd7X\":[\"Fehler beim Löschen der Images\"],\"fJpv9x\":[\"Fehler beim Deaktivieren der Images\"],\"fj5byd\":[\"keine Angabe\"],\"fxnDd7\":[\"Fehler beim Generieren der temporären URL: \",[\"generalError\"]],\"fzfAAa\":[\"Ingress\"],\"g+Jead\":[\"IPv6\"],\"g1IxCo\":[\"RAM muss eine ganze Zahl ≥ 128 MB sein.\"],\"g3BSCe\":[\"Swap-Disk muss eine ganze Zahl ≥ 0 sein.\"],\"g3UF2V\":[\"Akzeptieren\"],\"g8Yxlg\":[\"Temporäre URL für \\\"\",[\"objectName\"],\"\\\" wurde in die Zwischenablage kopiert.\"],\"g9m7gK\":[\"ACL-Einträge steuern, wer von diesem Container lesen oder in ihn schreiben kann. Mehrere Einträge werden durch Kommas getrennt. Änderungen werden sofort nach dem Speichern wirksam.\"],\"gFKJBP\":[\"Ordnername darf keine führenden oder nachfolgenden Leerzeichen haben\"],\"gGdfWx\":[\"Der Compute-Dienst ist für dieses Projekt derzeit nicht verfügbar. Bitte versuchen Sie es später erneut.\"],\"gHTJc/\":[\"Object Storage\"],\"gMYsdZ\":[\"Auf \\\"Geteilt\\\" setzen\"],\"gU7JFm\":[\"Sicherheitsgruppenregel wird erstellt...\"],\"gYe+hC\":[\"Zum Hochladen klicken\"],\"go0J2x\":[\"Fehler beim Kopieren der temporären URL in die Zwischenablage\"],\"grs4+e\":[\"Compute-Übersicht\"],\"gztCjq\":[\"Die angegebene Projekt-ID ist ungültig. Bitte überprüfen Sie Ihre Eingabe.\"],\"h3P8z+\":[\"Das Löschen des Flavor ist fehlgeschlagen. Bitte versuchen Sie es erneut\"],\"h47p9L\":[\"—\"],\"h8h6oz\":[\"Images deaktiviert\"],\"h99+4y\":[\"Floating IP zuweisen\"],\"hH3kDo\":[\"Image-Details werden geladen...\"],\"hHL/wm\":[\"Image-Instanz \\\"\",[\"imageName\"],\"\\\" wurde aktualisiert\"],\"hLp49h\":[\"<0>\",[\"deleteWord\"],\"</0> zur Bestätigung eingeben:\"],\"hPz54a\":[\"Fehler beim Herunterladen\"],\"hQr1Cr\":[\"Image deaktivieren\"],\"hXUWyd\":[\"Wert ist erforderlich\"],\"hYgDIe\":[\"Erstellen\"],\"he3ygx\":[\"Kopieren\"],\"he4q+i\":[\"z. B. b90f9c4bc76140e18540b2cec1299e2a\"],\"hgpMHD\":[\"Gesamtspeicher\"],\"hkjZ7P\":[\"Fehler beim Aktualisieren der Sichtbarkeit für \\\"\",[\"imageName\"],\"\\\":\"],\"hrBow7\":[\"Netzwerk-ID\"],\"i0qMbr\":[\"Startseite\"],\"i30J2U\":[\"Keine Projekte gefunden\"],\"i41Xuw\":[\"Feste IP-Adresse auswählen\"],\"i5MEDc\":[\"Fehler beim Verschieben des Objekts: \",[\"copyErrorMessage\"]],\"i6/ygf\":[\"Eine Eigenschaft mit diesem Schlüssel existiert bereits\"],\"i9TIyi\":[\"Entfernte Sicherheitsgruppe\"],\"i9qiyR\":[\"Läuft ab in\"],\"iH8pgl\":[\"Zurück\"],\"igVDFt\":[\"Anhängen\"],\"iqUvrS\":[\"Benutzer C/D/I\"],\"izMhIO\":[\"Benutzer \",[\"userId\"],\" (beliebiges Projekt)\"],\"j9hkgJ\":[\"Regeln\"],\"jBIkmi\":[\"QCOW2, Raw, VMDK, VHD, VHDX, VDI, AMI, ARI, AKI, ISO, PLOOP\"],\"jIPNJG\":[\"Grundlegende Informationen\"],\"jK6wqe\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" und \",[\"deletedCount\"],\" Objekt wurden dauerhaft gelöscht.\"],\"jKopCP\":[\"Netzwerk & Routing\"],\"jMc/mo\":[\"Beim Erstellen von Metadata ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"jNzyQo\":[\"Objekt-Versionierung\"],\"jPxavx\":[\"Security Group konnte nicht aktualisiert werden\"],\"jS4B2+\":[\"Container-Name stimmt nicht überein\"],\"jSG7wx\":[\"Bitte geben Sie eine gültige Anzahl von Minuten größer als 0 ein\"],\"jhU93c\":[\"Das Image \\\"\",[\"imageName\"],\"\\\" konnte nicht erstellt werden: \"],\"js24f6\":[\"Ordner löschen:\"],\"jtnAf8\":[\"Geschützte Images (können nicht gelöscht werden)\"],\"jyqLKs\":[\"Dieses Mitglied hat bereits Zugriff auf dieses Image.\"],\"k0vAWv\":[\"Objekt-Anzahl-Quota\"],\"k5nYwm\":[\"vCPU\"],\"k7ENJG\":[[\"rowDisplayName\"],\" vorschauen\"],\"k99j0U\":[\"Upload abbrechen\"],\"kA2lMP\":[\"Externes Netzwerk\"],\"kCLnJG\":[\"Alle leeren\"],\"kGmM/p\":[\"Sie haben keine Berechtigung, Metadata für diesen Flavor zu erstellen.\"],\"kIuDMT\":[\"Konfigurieren Sie die Eingangs- und Ausgangsregeln, die steuern, welcher Datenverkehr für diese Sicherheitsgruppe erlaubt ist.\"],\"kKK8AH\":[\"Verfügbare Quota:\"],\"kQYfgO\":[\"Ein oder mehrere Container konnten nicht geleert werden: \",[\"errorMessage\"]],\"kRfwU+\":[\"Nichts zu tun. Container ist bereits leer.\"],\"kqJVBO\":[\"Sie haben keine Berechtigung, diese Sicherheitsgruppe zu teilen.\"],\"kuYWaD\":[\"Reservierte Schlüssel: web-index, web-listings, quota-count, quota-bytes\"],\"kzkYE6\":[\"Checking S3 credentials...\"],\"l75CjT\":[\"Ja\"],\"lAsm87\":[\"Dieses Image vor dem Löschen schützen\"],\"lBVhQs\":[\"Etwas ist schiefgelaufen.\"],\"lN/Z9n\":[\"Security Group-Name ist erforderlich\"],\"lN3xvy\":[\"Regel löschen\"],\"lQ3EIe\":[\"Maximale Löschvorgänge pro Anfrage\"],\"lWTy+Y\":[\"Image kann nicht erstellt werden\"],\"lWxDDh\":[\"Flavor Name\"],\"lZvIXd\":[\"Die Beschreibung darf höchstens 255 Zeichen lang sein.\"],\"lhIa6x\":[\"Fehler beim Laden der Extra-Specs. Bitte versuchen Sie es erneut.\"],\"lq/mBZ\":[\"Objekt-Informationen werden geladen...\"],\"lw1412\":[\"Sie wurden aufgrund von Inaktivität abgemeldet.\"],\"lxentK\":[\"Ein unerwarteter Fehler ist aufgetreten\"],\"m16xKo\":[\"Hinzufügen\"],\"m6X3ro\":[\"Gruppenname\"],\"mQSO1Y\":[\"Port Forwarding\"],\"mSLePW\":[\"Sie haben keine Berechtigung, auf Flavors für dieses Projekt zuzugreifen.\"],\"mSfwLL\":[\"Projekt-ID\"],\"mYnJeY\":[\"Der Text muss “release” in Kleinbuchstaben entsprechen.\"],\"mpby9d\":[\"Support kontaktieren\"],\"mqljvE\":[\"Metadaten kopieren\"],\"mvz5Eo\":[\"URL für öffentlichen Zugriff\"],\"mxPfpY\":[\"Neuen Ordner hier erstellen\"],\"mzI/c+\":[\"Herunterladen\"],\"n0ZttO\":[\"Root Disk (GiB)\"],\"n1ekoW\":[\"Einloggen\"],\"n1gB0L\":[\"Floating IP \",[\"floating_ip_address\"],\" bearbeiten\"],\"n22YIM\":[\"Beschreibung bearbeiten\"],\"n2IuBI\":[\"Eine temporäre URL gewährt zeitlich begrenzten Lesezugriff auf dieses Objekt ohne Authentifizierung. Jeder mit dem Link kann es herunterladen, bis es abläuft.\"],\"n3eQzA\":[\"Diese Eigenschaft ist reserviert und kann nicht geändert werden\"],\"n46oLW\":[\"Fehler beim Entfernen des Mitglieds\"],\"n9jJG6\":[\"Mitgliedszugriff entfernen\"],\"nETBrc\":[\"Egress\"],\"nLvo6K\":[\"Details der RBAC-Richtlinie:\"],\"nUuaq8\":[\"Fehler beim Aktualisieren des Containers: \",[\"errorMessage\"]],\"nW/hX9\":[\"Allgemeine Image-Daten\"],\"nZbdB+\":[\"Upload abgebrochen\"],\"ne/GWZ\":[\"Innerhalb eines Projekts werden Objekte in Containern gespeichert. Container sind der Ort, an dem Sie Zugriffsberechtigungen und Quotas definieren.\"],\"neiJm0\":[\"Flavors\"],\"nkpZyk\":[\"Container \\\"\",[\"containerName\"],\"\\\" war bereits leer.\"],\"nnxwBn\":[\"Es gibt keine Regeln für diese Sicherheitsgruppe\"],\"ntNlXu\":[\"Zugriff auflisten\"],\"o/VDOG\":[\"Image kann nicht gelöscht werden\"],\"o6M6l0\":[\"Security Group konnte nicht erstellt werden\"],\"oDkgME\":[\"Sie sind nicht berechtigt, Flavors zu erstellen. Bitte melden Sie sich erneut an.\"],\"oEGiW3\":[\"Wird hochgeladen... \",[\"progressPct\"],\"%\"],\"ocUvR+\":[\"False\"],\"odVI9Y\":[\"Container gelöscht\"],\"olfSYj\":[\"Zugriffskontrolle aktualisiert\"],\"onHi/J\":[\"Es wird aus Ihrer Image-Liste entfernt.\"],\"p4nMut\":[\"Swap (MiB)\"],\"p6CSHM\":[\"Objekte löschen\"],\"p7DzCB\":[\"Fehler beim Aktualisieren der Zugriffskontrolle\"],\"pFg+7w\":[\"Aktualisiert:\"],\"pOPvlj\":[\"Bereits aktiv (wird übersprungen)\"],\"pU25+T\":[\"Upload von \\\"\",[\"objectName\"],\"\\\" wurde abgebrochen.\"],\"pbzA+s\":[\"Optionale Beschreibung\"],\"pebLmQ\":[\"Zugriff für \",[\"memberIdDisplay\"],\" entfernen\"],\"plnnns\":[[\"successCount\"],\" Image(s) gelöscht, aber \",[\"failedCount\"],\" Image(s) konnten nicht gelöscht werden.\"],\"poCbZw\":[\"ACLs werden geladen...\"],\"podzPY\":[\"Projekt-ID\"],\"psPHye\":[\"Geteiltes Image akzeptieren\"],\"pubQie\":[\"Value eingeben\"],\"q0Rla3\":[\"Projektzugriff hinzufügen\"],\"q44uUq\":[\"Container teilweise geleert\"],\"q5sTNZ\":[\"<0>Kein Temp-URL-Schlüssel konfiguriert.</0> Ein temporärer URL-Schlüssel muss auf Account- oder Container-Ebene gesetzt werden, bevor temporäre URLs generiert werden können. Wenden Sie sich an Ihren Administrator, um <1>X-Account-Meta-Temp-URL-Key</1> oder <2>X-Container-Meta-Temp-URL-Key</2> zu konfigurieren.\"],\"q6K46F\":[\"Schlüssel existiert bereits\"],\"q88/6A\":[\"Fehler beim Erstellen des Ordners\"],\"qAkkjP\":[\"Maximale Objekt-Namenslänge\"],\"qEDO1j\":[\"Dies ist ein <0>dynamisches großes Objekt</0>. Nur das Manifest wird gelöscht — die zugehörigen Segment-Objekte (gespeichert unter dem Manifest-Präfix) werden <1>nicht</1> automatisch entfernt und müssen separat gelöscht werden.\"],\"qFDA8L\":[\"Zugriff auf Image ablehnen\"],\"qJb6G2\":[\"Erneut versuchen\"],\"qQ1QBh\":[\"Hardware-Spezifikationen\"],\"qST5TS\":[\"Fehler – Image-Details\"],\"qUlxA+\":[\"Ordner \\\"\",[\"folderName\"],\"\\\" wurde dauerhaft gelöscht.\"],\"qaAo9Y\":[\"Es ist ein Serverfehler beim Erstellen des Flavors aufgetreten. Bitte versuchen Sie es später erneut.\"],\"qh5W8q\":[\"Richtlinie entfernen\"],\"qs+BrU\":[\"Sie haben keine Berechtigung, Projektzugriffe von diesem Flavor zu entfernen.\"],\"qtoOYG\":[\"Kein Limit\"],\"quU9wK\":[[\"failedCount\"],\" von \",[\"totalCount\"],\" Image(s) konnten nicht deaktiviert werden. Einige Images sind möglicherweise bereits deaktiviert oder in einem ungültigen Zustand.\"],\"qvF2D8\":[\"Keine Images verfügbar. Filtern Sie neu oder erstellen Sie ein neues Image.\"],\"qxxo7y\":[\"Keine Richtlinien entsprechen Ihrer Suche\"],\"qyNaF7\":[\"Geben Sie einen Zeitstempel wie \\\"YYYY-MM-DD HH:mm:ss\\\" ein, um die automatische Löschung zu planen\"],\"qzIZOL\":[\"Ungültiges Dateiformat. Unterstützte Formate: \",[\"supportedFileFormats\"]],\"qzhUb9\":[\"Erste \",[\"maxOptions\"],\" von \",[\"totalCount\"],\" werden angezeigt — verfeinern Sie Ihre Suche, um die Ergebnisse einzugrenzen\"],\"r5SQFW\":[\"Container-Name muss \",[\"maxContainerNameLength\"],\" Zeichen oder weniger haben\"],\"r9Aac8\":[\"Ephemeral Disk\"],\"rAtQcX\":[\"Sie können das Objekt umbenennen, indem Sie den Namen hier ändern.\"],\"rD9yV1\":[\"Zu deaktivierende Images:\"],\"rIe0oV\":[\"Fehler beim Hinzufügen des Zugriffs für das Project zum Flavor. Bitte versuchen Sie es erneut.\"],\"rIi6x4\":[\"Der Flavor konnte nicht gefunden werden. Möglicherweise wurde er bereits gelöscht.\"],\"rJe6vw\":[\"7 Tage\"],\"rbuO5A\":[\"Diese Sicherheitsgruppe ist bereits mit dem angegebenen Projekt geteilt.\"],\"rcBt6T\":[\"Failed to create credential: \",[\"errorMessage\"]],\"rdUucN\":[\"Vorschau\"],\"rhaNn7\":[\"Container werden geladen...\"],\"riR9oD\":[\"Hinweis: Für <0>statische und dynamische große Objekte</0> werden nur die Manifeste gelöscht — ihre Segmente außerhalb dieses Ordnerpräfixes sind nicht betroffen.\"],\"rlgAtt\":[\"Das Objekt wird zu diesem Pfad verschoben. Navigieren Sie oben durch die Ordner, um das Ziel zu ändern.\"],\"rp0Bd0\":[\"Compute\"],\"rrjuul\":[\"Weitere Details finden Sie in der <0>Dokumentation</0>.\"],\"rvXsSb\":[\"Der Projektzugriff für \\\"\",[\"tenantIdToRemove\"],\"\\\" wurde erfolgreich entfernt.\"],\"rwBVXS\":[\"Zu löschende Images (\",[\"deletableCount\"],\")\"],\"ryf/ee\":[\"Images aktiviert\"],\"ryxYVo\":[\"Zu deaktivierende Images (\",[\"activeCount\"],\")\"],\"s/s1lz\":[\"Jeder Benutzer kann eine HEAD- oder GET-Operation auf dem Container ausführen, sofern der Benutzer auch Lesezugriff auf Objekte hat. Es ist kein Token erforderlich.\"],\"s2ubkU\":[\"Flavor ID\"],\"s4Vnq2\":[\"Wird geleert...\"],\"sNVNmf\":[\"MAC-Adresse\"],\"sPFHpI\":[\"Disk\"],\"sSNyf3\":[\"Willkommen beim <0>Aurora-Dashboard</0>, Ihrer Cloud-Management-Lösung der nächsten Generation. Wir sind bestrebt, die Art und Weise, wie Sie mit Ihrer Cloud-Infrastruktur interagieren und diese verwalten, zu vereinfachen. Mit Effizienz, Skalierbarkeit und Benutzerfreundlichkeit im Kern konzipiert, ermöglicht Aurora Ihnen, Prozesse zu optimieren und das volle Potenzial Ihrer Cloud-Ressourcen auszuschöpfen.\"],\"sWBLli\":[\"Eigenschaft hinzufügen\"],\"sXd+qS\":[\"Eigenschaften von \\\"\",[\"objectName\"],\"\\\" wurden erfolgreich aktualisiert.\"],\"sa4CV6\":[\"Alle Benutzer aus Projekt \",[\"projectId\"]],\"shKIZu\":[\"Zu aktivierende Images (\",[\"deactivatedCount\"],\")\"],\"sheDTJ\":[\"Bitte beachten Sie: Für <0>dynamische</0> und <1>statische große Objekte</1> werden nur die Manifeste gelöscht. Die zugehörigen Segmente werden nicht gelöscht.\"],\"sihD20\":[\"Images werden geladen...\"],\"sjMCOP\":[\"Zuletzt geändert\"],\"slWh5C\":[\"Floating IP \",[\"floating_ip_address\"],\" einem Port zuordnen\"],\"sxbP3b\":[\"Objekt-Anzahl\"],\"t/YqKh\":[\"Entfernen\"],\"t0X9+8\":[\"Container-Name\"],\"t1POAD\":[\"Keine benutzerdefinierten Metadaten-Eigenschaften gefunden. Klicken Sie auf \\\"Eigenschaft hinzufügen\\\", um eine zu erstellen.\"],\"t1fq6V\":[\"Der Server hat ein unerwartetes Datenformat zurückgegeben.\"],\"t7ff15\":[\"gültiges Token erforderlich: false\"],\"t95VRV\":[\"Über das Aurora-Dashboard\"],\"tASa/P\":[\"Beim Löschen des Flavors ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"tIrNgH\":[\"Beim Abrufen der Metadata ist ein Serverfehler aufgetreten. Bitte versuchen Sie es später erneut.\"],\"tLerHy\":[\"Ephemeral Disk muss eine ganze Zahl ≥ 0 sein.\"],\"tM5SEI\":[\"ACLs für Container \\\"\",[\"containerName\"],\"\\\" wurden erfolgreich aktualisiert.\"],\"tOkmLM\":[\"Fehler beim Kopieren des Objekts\"],\"tV/Ozb\":[\"Port-Bereich\"],\"tVSmFT\":[\"Weitere werden geladen...\"],\"tX5yOZ\":[\"Neuer Ordner\"],\"tasfos\":[\"entfernen\"],\"tbwGSx\":[\"Minimale Festplatte (GB)\"],\"tejJLY\":[\"Floating IP wird zugeordnet...\"],\"tfAKBU\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht hochgeladen werden: \",[\"errorMessage\"]],\"tfDRzk\":[\"Speichern\"],\"tfxu04\":[\"Zugriff entfernen für \",[\"tenantId\"]],\"thHAVL\":[\"Akzeptierte Images\"],\"tiflqy\":[\"Image kann nicht reaktiviert werden\"],\"tlfxPP\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht kopiert werden: \",[\"errorMessage\"]],\"tmpGvt\":[\"production, linux\"],\"u+VWhB\":[\"In die Zwischenablage kopiert!\"],\"u2xIeO\":[\"Fehler beim Aktualisieren der ACLs: \",[\"errorMessage\"]],\"u5HztT\":[\"RX/TX Factor\"],\"u77/s4\":[\"Floating IPs\"],\"u7En0V\":[\"Metadata hinzufügen\"],\"uAI0yI\":[\"Objekt löschen:\"],\"uAQUqI\":[\"Status\"],\"uLtFAr\":[\"Container \\\"\",[\"containerName\"],\"\\\" konnte nicht aktualisiert werden: \",[\"errorMessage\"]],\"uSdnuQ\":[\"VCPUs müssen eine ganze Zahl ≥ 1 sein.\"],\"ujK/QN\":[\"Objekte werden geladen...\"],\"uly9ET\":[\"Regeldetails:\"],\"up0ZSW\":[\"Fingerprint\"],\"uuKb0T\":[\"Die Beschreibung muss weniger als 65535 Zeichen haben.\"],\"v0hPHE\":[\"Details anzeigen\"],\"v3djpU\":[\"Verschieben/Umbenennen\"],\"v9Dn8m\":[\"Das Aurora-Dashboard ist mehr als nur ein Werkzeug – es ist Ihr Partner bei der Navigation in der Cloud. Egal, ob Sie ein kleines Startup oder ein globales Unternehmen sind, Aurora bietet Ihnen die Flexibilität, Leistung und Einfachheit, die Sie benötigen, um Ihre Ziele zu erreichen.\"],\"vBUQNE\":[\"Das Metadata konnte nicht gefunden werden. Möglicherweise wurde es bereits gelöscht.\"],\"vEkTR9\":[\"Quota\"],\"vH2C/2\":[\"Swap\"],\"vLYu4Z\":[\"Nichts zu tun. Container scheint leer zu sein — die Objekt-Anzahl ist möglicherweise aufgrund einer kürzlichen Operation noch nicht synchronisiert.\"],\"vR4HmN\":[\"Lädt Instanzen...\"],\"vTh35P\":[\"Container erstellen\"],\"vXmL4D\":[\"Legen Sie Ihre Image-Datei hier ab\"],\"vZUKSz\":[\"Floating IP \",[\"floating_ip_address\"],\" trennen\"],\"vbajgL\":[\"Öffentlicher Flavor\"],\"vcQSZh\":[\"Dieser Ordner ist leer — verwenden Sie \\\"Neuer Ordner\\\", um einen zu erstellen.\"],\"vcXmqy\":[\"Netzwerk-Übersicht\"],\"vcvCXq\":[\"Fehler – Flavor-Details\"],\"vg84cD\":[[\"allCount\"],\" Elemente\"],\"vmRPFm\":[\"Sicherheitsgruppe teilen\"],\"vmYyLY\":[\"Entferntes IP-Präfix\"],\"vp5vfW\":[\"1 Stunde\"],\"vpt8cE\":[\"URL generieren\"],\"vrPCbw\":[\"Image-ID\"],\"w3bAcf\":[\"Diese Aktion ist dauerhaft. Die Adresse wird aus Ihrem Projekt entfernt und dem öffentlichen Pool zurückgegeben. Dies kann nicht rückgängig gemacht werden.\"],\"w9+8d7\":[\"Projektzugriff entfernen\"],\"wEfZld\":[\"Neuen Flavor erstellen\"],\"wFaT8w\":[\"Fehler beim Leeren der Container\"],\"wMHvYH\":[\"Value\"],\"wPrtGF\":[\"Key eingeben\"],\"wTg+FY\":[\"Maximale Dateigröße\"],\"wXxPjv\":[\"S3 Object Storage — Setup Required\"],\"wa1Bcq\":[\"Project ID eingeben\"],\"wbqM4L\":[[\"customMinutes\"],\" Minuten\"],\"wcUecy\":[\"Sie haben keine Berechtigung, Metadata für diesen Flavor anzusehen.\"],\"we28Pq\":[\"ACLs-Vorschau ausblenden\"],\"wlQNTg\":[\"Members\"],\"wlUDbB\":[\"Zuletzt aktualisiert: \",[\"formattedDate\"]],\"wrXcuy\":[\"Objektname\"],\"wrk/xj\":[\"Image-Details\"],\"wyIOMP\":[\"Image-Name ist erforderlich\"],\"wzqqS+\":[\"Hauptmerkmale\"],\"x/XQrD\":[\"Beliebiger Dateityp\"],\"x1bK0h\":[\"Mit den aktuellen Suchkriterien sind keine Container verfügbar. Versuchen Sie, Ihren Suchbegriff anzupassen.\"],\"x3T4pq\":[\"Die Container-Metadaten melden Objekte, aber keine wurden aufgelistet. Dies kann eine vorübergehende Synchronisierungsverzögerung sein — bitte warten Sie einen Moment und versuchen Sie es erneut.\"],\"x5l/TK\":[\"Bereits aktiv (wird übersprungen):\"],\"x9AdZ8\":[\"property_key\"],\"xNG/3n\":[\"Floating IP-Adresse\"],\"xNZKYy\":[[\"failedCount\"],\" von \",[\"totalCount\"],\" Image(s) konnten nicht gelöscht werden. Einige Images sind möglicherweise geschützt oder werden verwendet.\"],\"xqhyRT\":[\"Objekt hochgeladen\"],\"xw2UtT\":[\"Neues Image erstellen\"],\"y+KBOY\":[\"z. B. production, linux, ubuntu\"],\"y02Bu1\":[\"Container:\"],\"y0u86k\":[\"Der angeforderte Flavor konnte nicht gefunden werden. Er wurde möglicherweise gelöscht oder Sie haben keinen Zugriff darauf.\"],\"y1GYnY\":[\"\\\"\",[\"objectName\"],\"\\\" konnte nicht verschoben werden: \",[\"errorMessage\"]],\"yPWFWy\":[\"ICMP-Typ\"],\"yTtJTy\":[\"Image-Metadaten bearbeiten\"],\"yYxB17\":[\"Alle Filter löschen\"],\"ylfbpz\":[\"Der Key für die Extra-Spec ist erforderlich und darf nicht leer sein.\"],\"yp0UjB\":[\"Ethertype\"],\"yqPflB\":[\"... und \",[\"hiddenCount\"],\" weitere\"],\"yu9G3x\":[\"Security Group bearbeiten\"],\"yz7wBu\":[\"Schließen\"],\"z+zpLP\":[\"gültiges Token erforderlich: true\"],\"z1JceR\":[\"Zurück zu Floating IPs\"],\"z45o5B\":[\"Objekt-Anzahl\"],\"z9NAjZ\":[\"Objekt gelöscht\"],\"zCD96i\":[\"Sie sind nicht berechtigt, Flavor-Details anzuzeigen. Bitte melden Sie sich erneut an.\"],\"zDS0JC\":[\"Der Name muss 2-50 Zeichen lang sein.\"],\"zWb/Nn\":[\"Maximale Header-Größe\"],\"zc5dcw\":[\"Anmeldung fehlgeschlagen. Bitte überprüfen Sie Ihre Anmeldeinformationen und versuchen Sie es erneut.\"],\"zga9sT\":[\"OK\"],\"zhM8FP\":[\"Zugriff für einen Benutzer aus einem anderen Projekt gewähren.\"],\"zm7+/D\":[\"Sie sind dabei, <0>\",[\"activeCount\"],\" Image(s)</0> zu deaktivieren. Deaktivierte Images können nicht zum Starten neuer Instances verwendet werden.\"],\"zwBp5t\":[\"Privat\"]}")as Messages;
