+++ apps/aurora-portal/src/locales/en/messages.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+0B+ue\":[\"Projects\"],\"+9CXS9\":[\"Deactivate Images\"],\"+Jcye3\":[\"Key Name\"],\"+Lt5cp\":[\"You are not authorized to add tenant access. Please log in again.\"],\"+NwLgN\":[\"Activating this image will allow it to be used to launch new instances again.\"],\"+Nx1wc\":[\"Failed to load Floating IPs\"],\"+OEi73\":[\"Object Storage (Swift)\"],\"+YQ9qu\":[\"Container: \",[\"containerName\"]],\"+nQTmZ\":[\"This tenant does not have access to the flavor.\"],\"+p6nHr\":[\"Failed to load object metadata: \",[\"metadataErrorMessage\"]],\"+zy2Nq\":[\"Type\"],\"/1MfrG\":[\"Could not download \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"/2a/eI\":[\"Loading Flavor...\"],\"/9Squ9\":[\"You don't have permission to view this flavor's details.\"],\"/BZLRP\":[\"To confirm this action, type the word <0>“detach”</0> in the field below.\"],\"/EcdUM\":[\"Your action is required\"],\"/HgF9q\":[\"Sort by\"],\"/InK0O\":[\"Total size\"],\"/LqWNN\":[\"Could not delete \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"/NeNjH\":[\"Container \\\"\",[\"containerName\"],\"\\\" properties were successfully updated.\"],\"/Nmxy/\":[\"No key pairs available.\"],\"/QIkBY\":[\"<0>Secure & Reliable:</0> Your data and operations are safeguarded with enterprise-grade security and robust reliability.\"],\"/Qox3b\":[\"A folder with this name already exists\"],\"/Z2leb\":[\"No containers found.\"],\"/Z5n1b\":[\"Create folder below:\"],\"/bUiYk\":[\"Router ID\"],\"/eFtWI\":[\"RBAC Policies\"],\"/xnbdQ\":[\"The specified user has access. A token for the user (scoped to any project) must be included in the request.\"],\"01/uUD\":[\"Keep segments (delete manifest only)\"],\"07WXfc\":[\"Server returned unexpected data format for extra specs.\"],\"0BSSYj\":[\"Server error occurred while removing tenant access. Please try again later.\"],\"0Gd0NU\":[\"Shared\"],\"0P2gFy\":[\"The page you are looking for does not exist.\"],\"0WsqO0\":[\"Containers Emptied\"],\"0cVgUw\":[\"Filter by\"],\"0dZfcJ\":[[\"allContainersCount\"],\" containers\"],\"0eY8Mz\":[\"There are no Floating IPs available for this project. Floating IPs allow you to map public IP addresses to instances.\"],\"0j9NA7\":[\"<0>Are you sure?</0> All objects in the container will be deleted. This cannot be undone.\"],\"0kCt7e\":[\"The flavor data provided is invalid. Please check your input.\"],\"0kc0zi\":[\"Server error occurred while deleting the extra spec. Please try again later.\"],\"0o0OhW\":[\"No objects found.\"],\"0p+s6m\":[\"Type: \",[\"typeValue\"],\", Code: \",[\"codeValue\"]],\"0u9jhd\":[\"Detaching this Floating IP will remove its association with the current port. The instance will no longer be reachable through this address.\"],\"16085O\":[\"IP Version\"],\"1H2g6v\":[\"Moving object...\"],\"1NS3nd\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" container\"],\"other\":[\"#\",\" containers\"]}],\" successfully emptied. \",[\"totalDeleted\",\"plural\",{\"one\":[\"#\",\" object\"],\"other\":[\"#\",\" objects\"]}],\" deleted in total.\"],\"1RwosK\":[\"Target project ID is required\"],\"1UzENP\":[\"No\"],\"1VDqZj\":[\"<0>Future-Ready:</0> Aurora is designed to evolve with the latest trends in cloud technology, ensuring your solution is always cutting-edge.\"],\"1iQtS2\":[\"Showing first \",[\"actualObjectCount\"],\" of \",[\"total\"],\" objects\"],\"1iUuTT\":[\"Your session has expired. Please log in again.\"],\"1ojTVo\":[\"Select a DNS domain.\"],\"1pGUZa\":[\"Session expires in\"],\"1pdLQw\":[\"Image not found\"],\"1rLu3+\":[\"Could not empty container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"1rPB1p\":[\"The flavor or tenant could not be found. Please verify they exist.\"],\"1t/NnN\":[\"Reject\"],\"1zZ1IK\":[\"Hi\"],\"20E+79\":[\"You need to login to access this page.\"],\"20Kpaw\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" has been successfully deleted.\"],\"20axE5\":[\"Shared by Project\"],\"23wBCX\":[\"Public Read Access\"],\"2G6hLq\":[\"Delete \",[\"specKey\"]],\"2Inn83\":[\"Bulk upload of archive files\"],\"2TtIL2\":[\"Stored as X-Object-Meta-* headers. Keys are case-insensitive.\"],\"2cJIlz\":[\"Floating Network ID\"],\"2d/OiW\":[\"Enter your username\"],\"2dnZwV\":[\"Could not delete folder \\\"\",[\"folderName\"],\"\\\": \",[\"errorMessage\"]],\"2gH+i8\":[\"You are not authorized to delete flavors. Please log in again.\"],\"2lq0gq\":[\"<0>Properties of</0> <1>\",[\"displayName\"],\"</1>\"],\"2mbisJ\":[\"Metadata \\\"\",[\"trimmedKey\"],\"\\\" has been added successfully.\"],\"2pnrGl\":[\"Expected format: YYYY-MM-DD HH:MM:SS\"],\"2q/Q7x\":[\"Visibility\"],\"2ysnjX\":[\"<0>Enhanced Productivity:</0> By reducing operational complexity, Aurora helps your team focus on what truly matters—innovating and driving business success.\"],\"2zceEg\":[\"This action cannot be undone. The image will be permanently deleted.\"],\"33F2A+\":[\"Type container name to confirm\"],\"3AUpb4\":[\"Delete All (\",[\"selectedCount\"],\")\"],\"3Qn0me\":[\"Add Member\"],\"3n+vCm\":[\"Custom duration (minutes)\"],\"3nWqQW\":[\"You are not authorized to view extra specs. Please log in again.\"],\"3nh/7E\":[\"If checked, this flavor will be available to all tenants. If unchecked, access must be explicitly granted to specific tenants.\"],\"3oChIh\":[\"<0>Unified Cloud Management:</0> Consolidates all your cloud assets into one intuitive interface.\"],\"3oc18/\":[\"Private flavors could not be loaded. You may be seeing an incomplete list.\"],\"3q1GLx\":[\"Pending file upload...\"],\"3x7Sws\":[\"Loading Security Group Details...\"],\"4/axPe\":[\"Check out our source code on GitHub!\"],\"47eI0x\":[\"Description must be at least 1 character.\"],\"4EZrJN\":[\"Rules\"],\"4O2AH3\":[\"Member \\\"\",[\"memberIdToRemove\"],\"\\\" has been removed successfully.\"],\"4fh0Wj\":[\"Boot size\"],\"4fvDRe\":[\"Images to activate:\"],\"4fvcmm\":[\"Object will be uploaded as: <0>\",[\"selectedObjectName\"],\"</0>\"],\"4h3Eyf\":[\"\\\"\",[\"objectName\"],\"\\\" was successfully uploaded.\"],\"4kjaAc\":[\"No server groups available.\"],\"4mbrAq\":[\"1 minute\"],\"4opp4r\":[\"Security Groups\"],\"4pOfUd\":[\"Our Mission\"],\"4t33sh\":[\"Failed to Update Object\"],\"4uXhtt\":[\"CIDR\"],\"4utWB4\":[\"Server Role:\"],\"5/wyf8\":[\"Enter a floating IP\"],\"56IxdF\":[\"Failed to load container objects: \",[\"errorMessage\"]],\"5BLR6Q\":[\"IPv4\"],\"5Hdka/\":[[\"allContainersCount\"],\" container\"],\"5JDSvn\":[\"Max meta value length\"],\"5M4Te3\":[\"DNS\"],\"5MF8U2\":[\"Failed to Update Container\"],\"5W2i10\":[\"Manage OpenStack with \"],\"5Yrl6N\":[\"Loading Server Groups...\"],\"5aNQ3F\":[\"\\\"\",[\"objectName\"],\"\\\" was successfully copied to \",[\"destination\"],\".\"],\"5g7owI\":[\"Updating Floating IP...\"],\"5y3O+A\":[\"Deactivate\"],\"6+7EwD\":[\"Serve objects as index when file name is:\"],\"6+OdGi\":[\"Protocol\"],\"6/xipy\":[\"Container Format\"],\"644xgx\":[\"Protected\"],\"6BDqha\":[\"Limits\"],\"6CDYXS\":[\"Static website serving\"],\"6GBt0m\":[\"Metadata\"],\"6H/Lg1\":[\"This is a public image. All users have access to it. Explicit sharing is not needed.\"],\"6KRclz\":[\"Folder Created\"],\"6Kjltl\":[\"Access Control for container:\"],\"6OopEX\":[\"Container Emptied\"],\"6Rnrsz\":[\"Manage Access - \",[\"flavorName\"]],\"6X/9Di\":[\"\\\"\",[\"objectName\"],\"\\\" was successfully moved to \",[\"destination\"],\".\"],\"6YtxFj\":[\"Name\"],\"6jAi8c\":[\"Range\"],\"6luZQA\":[\"Object Moved\"],\"6oolxV\":[\"This extra spec keys already exist. Please use different keys.\"],\"6qzsuS\":[\"Write ACLs\"],\"6sxz+g\":[\"Port Name\"],\"6w+VnM\":[\"Container Created\"],\"6z9W13\":[\"Restart\"],\"76RKuS\":[\"ICMP Code\"],\"78+riR\":[\"You are not authorized to remove tenant access. Please log in again.\"],\"7AfIPZ\":[\"Floating Network\"],\"7BpykL\":[\"Failed to create extra specs. Please try again.\"],\"7L01XJ\":[\"Actions\"],\"7NC3vm\":[\"Subnet\"],\"7NSdfG\":[\"Emptying container \",[\"progressCurrent\"],\" of \",[\"progressTotal\"],\", please wait...\"],\"7Q24LN\":[\"Policy\"],\"7T1fHv\":[\"Failed to remove member \\\"\",[\"memberIdToRemove\"],\"\\\"\"],\"7UlHhT\":[\"Metadata \\\"\",[\"keyToDelete\"],\"\\\" has been deleted successfully.\"],\"7XQ3QJ\":[\"Denied referrer: \",[\"host\"]],\"7ZnTL8\":[\"Failed to update object: \",[\"mutationErrorMessage\"]],\"7a4DvD\":[\"No servers available.\"],\"7d1a0d\":[\"Public\"],\"7flw0l\":[\"Tenant access for \\\"\",[\"trimmedTenantId\"],\"\\\" has been added successfully.\"],\"7sMeHQ\":[\"Key\"],\"88kg0+\":[\"Created At\"],\"8AriEH\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been created\"],\"8TSI9h\":[\"Deactivating this image will prevent it from being used to launch new instances. Existing instances will not be affected.\"],\"8Tg/JR\":[\"Custom\"],\"8ZOb7O\":[[\"numberDeleted\"],\" object was permanently deleted.\"],\"8ZsakT\":[\"Password\"],\"8c3/77\":[\"Max meta name length\"],\"8jLXs3\":[\"Versioned writes\"],\"8s0tOH\":[\"You don't have permission to add tenant access to this flavor.\"],\"8t1+HU\":[\"Deactivated \",[\"successCount\"],\" image(s), but \",[\"failedCount\"],\" image(s) could not be deactivated.\"],\"8wdCNd\":[\"tcp, udp, icmp, or protocol number\"],\"8zAn1f\":[\"Failed to delete flavor. Please try again.\"],\"98Fs4G\":[\"Creating image...\"],\"9J93Xr\":[\"Container name cannot contain slashes\"],\"9SX0bO\":[\"The image \\\"\",[\"imageName\"],\"\\\" could not be updated: \"],\"9X8lAk\":[\"Allocate\"],\"9doWrf\":[\"Failed to add member\"],\"9dsDHD\":[\"The image \\\"\",[\"imageId\"],\"\\\" could not be re-activated: \",[\"message\"]],\"9iz2XW\":[\"Unable to Update Image\"],\"9njIiV\":[\"Failed to Activate Images\"],\"9rz81C\":[\"Device ID\"],\"9v5VLp\":[\"No custom properties defined\"],\"9vSW3U\":[\"Delete Recursively\"],\"9x6EkK\":[\"This is a public flavor. All tenants have access to it.\"],\"A7CVME\":[\"Select disk format first\"],\"AB4Tnl\":[\"Please select a file to upload\"],\"AGXLLY\":[\"Unable to Upload Image File\"],\"AJRhSM\":[\"Root Disk must be an integer ≥ 0.\"],\"AN0DBJ\":[\"Press Enter to add\"],\"AX9Juz\":[\"ID must only contain alphanumeric characters, hyphens, underscores, and dots.\"],\"AZyHwC\":[\"Must be a valid IPv4 or IPv6 address (for example: 172.24.4.228 or 2001:db8::1).\"],\"Ac6dy9\":[\"Type name\"],\"AdtLNV\":[\"Ensure ACL entries are valid — correct project IDs, user IDs, and formats are your responsibility. Invalid entries may silently grant or deny unintended access.\"],\"AeXO77\":[\"Account\"],\"Afh/Lb\":[\"Select destination folder\"],\"AlbUVn\":[\"<0>Optimized Scalability:</0> Built for businesses of all sizes, Aurora grows with you, supporting simple environments and intricate multi-cloud setups alike.\"],\"Alx2/L\":[\"Open in new tab\"],\"AuQtzx\":[\"Must be a non-negative integer\"],\"AxZkIr\":[\"Disk (GiB)\"],\"B2Czeb\":[\"Min. RAM\"],\"B2i9cQ\":[\"Objects to be deleted (\",[\"totalCount\"],\")\"],\"B3toQF\":[\"Objects\"],\"B4Jzm7\":[\"Ceph\"],\"BCJPTn\":[\"Grant access to all users from that project.\"],\"BCXapL\":[\"Failed to load container properties: \",[\"errorMessage\"]],\"BJt+PJ\":[\"Failed to Delete Container\"],\"BMTd81\":[\"This action cannot be undone. The target project will lose access to this security group immediately.\"],\"BOQYRn\":[\"Loading Key Pairs...\"],\"BP4Fwj\":[\"Error Loading Objects: \",[\"errorMessage\"]],\"BSaBkZ\":[\"Objects — \",[\"containerName\"]],\"BYH/2L\":[\"Unable to Deactivate Image\"],\"BZpsYm\":[\"Failed to load containers: \",[\"errorMessage\"]],\"BgMp/T\":[\"Invalid format combination for selected disk format\"],\"BoIAP6\":[\"The ID of the network associated with the floating IP.\"],\"BoPocW\":[\"MD5 checksum\"],\"BrrIs8\":[\"Storage\"],\"CA8ZeT\":[\"Image \\\"\",[\"imageName\"],\"\\\" visibility updated to \",[\"visibility\"]],\"CBFSfX\":[\"Please fix the validation errors below.\"],\"CFMxC8\":[\"Images Deleted\"],\"CMVP7y\":[\"This action cannot be undone. The rule will be permanently deleted.\"],\"CgZxr7\":[\"Min RAM (MB)\"],\"ChOuUj\":[\"Floating IP not found\"],\"Cj2Gtd\":[\"Size\"],\"ClGcRq\":[\"Containers\"],\"Cu6xuZ\":[\"This is a <0>dynamic large object (DLO)</0> manifest. Metadata changes apply to the manifest only — segment objects are not affected.\"],\"CunRry\":[\"Invalid project ID format. Must be 32 hexadecimal characters (e.g., b90f9c4bc76140e18540b2cec1299e2a) or UUID format (e.g., 12345678-1234-1234-1234-123456789abc)\"],\"Cxgv2U\":[\"Min. Disk\"],\"D+Se/Q\":[\"Empty container:\"],\"D/8vkD\":[\"It will appear in your image list.\"],\"D1F789\":[\"<0>Are you sure?</0> The container will be deleted. This cannot be undone.\"],\"D3IRXw\":[\"Detaching Floating IP...\"],\"D7qT9F\":[\"Why Choose Aurora?\"],\"DDRhQm\":[\"Your session has expired.\"],\"DJT9tB\":[\"Account quotas\"],\"DKkOPx\":[\"Extra Specs\"],\"DNVql8\":[\"Full lifecycle management of Floating IPs, including attachment, port association/disassociation, DNS settings, and deletion\"],\"DcMIiu\":[\"Could not update ACLs for container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"Df0YHr\":[\"Update Security Group\"],\"Dh1qvV\":[\"You are about to delete \",[\"deletableCount\"],\" image(s). This action cannot be undone.\"],\"Dia2Ue\":[\"There are no RBAC policies for this security group\"],\"Do5/uH\":[\"The flavor or tenant could not be found. It may have already been removed.\"],\"Dqnh7K\":[\"Specific referrer: \",[\"host\"]],\"Dt5W9T\":[\"Remove RBAC Policy\"],\"DvB4XF\":[\"Drop your file here\"],\"E/QGRL\":[\"Disabled\"],\"E4QYe7\":[\"Suggested Images\"],\"E6nRW7\":[\"Copy URL\"],\"EF2EU9\":[\"Deleting...\"],\"EPMHs9\":[\"You don't have permission to delete flavors in this project.\"],\"EQnVgi\":[\"Flavor service is not available for this project.\"],\"EdQY6l\":[\"None\"],\"Ef7StM\":[\"Unknown\"],\"Enpdmy\":[\"Type <0>remove</0> to confirm:\"],\"EoKe5U\":[\"Domain\"],\"Eq5PsT\":[\"Type \\\"detach\\\" to confirm\"],\"EqSPkP\":[\"Loading Flavors...\"],\"Erlvqg\":[\"Object name cannot have leading or trailing whitespace\"],\"ExLULX\":[\"Image Name\"],\"EztMB8\":[\"Failed to fetch flavors from server.\"],\"F02e8I\":[\"No custom metadata. Click \\\"Add Property\\\" to create one.\"],\"F6YIQe\":[\"Efficient bulk deletion\"],\"FKL6Jv\":[\"e.g. .r:*,.rlistings\"],\"FNcMGM\":[\"Creation Date\"],\"FORsv1\":[\"Find detailed guides and references in our docs.\"],\"FOcBn3\":[\"Detach\"],\"FPsvA8\":[\"Got it!\"],\"FQBaXG\":[\"Activate\"],\"FRtmJJ\":[\"Storage container not found\"],\"FSbpS7\":[\"CPU\"],\"FjONW3\":[\"Error Loading Flavor\"],\"FjPnAE\":[\"Error loading security group\"],\"Flugry\":[[\"progressPct\"],\"%\"],\"FwSyEp\":[\"The specified project does not exist or you don't have permission to share with it.\"],\"Fzrzfe\":[\"Folder name is required\"],\"G6AP+o\":[\"Shared:\"],\"GEgjm+\":[\"Loading Objects...\"],\"GPuCEo\":[\"Leave empty for all types\"],\"GSIPwA\":[\"Temporary URL\"],\"GbKqnI\":[\"Activated \",[\"successCount\"],\" image(s), but \",[\"failedCount\"],\" image(s) could not be activated.\"],\"Gfx1qQ\":[\"Unable to Load Content\"],\"GxkJXS\":[\"Uploading...\"],\"Gyd3No\":[\"No specific tenant access configured for this private flavor. Click \\\"Add Tenant Access\\\" to grant access.\"],\"H+a5j6\":[\"Release\"],\"H0MuMF\":[\"Aurora empowers you with robust tools to streamline your cloud management effortlessly. Experience a seamless way to provision, configure, and scale your resources across OpenStack environments.\"],\"H4Qwmp\":[\"No objects match your search. Try adjusting your search term.\"],\"H7u085\":[\"No projects have access to this image yet. Click \\\"Add Project Access\\\" to grant access.\"],\"HAkrpK\":[\"At Aurora, our mission is to provide a centralized platform that unifies cloud management. We aim to simplify the complexities of provisioning, configuring, and scaling resources across diverse cloud environments while enabling seamless growth for your business.\"],\"HBI6nY\":[\"Need help?\"],\"HBpi4q\":[\"Loading Images...\"],\"HM56Bx\":[\"Creating...\"],\"HNlEFZ\":[\"delete\"],\"HQH8HM\":[\"Could not update \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"HVdrr1\":[\"ANY referrer\"],\"HivZR9\":[\"Create Credential\"],\"Hivb/4\":[\"Server is experiencing issues. Please try again later.\"],\"Hiw1Ha\":[\"No containers found\"],\"HlwgQN\":[\"Object \\\"\",[\"objectName\"],\"\\\" was permanently deleted.\"],\"HuA8iQ\":[\"Allocating Floating IP...\"],\"HxTYrE\":[\"The flavor could not be found. It may have been deleted.\"],\"I5kZVK\":[\"Remote Source\"],\"INUP6f\":[\"<0>Effortless Resource Provisioning:</0> Quickly provision, configure, and deploy resources like servers, networks, and volumes with just a few clicks.\"],\"IOkHLC\":[\"Failed to copy object: \",[\"errorMessage\"]],\"IUwGEM\":[\"Save Changes\"],\"IWF68U\":[\"Storage Overview\"],\"IZ6Mh2\":[\"Enter your domain\"],\"IbYr/u\":[\"Content type\"],\"J4DKSM\":[\"Container format is required\"],\"J6EOll\":[\"Move/Rename object:\"],\"J7+bZb\":[\"Folder Deleted\"],\"J9QcnV\":[\"Successfully activated \",[\"successCount\"],\" of \",[\"totalCount\"],\" image(s)\"],\"J9cmxx\":[\"Failed to update visibility to \",[\"newVisibility\"]],\"JB0bhm\":[\"Get Involved\"],\"JNGYAW\":[\"Container name is required\"],\"JT3I1g\":[\"Delete Flavor\"],\"JeRXll\":[\"This key is reserved and managed separately\"],\"JfWCsP\":[\"Partial Deactivation Success\"],\"Jh4rAZ\":[\"Error loading image\"],\"Jim5X9\":[\"Stateful\"],\"Jj0eys\":[\"View Repository\"],\"JoECY1\":[\"The extra spec data provided is invalid. Please check your input.\"],\"JpZn1L\":[\"Already deactivated (will be skipped)\"],\"JrmKyf\":[\"Failed: \",[\"errorDetails\"]],\"JtHgVz\":[\"Delete Images\"],\"K+e/0e\":[\"RAM (MiB)\"],\"K3bUTE\":[\"Minimum disk must be 0 or greater\"],\"K8Qnlj\":[\"Moving...\"],\"K9eC8x\":[\"This could be due to insufficient permissions or a temporary service issue. Please check your access rights or try refreshing the page.\"],\"KDw4GX\":[\"Try again\"],\"KJC+M7\":[\"Server error occurred while fetching flavor details. Please try again later.\"],\"KOpPMt\":[\"Total size quota\"],\"KSW/GC\":[\"There are no flavors available for this project with the current filters applied. Try adjusting your filter criteria or create a new flavor.\"],\"KZN4Lc\":[\"Delete All\"],\"Km4AGG\":[\"Creating security group...\"],\"KoQP4F\":[\"Server error occurred while adding tenant access. Please try again later.\"],\"KsIM0b\":[\"Boot RAM\"],\"KsnZ3m\":[\"Folder \\\"\",[\"folderName\"],\"\\\" was successfully created.\"],\"KzUd7m\":[\"new-folder-name\"],\"LI8Z2I\":[\"Download \",[\"rowDisplayName\"]],\"LK0pQN\":[\"Disk format is required\"],\"LMdsuJ\":[\"Port (from)\"],\"LQQCas\":[\"Folder \\\"\",[\"folderName\"],\"\\\" and \",[\"deletedCount\"],\" object was permanently deleted.\"],\"LiyMGv\":[\"Enter the Cloud\"],\"Llcakz\":[\"Updated At\"],\"LqMb+g\":[\"To confirm this action, type the word <0>\\\"release\\\"</0> in the field below.\"],\"LtI9AS\":[\"Owner\"],\"Lylr9Z\":[\"Object Copied\"],\"M470oJ\":[\"The flavor could not be found or has no extra specs.\"],\"M5Epeo\":[\"Edit Image Details\"],\"M5RhXF\":[\"Removing...\"],\"M5rEN5\":[\"Session Expired\"],\"M9H+/G\":[\"projects\"],\"MEIAzV\":[\"Unnamed\"],\"MJtNLd\":[\"Images to delete:\"],\"MOug+V\":[\"Enter a tag and press Enter or click Add\"],\"MRB7nI\":[\"Direction\"],\"MXoA/6\":[\"Upload Object\"],\"MXw7Fr\":[\"Server Name\"],\"MZGbkp\":[\"VCPUs\"],\"MbKJNP\":[\"You don't have permission to access flavor access information for this flavor.\"],\"MgZyuJ\":[\"You are about to activate <0>\",[\"deactivatedCount\"],\" image(s)</0>. Activated images will be available for launching new instances.\"],\"MmtQVF\":[\"Invalid value for public flavor setting.\"],\"Mt6sRo\":[\"You are not authorized to access flavor access information. Please log in again.\"],\"MtzSbv\":[\"Object name is required\"],\"MuKU9V\":[\"Failed to load objects: \",[\"errorMessage\"]],\"N2S1rs\":[\"Empty\"],\"N5I2RJ\":[\"Type \\\"release\\\" to confirm\"],\"N5vGcw\":[\"Enter your credentials to access your account\"],\"NH2fsP\":[\"Already deactivated (will be skipped):\"],\"NOdFZR\":[\"Generating...\"],\"NQU1Nn\":[\"Copy container name\"],\"NRMm0E\":[\"This tenant already has access to the flavor.\"],\"NRP2uq\":[\"Share object:\"],\"NRVSdy\":[\"Member ID\"],\"NW4PIb\":[\"Could not create folder \\\"\",[\"folderName\"],\"\\\": \",[\"errorMessage\"]],\"NZJhro\":[\"Object name cannot contain slashes\"],\"Nc7QKU\":[\"Fixed IP Address\"],\"NeUjqc\":[\"Enable file listing\"],\"NixRmA\":[\"Min Disk (GB)\"],\"NlcF/v\":[\"No flavor selected for deletion.\"],\"NopYGU\":[\"Disk Format\"],\"Np28ib\":[\"or drag and drop\"],\"Nu4oKW\":[\"Description\"],\"Nvfd2b\":[\"Versioning is enabled\"],\"O80bQY\":[\"Loading object properties...\"],\"O8tK4v\":[\"Add rule\"],\"ONWvwQ\":[\"Upload\"],\"OR475H\":[\"Network\"],\"OSlLnz\":[\"Image Visibility\"],\"OYHzN1\":[\"Tags\"],\"OZImTR\":[\"Container listing limit\"],\"OaSktR\":[\"Device Owner\"],\"OlmKCg\":[\"A flavor with this ID or name already exists. Please use different values.\"],\"OvEjsP\":[\"Copying...\"],\"Ovofy+\":[\"Release Floating IP \",[\"floating_ip_address\"]],\"OxDN2m\":[\"Failed to create flavor. Please try again.\"],\"OxaeYj\":[\"We are building Aurora Dashboard to serve you better. Your feedback is invaluable in shaping a tool that meets the unique needs of businesses like yours. Stay connected and join us as we redefine cloud management.\"],\"Oxl1UN\":[\"If there is no index file, the URL displays a list of objects in the container.\"],\"PAKSdy\":[\"Enter a floating IP or leave blank to auto-assign one\"],\"PEGvy+\":[\"Minimum RAM must be 0 or greater\"],\"PHt+EV\":[\"Type <0>delete</0> to confirm:\"],\"PIbPRX\":[\"RX/TX Factor must be an integer ≥ 1.\"],\"PLwzWR\":[\"All containers\"],\"PYQUjU\":[\"Failed to load metadata configuration.\"],\"PZnUbs\":[\"Please log in again to continue.\"],\"PgNNGl\":[\"More Actions\"],\"PiH3UR\":[\"Copied!\"],\"PiyQJ/\":[\"No flavors found\"],\"PkfPsB\":[\"Enter the ID of the project you want to share this security group with. You can find project IDs in the account/project switcher or in the Identity service.\"],\"Pkw7J9\":[\"This folder is empty.\"],\"PsEGri\":[\"Ubuntu 22.04 LTS\"],\"PtjzS+\":[\"Associates on the selected port. If the port has multiple IPs, select the desired fixed IP address.\"],\"PzgYM9\":[\"Checksum\"],\"Q1W//7\":[\"No services available for this project.\"],\"Q2xmVl\":[\"Symlinks\"],\"Q9f2QF\":[[\"numberDeleted\"],\" objects were deleted successfully, but some deletions failed.\"],\"QAUa4B\":[\"Enter a single port, or define a range by also filling \\\"Port (to)\\\". \\\"Port (to)\\\" is optional.\"],\"QEtDlS\":[\"Copying object...\"],\"QNHur0\":[\"Failed to load container ACLs: \",[\"errorMessage\"]],\"QQ8wUG\":[\"This action cannot be undone. The flavor will be permanently deleted.\"],\"QV1ZPO\":[\"Key is required\"],\"QWdKwH\":[\"Move\"],\"QYiqYb\":[\"Failed to update access status\"],\"Qb+14I\":[\"This action cannot be undone. The security group will be permanently deleted.\"],\"QetsXP\":[\"Upload failed: \",[\"uploadError\"]],\"Qg4EG6\":[\"Unable to connect to the compute service. Please check your connection and try again.\"],\"QuJSSl\":[\"Failed to create the flavor. Please try again.\"],\"QvqBQa\":[\"Target container\"],\"Qx7DM7\":[\"Capabilities\"],\"QxBGbh\":[\"Protected (will be skipped):\"],\"R0xWJL\":[\"Cannot delete. Container contains objects. Please empty it first.\"],\"R6kcsL\":[\"Must be a valid PQDN or FQDN (alphanumeric and hyphens only, cannot start or end with hyphen).\"],\"R6u5CR\":[\"Failed to activate \",[\"failedCount\"],\" of \",[\"totalCount\"],\" image(s). Some images may already be active or in an invalid state.\"],\"RByeNR\":[\"Your session expired. Please login again.\"],\"RCr0yv\":[\"Failed to load flavor details. Please try again.\"],\"RFDYCD\":[\"Minimum disk size required to boot this image\"],\"RGhYAo\":[\"RAM\"],\"RGrgxg\":[\"Could not delete container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"RGwfoL\":[\"Max meta count\"],\"RNBvdl\":[\"Max SLO segments\"],\"RSFkXF\":[\"Activate Image\"],\"RSMPjT\":[\"You are currently on the dashboard route.\"],\"RSg/pq\":[\"Failed to Delete Object\"],\"RTQFAw\":[\"You are not authorized to create extra specs. Please log in again.\"],\"Rih53k\":[\"Max container name length\"],\"RkXlPZ\":[\"GitHub\"],\"Rlp5zj\":[\"Create Flavor\"],\"S0kLOH\":[\"ID\"],\"S1iTXO\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been deleted\"],\"S3olSf\":[\"No extra specs found. Click \\\"Add Metadata\\\" to create one.\"],\"S5CUKP\":[\"Member ID (project UUID) is required.\"],\"S63NbU\":[\"The image \\\"\",[\"imageId\"],\"\\\" could not be deactivated: \",[\"message\"]],\"S8/j2h\":[\"Failed to Empty Container\"],\"SBGiGm\":[\"Read ACLs\"],\"SCY5an\":[\"Failed to Move Object\"],\"SFo0kK\":[\"All Images\"],\"SIfYq6\":[\"Edit Metadata\"],\"SLEH7X\":[\"Enter DNS name\"],\"STc+7E\":[\"Max containers per extraction\"],\"SU0uxT\":[\"Upload object to:\"],\"SUSS9i\":[\"Container name\"],\"SVLToM\":[\"Type \\\"remove\\\" to confirm\"],\"SXCuo2\":[\"View Docs\"],\"SZw9tS\":[\"View Details\"],\"Sb/VT5\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully emptied. \",[\"deletedCount\"],\" objects deleted.\"],\"SfW/3r\":[\"There are no groups\"],\"Sgz1vJ\":[\"Member \\\"\",[\"trimmedMemberId\"],\"\\\" has been added successfully.\"],\"Smk7M2\":[\"Error loading floating IP\"],\"SuX2Ca\":[\"Basic Info\"],\"SysqAR\":[\"Flavor Details\"],\"T6Gm5y\":[\"Select an external network\"],\"T7mgdd\":[\"Successfully deleted \",[\"successCount\"],\" of \",[\"totalCount\"],\" image(s)\"],\"T8N6oi\":[\"Property Key\"],\"T9Mtpi\":[\"Tenant ID\"],\"TM93nK\":[\"Delete Security Group Rule\"],\"TPMaxo\":[\"Type “release” to confirm\"],\"TQn3hH\":[\"Failed to create image. Please try again.\"],\"TZJiVf\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully emptied. \",[\"deletedCount\"],\" object deleted.\"],\"TfC9O+\":[\"Last modified (UTC)\"],\"TfdeUd\":[\"Failed to delete the extra spec. Please try again.\"],\"TpGxnq\":[\"Enter member ID\"],\"TvY/XA\":[\"Documentation\"],\"Tx4Ym+\":[\"Enter a valid PQDN or FQDN (max 63 characters) to associate with the floating IP. A and PTR records are created automatically.\"],\"TyODHt\":[\"Save Metadata\"],\"U/oahm\":[\"URL Copied\"],\"U2wTy/\":[\"Note: The 'stateful' attribute cannot be changed if this security group is currently in use by one or more ports.\"],\"U4fmHG\":[\"The text must match “detach” in lowercase.\"],\"U6L+P/\":[\"Inactivity Timeout\"],\"U9q4M7\":[\"Back to Security Groups\"],\"UGhVPl\":[\"Object Type\"],\"UJVf0u\":[\"Loading Image...\"],\"UK2mpr\":[\"Generating temporary URL...\"],\"UKwOYH\":[\"Image File\"],\"UO3hJ2\":[\"Temporary URLs\"],\"UQ7Wyv\":[\"Manage Access for Image - \",[\"imageName\"]],\"URmyfc\":[\"Details\"],\"USiuNX\":[\"Container quotas\"],\"UVFHGY\":[\"e.g. PROJECT_ID:USER_ID\"],\"UVSFVV\":[\"Reject Shared Image\"],\"UYSopm\":[\"Minimum RAM (MB)\"],\"UbRKMZ\":[\"Pending\"],\"UdcGJu\":[\"Activate Images\"],\"UiNv/G\":[\"S3 Object Storage requires EC2 credentials (access key + secret key) to authenticate your requests. You need to create credentials before accessing S3 resources.\"],\"Uj+n/2\":[\"Failed to Delete Folder\"],\"UkVkoq\":[\"Leave empty for all codes\"],\"UmQ3/m\":[\"Deactivate Selected\"],\"Uwo8Xw\":[\"This image was shared with you by <0>\",[\"ownerProject\"],\"</0> on \",[\"sharedAt\"],\".\"],\"UztfYZ\":[\"Select port to associate\"],\"V/8B9A\":[\"I confirm that all existing versions will also be deleted\"],\"V/SINY\":[\"Update object\"],\"V1TzeS\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully deleted.\"],\"V66Jih\":[\"Access Status\"],\"V7fN5X\":[\"Copy object:\"],\"V804LY\":[\"Updating security group...\"],\"VCM3KS\":[\"Add Project Access\"],\"VKmlZ+\":[\"Containers to be emptied (\",[\"totalCount\"],\")\"],\"VLI9eO\":[\"Loading Floating IP Details...\"],\"VV1fdg\":[\"Any user has read access to objects. No token is required in the request.\"],\"VaA9mu\":[\"24 hours\"],\"VakxP/\":[\"Failed to Upload Object\"],\"Vg0k6h\":[\"Showing \",[\"filteredCount\"],\" of \",[\"totalCount\"],\" \",[\"itemName\"]],\"Vh/Uj5\":[\"Target path\"],\"Vj8XFg\":[\"Failed to Create Container\"],\"Vl4XTj\":[\"Folder name cannot contain slashes\"],\"Vmojta\":[\"Access status updated to \\\"\",[\"newStatus\"],\"\\\".\"],\"VoxR3s\":[\"Object was copied but could not be deleted from the source: \",[\"deleteErrorMessage\"]],\"Vz+7ZA\":[\"Could not create container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"Vzlopx\":[\"Delete container:\"],\"W0MCSG\":[\"Accept access to image\"],\"W5FkH9\":[\"Enter container name\"],\"W9PZE0\":[\"Objects Deleted\"],\"W9kfjU\":[\"QoS Policy ID\"],\"WCKEqI\":[\"This is a <0>static large object (SLO)</0> manifest. Metadata changes apply to the manifest only — segment objects are not affected.\"],\"WCLyHI\":[\"No Floating IPs found\"],\"WErCZy\":[\"Minimum RAM required to boot this image\"],\"WRZ3Mt\":[\"Loading container properties...\"],\"WYb0Td\":[\"<0>Are you sure?</0> All objects in the selected containers will be permanently deleted. This cannot be undone.\"],\"WYiUDa\":[\"Loading Containers...\"],\"Wbg1jv\":[\"Copy \",[\"text\"],\" to clipboard\"],\"Wca9WC\":[\"Failed to load security groups\"],\"WlpcJv\":[\"DNS Domain\"],\"WoSkGY\":[\"Remote IP\"],\"WrUky8\":[\"Share (Temporary URL)\"],\"WyKwnD\":[\"Store old object versions\"],\"WzVwU0\":[\"Target Project ID\"],\"X2OnDx\":[\"Ephemeral Disk (GiB)\"],\"X70LXS\":[[\"numberDeleted\"],\" objects were permanently deleted.\"],\"XLk16/\":[\"Together, we can unlock the true potential of your cloud infrastructure.\"],\"XYZLy9\":[\"Key contains invalid characters\"],\"XvjC4F\":[\"Saving...\"],\"XwxJJB\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully created.\"],\"XxjLdW\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" container was already empty.\"],\"other\":[\"#\",\" containers were already empty.\"]}]],\"Y+2SDm\":[\"Delete Security Group \\\"\",[\"securityGroupName\"],\"\\\"\"],\"Y1YKad\":[\"Edit Details\"],\"YIix5Y\":[\"Search...\"],\"YNgcgc\":[\"Loading Flavor Details...\"],\"YRexkb\":[\"Object Updated\"],\"YUU0QW\":[\"Flavor ID is required and cannot be empty.\"],\"YZmsaT\":[\"Partial Activation Success\"],\"YiMCKk\":[\"Show ACLs Preview\"],\"Yin3uB\":[\"Releasing Floating IP...\"],\"YjAOtb\":[\"Create Security Group\"],\"YrAy/S\":[\"You don't have permission to delete extra specs for this flavor.\"],\"YsOJlj\":[\"Server error occurred while fetching flavor access information. Please try again later.\"],\"YsrbQh\":[\"Owner Project ID\"],\"YuC9dj\":[\"Associate\"],\"YuGQWb\":[\"Rule Type\"],\"Z/eWPC\":[\"The object will be copied to this path. Navigate folders above to change the destination.\"],\"Z2fZGD\":[\"No project selected\"],\"Z3FXyt\":[\"Loading...\"],\"Z42tfY\":[\"Folders in object storage are virtual — they are created as zero-byte placeholder objects with a trailing slash. The folder will appear once created.\"],\"Z5r9vC\":[\"Partial Delete Success\"],\"Z8lGw6\":[\"Share\"],\"ZAx+d1\":[\"Max meta overall size\"],\"ZAy0zp\":[\"Successfully deactivated \",[\"successCount\"],\" of \",[\"totalCount\"],\" image(s)\"],\"ZUmOzn\":[\"Server returned unexpected data format for flavor details.\"],\"ZcWMT1\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been deactivated\"],\"Zgp2Sm\":[\"No projects available.\"],\"ZhVSpK\":[\"Failed to remove tenant access from flavor. Please try again.\"],\"Zq6Y5u\":[\"DNS name must be at most 63 characters.\"],\"ZvIpwi\":[\"Select a security group...\"],\"Zw49f9\":[\"folder-name\"],\"Zw8Q49\":[\"Security group not found\"],\"a/nTb8\":[\"Create Image\"],\"a12lSo\":[\"Port (to)\"],\"a13wDR\":[\"Type to search containers...\"],\"a3LDKx\":[\"Security\"],\"a4A2uB\":[\"Account listing limit\"],\"a4N/Bg\":[\"Load More\"],\"a7C4YS\":[\"Container Updated\"],\"a88X3d\":[\"<0>Are you sure?</0> Object <1>\\\"\",[\"displayName\"],\"\\\"</1> will be permanently deleted. This cannot be undone.\"],\"aG9OiI\":[\"Sharing Details\"],\"aI8Tgp\":[\"Owning Project ID\"],\"aL1w5Z\":[\"Used\"],\"aOeFR+\":[\"Empty Containers\"],\"aSsVD3\":[\"Public read access is not enabled. Before configuring static website serving, go to <0>Manage Access</0> and enable public read access.\"],\"aTqCTq\":[\"Image file is required\"],\"aV6KPH\":[\"Prevent accidental deletion\"],\"aiqFbS\":[\"<0>Are you sure?</0> The selected objects will be permanently deleted. This cannot be undone.\"],\"an5hVd\":[\"Images\"],\"ao/ZJi\":[\"Deleting folder and all its contents...\"],\"aqagJH\":[\"Unable to Update Image Visibility\"],\"arel2K\":[\"No objects found\"],\"azXlY+\":[\"Access Status:\"],\"b0uU1G\":[\"Store old object versions in container:\"],\"b2BLBa\":[\"Add Security Group Rule\"],\"bISG26\":[\"Failed to fetch flavor access information. Please try again.\"],\"bM1O3m\":[\"Image Instance\"],\"bQBMTH\":[\"This is a <0>static large object</0>. By default, all associated segment objects will also be permanently deleted.\"],\"bRgFkJ\":[\"Failed to upload file \\\"\",[\"fileName\"],\"\\\": \"],\"bYRFNi\":[\"Failed to Delete Objects\"],\"bc67JN\":[\"Custom Properties / Metadata\"],\"bmQLn5\":[\"Add Rule\"],\"bnql/K\":[\"Back to Images\"],\"boJ+Y1\":[\"Create Folder\"],\"boJlGf\":[\"Page Not Found\"],\"bpme7e\":[\"Flavor Not Found\"],\"bwRvnp\":[\"Action\"],\"bwhBhT\":[\"Security Group\"],\"byKna+\":[\"An unexpected error occurred. Please try again.\"],\"bzMKg7\":[\"Accepted\"],\"bzSI52\":[\"Discard\"],\"c+fUtV\":[\"Start typing to search for a container\"],\"c+xCSz\":[\"True\"],\"c1uL4p\":[\"The image \\\"\",[\"imageId\"],\"\\\" could not be deleted: \",[\"message\"]],\"c6b6fz\":[\"Delete Selected\"],\"cCfxH1\":[\"Downloading...\"],\"cJDQIO\":[\"Root Disk\"],\"cPKL6O\":[\"You don't have permission to create flavors in this project.\"],\"cWbW6w\":[\"Manage Access\"],\"cXuXkb\":[\"User \",[\"userId\"],\" from project \",[\"projectId\"]],\"chL5IG\":[\"Community\"],\"cj17eo\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been activated\"],\"cjEOmc\":[\"Share this security group with another project. The target project will be able to view and use this security group, but will not be able to modify or delete it.\"],\"cnGeoo\":[\"Delete\"],\"cpw++p\":[\"Static large object support\"],\"cqQyPB\":[\"Folder name\"],\"d+F6q9\":[\"Created\"],\"d+Ugpw\":[\"<0>Are you sure?</0> Folder <1>\\\"\",[\"folderDisplayName\"],\"\\\"</1> and all objects within it will be permanently deleted. This cannot be undone.\"],\"d/I0J3\":[\"Activate Selected\"],\"d0pLfy\":[\"Failed to delete security group\"],\"dEgA5A\":[\"Cancel\"],\"dFb5Nt\":[\"Id\"],\"dLFiER\":[\"Error Loading Containers: \",[\"errorMessage\"]],\"dOevLB\":[\"Expires in \",[\"selectedPresetLabel\"],\" — at \",[\"expiresAtFormatted\"]],\"dPBJAJ\":[\"Empty All (\",[\"selectedCount\"],\")\"],\"dPj4yB\":[\"Login to Your Account\"],\"dPoCVe\":[\"Type “detach” to confirm\"],\"dTNzBI\":[\"Key must contain at least one alphanumeric character\"],\"dVdc7N\":[\"You are not authorized to delete extra specs. Please log in again.\"],\"dd2ndz\":[\"Entries in ACLs are comma-separated. Examples:\"],\"diFNkW\":[\"Error loading component\"],\"e0NrBM\":[\"Project\"],\"eChIh7\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" has been successfully created.\"],\"eGEHJE\":[\"DNS Name\"],\"eKC+EC\":[\"-\"],\"ePK91l\":[\"Edit\"],\"eYlnXt\":[\"No images found\"],\"eh/k36\":[\"Select a rule type...\"],\"ekCRTP\":[\"Rejected\"],\"eks7oA\":[\"Port ID\"],\"eu70nA\":[\"Expires at (UTC)\"],\"eyRsaH\":[\"Root\"],\"ezT9KW\":[\"Container syncing\"],\"f+Uq1E\":[\"New object name\"],\"f0cwjH\":[\"Object properties\"],\"fCPhho\":[\"One or more objects could not be deleted:\"],\"fIvd7X\":[\"Failed to Delete Images\"],\"fJpv9x\":[\"Failed to Deactivate Images\"],\"fj5byd\":[\"N/A\"],\"fxnDd7\":[\"Failed to generate temporary URL: \",[\"generalError\"]],\"fzfAAa\":[\"Ingress\"],\"g+Jead\":[\"IPv6\"],\"g1IxCo\":[\"RAM must be an integer ≥ 128 MB.\"],\"g3BSCe\":[\"Swap Disk must be an integer ≥ 0.\"],\"g3UF2V\":[\"Accept\"],\"g8Yxlg\":[\"Temporary URL for \\\"\",[\"objectName\"],\"\\\" was copied to clipboard.\"],\"g9m7gK\":[\"ACL entries control who can read from or write to this container. Multiple entries are comma-separated. Changes take effect immediately after saving.\"],\"gFKJBP\":[\"Folder name cannot have leading or trailing whitespace\"],\"gGdfWx\":[\"The compute service is currently unavailable for this project. Please try again later.\"],\"gHTJc/\":[\"Object Storage\"],\"gMYsdZ\":[\"Set to \\\"Shared\\\"\"],\"gU7JFm\":[\"Creating security group rule...\"],\"gYe+hC\":[\"Click to upload\"],\"go0J2x\":[\"Failed to copy the temporary URL to the clipboard\"],\"grs4+e\":[\"Compute Overview\"],\"gztCjq\":[\"The tenant ID provided is invalid. Please check your input.\"],\"h3P8z+\":[\"Failed to delete the flavor. Please try again.\"],\"h47p9L\":[\"—\"],\"h8h6oz\":[\"Images Deactivated\"],\"h99+4y\":[\"Allocate Floating IP\"],\"hH3kDo\":[\"Loading Image Details...\"],\"hHL/wm\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been updated\"],\"hLp49h\":[\"Type <0>\",[\"deleteWord\"],\"</0> to confirm:\"],\"hPz54a\":[\"Failed to Download\"],\"hQr1Cr\":[\"Deactivate Image\"],\"hXUWyd\":[\"Value is required\"],\"hYgDIe\":[\"Create\"],\"he3ygx\":[\"Copy\"],\"he4q+i\":[\"e.g., b90f9c4bc76140e18540b2cec1299e2a\"],\"hgpMHD\":[\"Total Size\"],\"hkjZ7P\":[\"Failed to update visibility for \\\"\",[\"imageName\"],\"\\\": \"],\"hrBow7\":[\"Network ID\"],\"i0qMbr\":[\"Home\"],\"i30J2U\":[\"No projects found\"],\"i41Xuw\":[\"Select a fixed IP address\"],\"i5MEDc\":[\"Failed to move object: \",[\"copyErrorMessage\"]],\"i6/ygf\":[\"A property with this key already exists\"],\"i9TIyi\":[\"Remote Security Group\"],\"i9qiyR\":[\"Expires in\"],\"iH8pgl\":[\"Back\"],\"igVDFt\":[\"Attach\"],\"iqUvrS\":[\"User C/D/I\"],\"izMhIO\":[\"User \",[\"userId\"],\" (any project)\"],\"j9hkgJ\":[\"rules\"],\"jBIkmi\":[\"QCOW2, Raw, VMDK, VHD, VHDX, VDI, AMI, ARI, AKI, ISO, PLOOP\"],\"jIPNJG\":[\"Basic Information\"],\"jK6wqe\":[\"Folder \\\"\",[\"folderName\"],\"\\\" and \",[\"deletedCount\"],\" objects were permanently deleted.\"],\"jKopCP\":[\"Network & Routing\"],\"jMc/mo\":[\"Server error occurred while creating extra specs. Please try again later.\"],\"jNzyQo\":[\"Object versioning\"],\"jPxavx\":[\"Failed to update security group\"],\"jS4B2+\":[\"Container name does not match\"],\"jSG7wx\":[\"Please enter a valid number of minutes greater than 0\"],\"jhU93c\":[\"The image \\\"\",[\"imageName\"],\"\\\" could not be created: \"],\"js24f6\":[\"Delete folder:\"],\"jtnAf8\":[\"Protected images (cannot be deleted)\"],\"jyqLKs\":[\"This member already has access to this image.\"],\"k0vAWv\":[\"Object count quota\"],\"k5nYwm\":[\"vCPU\"],\"k7ENJG\":[\"Preview \",[\"rowDisplayName\"]],\"k99j0U\":[\"Cancel upload\"],\"kA2lMP\":[\"External Network\"],\"kCLnJG\":[\"Empty All\"],\"kGmM/p\":[\"You don't have permission to create extra specs for this flavor.\"],\"kIuDMT\":[\"Configure the ingress and egress rules that control which traffic is allowed for this security group.\"],\"kKK8AH\":[\"Remaining Quota:\"],\"kQYfgO\":[\"One or more containers could not be emptied: \",[\"errorMessage\"]],\"kRfwU+\":[\"Nothing to do. Container is already empty.\"],\"kqJVBO\":[\"You don't have permission to share this security group.\"],\"kuYWaD\":[\"Reserved keys: web-index, web-listings, quota-count, quota-bytes\"],\"kzkYE6\":[\"Checking S3 credentials...\"],\"l75CjT\":[\"Yes\"],\"lAsm87\":[\"Protect this image from being deleted\"],\"lBVhQs\":[\"Something went wrong:\"],\"lN/Z9n\":[\"Security group name is required\"],\"lN3xvy\":[\"Delete Rule\"],\"lQ3EIe\":[\"Max deletes per request\"],\"lWTy+Y\":[\"Unable to Create Image\"],\"lWxDDh\":[\"Flavor Name\"],\"lZvIXd\":[\"Description must be at most 255 characters.\"],\"lhIa6x\":[\"Failed to load extra specs. Please try again.\"],\"lq/mBZ\":[\"Loading object info...\"],\"lw1412\":[\"You have been logged out due to inactivity.\"],\"lxentK\":[\"An unexpected error occurred\"],\"m16xKo\":[\"Add\"],\"m6X3ro\":[\"Group Name\"],\"mQSO1Y\":[\"Port Forwarding\"],\"mSLePW\":[\"You don't have permission to access flavors for this project.\"],\"mSfwLL\":[\"Project ID\"],\"mYnJeY\":[\"The text must match “release” in lowercase.\"],\"mpby9d\":[\"Contact support\"],\"mqljvE\":[\"Copy metadata\"],\"mvz5Eo\":[\"URL for public access\"],\"mxPfpY\":[\"Create new folder here\"],\"mzI/c+\":[\"Download\"],\"n0ZttO\":[\"Root Disk (GiB)\"],\"n1ekoW\":[\"Sign In\"],\"n1gB0L\":[\"Edit Floating IP \",[\"floating_ip_address\"]],\"n22YIM\":[\"Edit Description\"],\"n2IuBI\":[\"A temporary URL grants time-limited read access to this object without requiring authentication. Anyone with the link can download it until it expires.\"],\"n3eQzA\":[\"This property is reserved and cannot be modified\"],\"n46oLW\":[\"Failed to remove member\"],\"n9jJG6\":[\"Remove member access\"],\"nETBrc\":[\"Egress\"],\"nLvo6K\":[\"RBAC Policy Details:\"],\"nUuaq8\":[\"Failed to update container: \",[\"errorMessage\"]],\"nW/hX9\":[\"General Image Data\"],\"nZbdB+\":[\"Upload Cancelled\"],\"ne/GWZ\":[\"Inside a project, objects are stored in containers. Containers are where you define access permissions and quotas.\"],\"neiJm0\":[\"Flavors\"],\"nkpZyk\":[\"Container \\\"\",[\"containerName\"],\"\\\" was already empty.\"],\"nnxwBn\":[\"There are no rules for this security group\"],\"ntNlXu\":[\"Listing access\"],\"o/VDOG\":[\"Unable to Delete Image\"],\"o6M6l0\":[\"Failed to create security group\"],\"oDkgME\":[\"You are not authorized to create flavors. Please log in again.\"],\"oEGiW3\":[\"Uploading... \",[\"progressPct\"],\"%\"],\"ocUvR+\":[\"False\"],\"odVI9Y\":[\"Container Deleted\"],\"olfSYj\":[\"Access Control Updated\"],\"onHi/J\":[\"It will be removed from your image list.\"],\"p4nMut\":[\"Swap (MiB)\"],\"p6CSHM\":[\"Delete Objects\"],\"p7DzCB\":[\"Failed to Update Access Control\"],\"pFg+7w\":[\"Updated:\"],\"pOPvlj\":[\"Already active (will be skipped)\"],\"pU25+T\":[\"Upload of \\\"\",[\"objectName\"],\"\\\" was cancelled.\"],\"pbzA+s\":[\"Optional description\"],\"pebLmQ\":[\"Remove access for \",[\"memberIdDisplay\"]],\"plnnns\":[\"Deleted \",[\"successCount\"],\" image(s), but \",[\"failedCount\"],\" image(s) could not be deleted.\"],\"poCbZw\":[\"Loading ACLs...\"],\"podzPY\":[\"Project id\"],\"psPHye\":[\"Accept Shared Image\"],\"pubQie\":[\"Enter value\"],\"q0Rla3\":[\"Add Tenant Access\"],\"q44uUq\":[\"Containers Partially Emptied\"],\"q5sTNZ\":[\"<0>No Temp URL key configured.</0> A temporary URL key must be set at the account or container level before temporary URLs can be generated. Contact your administrator to configure <1>X-Account-Meta-Temp-URL-Key</1> or <2>X-Container-Meta-Temp-URL-Key</2>.\"],\"q6K46F\":[\"Key already exists\"],\"q88/6A\":[\"Failed to Create Folder\"],\"qAkkjP\":[\"Max object name length\"],\"qEDO1j\":[\"This is a <0>dynamic large object</0>. Only the manifest will be deleted — its segment objects (stored under the manifest prefix) are <1>not</1> automatically removed and must be deleted separately.\"],\"qFDA8L\":[\"Reject access to image\"],\"qJb6G2\":[\"Try Again\"],\"qQ1QBh\":[\"Hardware Specifications\"],\"qST5TS\":[\"Error - Image Details\"],\"qUlxA+\":[\"Folder \\\"\",[\"folderName\"],\"\\\" was permanently deleted.\"],\"qaAo9Y\":[\"Server error occurred while creating the flavor. Please try again later.\"],\"qh5W8q\":[\"Remove Policy\"],\"qs+BrU\":[\"You don't have permission to remove tenant access from this flavor.\"],\"qtoOYG\":[\"No limit\"],\"quU9wK\":[\"Failed to deactivate \",[\"failedCount\"],\" of \",[\"totalCount\"],\" image(s). Some images may already be deactivated or in an invalid state.\"],\"qvF2D8\":[\"There are no images available for this project with the current filters applied. Try adjusting your filter criteria or create a new image.\"],\"qxxo7y\":[\"No policies match your search\"],\"qyNaF7\":[\"Enter a timestamp like \\\"YYYY-MM-DD HH:mm:ss\\\" to schedule automatic deletion\"],\"qzIZOL\":[\"Invalid file format. Supported formats: \",[\"supportedFileFormats\"]],\"qzhUb9\":[\"Showing first \",[\"maxOptions\"],\" of \",[\"totalCount\"],\" — refine your search to narrow results\"],\"r5SQFW\":[\"Container name must be \",[\"maxContainerNameLength\"],\" characters or fewer\"],\"r9Aac8\":[\"Ephemeral Disk\"],\"rAtQcX\":[\"You can rename the object by changing the name here.\"],\"rD9yV1\":[\"Images to deactivate:\"],\"rIe0oV\":[\"Failed to add tenant access to flavor. Please try again.\"],\"rIi6x4\":[\"The flavor could not be found. It may have already been deleted.\"],\"rJe6vw\":[\"7 days\"],\"rbuO5A\":[\"This security group is already shared with the specified project.\"],\"rcBt6T\":[\"Failed to create credential: \",[\"errorMessage\"]],\"rdUucN\":[\"Preview\"],\"rhaNn7\":[\"Loading containers...\"],\"riR9oD\":[\"Note: for <0>static and dynamic large objects</0> only the manifests are deleted — their segments outside this folder prefix are not affected.\"],\"rlgAtt\":[\"The object will be moved to this path. Navigate folders above to change the destination.\"],\"rp0Bd0\":[\"Compute\"],\"rrjuul\":[\"For more details, have a look at the <0>documentation</0>.\"],\"rvXsSb\":[\"Tenant access for \\\"\",[\"tenantIdToRemove\"],\"\\\" has been removed successfully.\"],\"rwBVXS\":[\"Images to be deleted (\",[\"deletableCount\"],\")\"],\"ryf/ee\":[\"Images Activated\"],\"ryxYVo\":[\"Images to be deactivated (\",[\"activeCount\"],\")\"],\"s/s1lz\":[\"Any user can perform a HEAD or GET operation on the container provided the user also has read access on objects. No token is required.\"],\"s2ubkU\":[\"Flavor ID\"],\"s4Vnq2\":[\"Emptying...\"],\"sNVNmf\":[\"MAC Address\"],\"sPFHpI\":[\"Disk\"],\"sSNyf3\":[\"Welcome to <0>Aurora Dashboard</0>, your next-generation cloud management solution. We are dedicated to simplifying how you interact with and manage your cloud infrastructure. Designed with efficiency, scalability, and usability at its core, Aurora empowers you to streamline operations and unlock the full potential of your cloud resources.\"],\"sWBLli\":[\"Add Property\"],\"sXd+qS\":[\"Properties of \\\"\",[\"objectName\"],\"\\\" were successfully updated.\"],\"sa4CV6\":[\"All users from project \",[\"projectId\"]],\"shKIZu\":[\"Images to be activated (\",[\"deactivatedCount\"],\")\"],\"sheDTJ\":[\"Please note: for <0>dynamic</0> and <1>static large objects</1> only the manifests are deleted. The related segments are not deleted.\"],\"sihD20\":[\"Loading images...\"],\"sjMCOP\":[\"Last Modified\"],\"slWh5C\":[\"Associate Floating IP \",[\"floating_ip_address\"],\" with Port\"],\"sxbP3b\":[\"Object Count\"],\"t/YqKh\":[\"Remove\"],\"t0X9+8\":[\"Container Name\"],\"t1POAD\":[\"No custom metadata properties found. Click \\\"Add Property\\\" to create one.\"],\"t1fq6V\":[\"Server returned unexpected data format.\"],\"t7ff15\":[\"valid token required: false\"],\"t95VRV\":[\"About Aurora Dashboard\"],\"tASa/P\":[\"Server error occurred while deleting the flavor. Please try again later.\"],\"tIrNgH\":[\"Server error occurred while fetching extra specs. Please try again later.\"],\"tLerHy\":[\"Ephemeral Disk must be an integer ≥ 0.\"],\"tM5SEI\":[\"ACLs for container \\\"\",[\"containerName\"],\"\\\" were successfully updated.\"],\"tOkmLM\":[\"Failed to Copy Object\"],\"tV/Ozb\":[\"Port Range\"],\"tVSmFT\":[\"Loading more...\"],\"tX5yOZ\":[\"New Folder\"],\"tasfos\":[\"remove\"],\"tbwGSx\":[\"Minimum Disk (GB)\"],\"tejJLY\":[\"Associating Floating IP...\"],\"tfAKBU\":[\"Could not upload \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"tfDRzk\":[\"Save\"],\"tfxu04\":[\"Remove access for \",[\"tenantId\"]],\"thHAVL\":[\"Accepted Images\"],\"tiflqy\":[\"Unable to Re-activate Image\"],\"tlfxPP\":[\"Could not copy \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"tmpGvt\":[\"production, linux\"],\"u+VWhB\":[\"Copied to clipboard!\"],\"u2xIeO\":[\"Failed to update ACLs: \",[\"errorMessage\"]],\"u5HztT\":[\"RX/TX Factor\"],\"u77/s4\":[\"Floating IPs\"],\"u7En0V\":[\"Add Metadata\"],\"uAI0yI\":[\"Delete object:\"],\"uAQUqI\":[\"Status\"],\"uLtFAr\":[\"Could not update container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"uSdnuQ\":[\"VCPUs must be an integer ≥ 1.\"],\"ujK/QN\":[\"Loading objects...\"],\"uly9ET\":[\"Rule Details:\"],\"up0ZSW\":[\"Fingerprint\"],\"uuKb0T\":[\"Description must be less than 65535 characters.\"],\"v0hPHE\":[\"Show Details\"],\"v3djpU\":[\"Move/Rename\"],\"v9Dn8m\":[\"Aurora Dashboard is more than just a tool—it's your partner in navigating the cloud. Whether you're a small startup or a global enterprise, Aurora provides the flexibility, power, and simplicity you need to achieve your goals.\"],\"vBUQNE\":[\"The extra spec could not be found. It may have already been deleted.\"],\"vEkTR9\":[\"Quota\"],\"vH2C/2\":[\"Swap\"],\"vLYu4Z\":[\"Nothing to do. Container appears empty — the object count may not have synced yet due to a recent operation.\"],\"vR4HmN\":[\"Loading Instances...\"],\"vTh35P\":[\"Create Container\"],\"vXmL4D\":[\"Drop your image file here\"],\"vZUKSz\":[\"Detach Floating IP \",[\"floating_ip_address\"]],\"vbajgL\":[\"Public Flavor\"],\"vcQSZh\":[\"This folder is empty — use New Folder to create one.\"],\"vcXmqy\":[\"Network Overview\"],\"vcvCXq\":[\"Error - Flavor Details\"],\"vg84cD\":[[\"allCount\"],\" items\"],\"vmRPFm\":[\"Share Security Group\"],\"vmYyLY\":[\"Remote IP Prefix\"],\"vp5vfW\":[\"1 hour\"],\"vpt8cE\":[\"Generate URL\"],\"vrPCbw\":[\"Image ID\"],\"w3bAcf\":[\"This action is permanent. The address will be removed from your project and returned to the public pool. This action cannot be undone.\"],\"w9+8d7\":[\"Remove tenant access\"],\"wEfZld\":[\"Create New Flavor\"],\"wFaT8w\":[\"Failed to Empty Containers\"],\"wMHvYH\":[\"Value\"],\"wPrtGF\":[\"Enter key\"],\"wTg+FY\":[\"Max file size\"],\"wXxPjv\":[\"S3 Object Storage — Setup Required\"],\"wa1Bcq\":[\"Enter tenant ID\"],\"wbqM4L\":[[\"customMinutes\"],\" minutes\"],\"wcUecy\":[\"You don't have permission to view extra specs for this flavor.\"],\"we28Pq\":[\"Hide ACLs Preview\"],\"wlQNTg\":[\"Members\"],\"wlUDbB\":[\"Last updated: \",[\"formattedDate\"]],\"wrXcuy\":[\"Object Name\"],\"wrk/xj\":[\"Image Details\"],\"wyIOMP\":[\"Image name is required\"],\"wzqqS+\":[\"Key Features\"],\"x/XQrD\":[\"Any file type\"],\"x1bK0h\":[\"There are no containers available with the current search criteria. Try adjusting your search term.\"],\"x3T4pq\":[\"The container metadata reports objects but none were listed. This may be a temporary synchronization delay — please wait a moment and try again.\"],\"x5l/TK\":[\"Already active (will be skipped):\"],\"x9AdZ8\":[\"property_key\"],\"xNG/3n\":[\"Floating IP Address\"],\"xNZKYy\":[\"Failed to delete \",[\"failedCount\"],\" of \",[\"totalCount\"],\" image(s). Some images may be protected or in use.\"],\"xqhyRT\":[\"Object Uploaded\"],\"xw2UtT\":[\"Create New Image\"],\"y+KBOY\":[\"e.g., production, linux, ubuntu\"],\"y02Bu1\":[\"Container:\"],\"y0u86k\":[\"The requested flavor could not be found. It may have been deleted or you may not have access to it.\"],\"y1GYnY\":[\"Could not move \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"yPWFWy\":[\"ICMP Type\"],\"yTtJTy\":[\"Edit Image Metadata\"],\"yYxB17\":[\"Clear all\"],\"ylfbpz\":[\"Extra spec key is required and cannot be empty.\"],\"yp0UjB\":[\"Ethertype\"],\"yqPflB\":[\"... and \",[\"hiddenCount\"],\" more\"],\"yu9G3x\":[\"Edit Security Group\"],\"yz7wBu\":[\"Close\"],\"z+zpLP\":[\"valid token required: true\"],\"z1JceR\":[\"Back to Floating IPs\"],\"z45o5B\":[\"Object count\"],\"z9NAjZ\":[\"Object Deleted\"],\"zCD96i\":[\"You are not authorized to view flavor details. Please log in again.\"],\"zDS0JC\":[\"Name must be 2-50 characters long.\"],\"zWb/Nn\":[\"Max header size\"],\"zc5dcw\":[\"Login failed. Please check your credentials and try again.\"],\"zga9sT\":[\"OK\"],\"zhM8FP\":[\"Grant access to a user from a different project.\"],\"zm7+/D\":[\"You are about to deactivate <0>\",[\"activeCount\"],\" image(s)</0>. Deactivated images cannot be used to launch new instances.\"],\"zwBp5t\":[\"Private\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"+0B+ue\":[\"Projects\"],\"+9CXS9\":[\"Deactivate Images\"],\"+Jcye3\":[\"Key Name\"],\"+Lt5cp\":[\"You are not authorized to add tenant access. Please log in again.\"],\"+NwLgN\":[\"Activating this image will allow it to be used to launch new instances again.\"],\"+Nx1wc\":[\"Failed to load Floating IPs\"],\"+OEi73\":[\"Object Storage (Swift)\"],\"+YQ9qu\":[\"Container: \",[\"containerName\"]],\"+nQTmZ\":[\"This tenant does not have access to the flavor.\"],\"+p6nHr\":[\"Failed to load object metadata: \",[\"metadataErrorMessage\"]],\"+zy2Nq\":[\"Type\"],\"/1MfrG\":[\"Could not download \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"/2a/eI\":[\"Loading Flavor...\"],\"/9Squ9\":[\"You don't have permission to view this flavor's details.\"],\"/BZLRP\":[\"To confirm this action, type the word <0>“detach”</0> in the field below.\"],\"/EcdUM\":[\"Your action is required\"],\"/HgF9q\":[\"Sort by\"],\"/InK0O\":[\"Total size\"],\"/LqWNN\":[\"Could not delete \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"/NeNjH\":[\"Container \\\"\",[\"containerName\"],\"\\\" properties were successfully updated.\"],\"/Nmxy/\":[\"No key pairs available.\"],\"/QIkBY\":[\"<0>Secure & Reliable:</0> Your data and operations are safeguarded with enterprise-grade security and robust reliability.\"],\"/Qox3b\":[\"A folder with this name already exists\"],\"/Z2leb\":[\"No containers found.\"],\"/Z5n1b\":[\"Create folder below:\"],\"/bUiYk\":[\"Router ID\"],\"/eFtWI\":[\"RBAC Policies\"],\"/xnbdQ\":[\"The specified user has access. A token for the user (scoped to any project) must be included in the request.\"],\"01/uUD\":[\"Keep segments (delete manifest only)\"],\"07WXfc\":[\"Server returned unexpected data format for extra specs.\"],\"0BSSYj\":[\"Server error occurred while removing tenant access. Please try again later.\"],\"0Gd0NU\":[\"Shared\"],\"0P2gFy\":[\"The page you are looking for does not exist.\"],\"0WsqO0\":[\"Containers Emptied\"],\"0cVgUw\":[\"Filter by\"],\"0dZfcJ\":[[\"allContainersCount\"],\" containers\"],\"0eY8Mz\":[\"There are no Floating IPs available for this project. Floating IPs allow you to map public IP addresses to instances.\"],\"0j9NA7\":[\"<0>Are you sure?</0> All objects in the container will be deleted. This cannot be undone.\"],\"0kCt7e\":[\"The flavor data provided is invalid. Please check your input.\"],\"0kc0zi\":[\"Server error occurred while deleting the extra spec. Please try again later.\"],\"0o0OhW\":[\"No objects found.\"],\"0p+s6m\":[\"Type: \",[\"typeValue\"],\", Code: \",[\"codeValue\"]],\"0u9jhd\":[\"Detaching this Floating IP will remove its association with the current port. The instance will no longer be reachable through this address.\"],\"16085O\":[\"IP Version\"],\"1H2g6v\":[\"Moving object...\"],\"1NS3nd\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" container\"],\"other\":[\"#\",\" containers\"]}],\" successfully emptied. \",[\"totalDeleted\",\"plural\",{\"one\":[\"#\",\" object\"],\"other\":[\"#\",\" objects\"]}],\" deleted in total.\"],\"1RwosK\":[\"Target project ID is required\"],\"1UzENP\":[\"No\"],\"1VDqZj\":[\"<0>Future-Ready:</0> Aurora is designed to evolve with the latest trends in cloud technology, ensuring your solution is always cutting-edge.\"],\"1iQtS2\":[\"Showing first \",[\"actualObjectCount\"],\" of \",[\"total\"],\" objects\"],\"1iUuTT\":[\"Your session has expired. Please log in again.\"],\"1ojTVo\":[\"Select a DNS domain.\"],\"1pGUZa\":[\"Session expires in\"],\"1pdLQw\":[\"Image not found\"],\"1rLu3+\":[\"Could not empty container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"1rPB1p\":[\"The flavor or tenant could not be found. Please verify they exist.\"],\"1t/NnN\":[\"Reject\"],\"1zZ1IK\":[\"Hi\"],\"20E+79\":[\"You need to login to access this page.\"],\"20Kpaw\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" has been successfully deleted.\"],\"20axE5\":[\"Shared by Project\"],\"23wBCX\":[\"Public Read Access\"],\"2G6hLq\":[\"Delete \",[\"specKey\"]],\"2Inn83\":[\"Bulk upload of archive files\"],\"2TtIL2\":[\"Stored as X-Object-Meta-* headers. Keys are case-insensitive.\"],\"2cJIlz\":[\"Floating Network ID\"],\"2d/OiW\":[\"Enter your username\"],\"2dnZwV\":[\"Could not delete folder \\\"\",[\"folderName\"],\"\\\": \",[\"errorMessage\"]],\"2gH+i8\":[\"You are not authorized to delete flavors. Please log in again.\"],\"2lq0gq\":[\"<0>Properties of</0> <1>\",[\"displayName\"],\"</1>\"],\"2mbisJ\":[\"Metadata \\\"\",[\"trimmedKey\"],\"\\\" has been added successfully.\"],\"2pnrGl\":[\"Expected format: YYYY-MM-DD HH:MM:SS\"],\"2q/Q7x\":[\"Visibility\"],\"2ysnjX\":[\"<0>Enhanced Productivity:</0> By reducing operational complexity, Aurora helps your team focus on what truly matters—innovating and driving business success.\"],\"2zceEg\":[\"This action cannot be undone. The image will be permanently deleted.\"],\"33F2A+\":[\"Type container name to confirm\"],\"3AUpb4\":[\"Delete All (\",[\"selectedCount\"],\")\"],\"3Qn0me\":[\"Add Member\"],\"3n+vCm\":[\"Custom duration (minutes)\"],\"3nWqQW\":[\"You are not authorized to view extra specs. Please log in again.\"],\"3nh/7E\":[\"If checked, this flavor will be available to all tenants. If unchecked, access must be explicitly granted to specific tenants.\"],\"3oChIh\":[\"<0>Unified Cloud Management:</0> Consolidates all your cloud assets into one intuitive interface.\"],\"3oc18/\":[\"Private flavors could not be loaded. You may be seeing an incomplete list.\"],\"3q1GLx\":[\"Pending file upload...\"],\"3x7Sws\":[\"Loading Security Group Details...\"],\"47eI0x\":[\"Description must be at least 1 character.\"],\"4EZrJN\":[\"Rules\"],\"4O2AH3\":[\"Member \\\"\",[\"memberIdToRemove\"],\"\\\" has been removed successfully.\"],\"4fh0Wj\":[\"Boot size\"],\"4fvDRe\":[\"Images to activate:\"],\"4fvcmm\":[\"Object will be uploaded as: <0>\",[\"selectedObjectName\"],\"</0>\"],\"4h3Eyf\":[\"\\\"\",[\"objectName\"],\"\\\" was successfully uploaded.\"],\"4kjaAc\":[\"No server groups available.\"],\"4mbrAq\":[\"1 minute\"],\"4opp4r\":[\"Security Groups\"],\"4pOfUd\":[\"Our Mission\"],\"4t33sh\":[\"Failed to Update Object\"],\"4uXhtt\":[\"CIDR\"],\"4utWB4\":[\"Server Role:\"],\"5/wyf8\":[\"Enter a floating IP\"],\"56IxdF\":[\"Failed to load container objects: \",[\"errorMessage\"]],\"5BLR6Q\":[\"IPv4\"],\"5Hdka/\":[[\"allContainersCount\"],\" container\"],\"5JDSvn\":[\"Max meta value length\"],\"5M4Te3\":[\"DNS\"],\"5MF8U2\":[\"Failed to Update Container\"],\"5Yrl6N\":[\"Loading Server Groups...\"],\"5aNQ3F\":[\"\\\"\",[\"objectName\"],\"\\\" was successfully copied to \",[\"destination\"],\".\"],\"5g7owI\":[\"Updating Floating IP...\"],\"5y3O+A\":[\"Deactivate\"],\"6+7EwD\":[\"Serve objects as index when file name is:\"],\"6+OdGi\":[\"Protocol\"],\"6/xipy\":[\"Container Format\"],\"644xgx\":[\"Protected\"],\"6BDqha\":[\"Limits\"],\"6CDYXS\":[\"Static website serving\"],\"6GBt0m\":[\"Metadata\"],\"6H/Lg1\":[\"This is a public image. All users have access to it. Explicit sharing is not needed.\"],\"6KRclz\":[\"Folder Created\"],\"6Kjltl\":[\"Access Control for container:\"],\"6OopEX\":[\"Container Emptied\"],\"6Rnrsz\":[\"Manage Access - \",[\"flavorName\"]],\"6X/9Di\":[\"\\\"\",[\"objectName\"],\"\\\" was successfully moved to \",[\"destination\"],\".\"],\"6YtxFj\":[\"Name\"],\"6jAi8c\":[\"Range\"],\"6luZQA\":[\"Object Moved\"],\"6oolxV\":[\"This extra spec keys already exist. Please use different keys.\"],\"6qzsuS\":[\"Write ACLs\"],\"6sxz+g\":[\"Port Name\"],\"6w+VnM\":[\"Container Created\"],\"6z9W13\":[\"Restart\"],\"76RKuS\":[\"ICMP Code\"],\"78+riR\":[\"You are not authorized to remove tenant access. Please log in again.\"],\"7AfIPZ\":[\"Floating Network\"],\"7BpykL\":[\"Failed to create extra specs. Please try again.\"],\"7L01XJ\":[\"Actions\"],\"7NC3vm\":[\"Subnet\"],\"7NSdfG\":[\"Emptying container \",[\"progressCurrent\"],\" of \",[\"progressTotal\"],\", please wait...\"],\"7Q24LN\":[\"Policy\"],\"7T1fHv\":[\"Failed to remove member \\\"\",[\"memberIdToRemove\"],\"\\\"\"],\"7UlHhT\":[\"Metadata \\\"\",[\"keyToDelete\"],\"\\\" has been deleted successfully.\"],\"7XQ3QJ\":[\"Denied referrer: \",[\"host\"]],\"7ZnTL8\":[\"Failed to update object: \",[\"mutationErrorMessage\"]],\"7a4DvD\":[\"No servers available.\"],\"7d1a0d\":[\"Public\"],\"7flw0l\":[\"Tenant access for \\\"\",[\"trimmedTenantId\"],\"\\\" has been added successfully.\"],\"7sMeHQ\":[\"Key\"],\"88kg0+\":[\"Created At\"],\"8AriEH\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been created\"],\"8TSI9h\":[\"Deactivating this image will prevent it from being used to launch new instances. Existing instances will not be affected.\"],\"8Tg/JR\":[\"Custom\"],\"8ZOb7O\":[[\"numberDeleted\"],\" object was permanently deleted.\"],\"8ZsakT\":[\"Password\"],\"8c3/77\":[\"Max meta name length\"],\"8jLXs3\":[\"Versioned writes\"],\"8s0tOH\":[\"You don't have permission to add tenant access to this flavor.\"],\"8t1+HU\":[\"Deactivated \",[\"successCount\"],\" image(s), but \",[\"failedCount\"],\" image(s) could not be deactivated.\"],\"8wdCNd\":[\"tcp, udp, icmp, or protocol number\"],\"8zAn1f\":[\"Failed to delete flavor. Please try again.\"],\"98Fs4G\":[\"Creating image...\"],\"9J93Xr\":[\"Container name cannot contain slashes\"],\"9SX0bO\":[\"The image \\\"\",[\"imageName\"],\"\\\" could not be updated: \"],\"9X8lAk\":[\"Allocate\"],\"9doWrf\":[\"Failed to add member\"],\"9dsDHD\":[\"The image \\\"\",[\"imageId\"],\"\\\" could not be re-activated: \",[\"message\"]],\"9iz2XW\":[\"Unable to Update Image\"],\"9njIiV\":[\"Failed to Activate Images\"],\"9rz81C\":[\"Device ID\"],\"9v5VLp\":[\"No custom properties defined\"],\"9vSW3U\":[\"Delete Recursively\"],\"9x6EkK\":[\"This is a public flavor. All tenants have access to it.\"],\"A7CVME\":[\"Select disk format first\"],\"AB4Tnl\":[\"Please select a file to upload\"],\"AGXLLY\":[\"Unable to Upload Image File\"],\"AJRhSM\":[\"Root Disk must be an integer ≥ 0.\"],\"AN0DBJ\":[\"Press Enter to add\"],\"AX9Juz\":[\"ID must only contain alphanumeric characters, hyphens, underscores, and dots.\"],\"AZyHwC\":[\"Must be a valid IPv4 or IPv6 address (for example: 172.24.4.228 or 2001:db8::1).\"],\"Ac6dy9\":[\"Type name\"],\"AdtLNV\":[\"Ensure ACL entries are valid — correct project IDs, user IDs, and formats are your responsibility. Invalid entries may silently grant or deny unintended access.\"],\"AeXO77\":[\"Account\"],\"Afh/Lb\":[\"Select destination folder\"],\"AlbUVn\":[\"<0>Optimized Scalability:</0> Built for businesses of all sizes, Aurora grows with you, supporting simple environments and intricate multi-cloud setups alike.\"],\"Alx2/L\":[\"Open in new tab\"],\"AuQtzx\":[\"Must be a non-negative integer\"],\"AxZkIr\":[\"Disk (GiB)\"],\"B2Czeb\":[\"Min. RAM\"],\"B2i9cQ\":[\"Objects to be deleted (\",[\"totalCount\"],\")\"],\"B3toQF\":[\"Objects\"],\"B4Jzm7\":[\"Ceph\"],\"BCJPTn\":[\"Grant access to all users from that project.\"],\"BCXapL\":[\"Failed to load container properties: \",[\"errorMessage\"]],\"BJt+PJ\":[\"Failed to Delete Container\"],\"BMTd81\":[\"This action cannot be undone. The target project will lose access to this security group immediately.\"],\"BOQYRn\":[\"Loading Key Pairs...\"],\"BP4Fwj\":[\"Error Loading Objects: \",[\"errorMessage\"]],\"BSaBkZ\":[\"Objects — \",[\"containerName\"]],\"BYH/2L\":[\"Unable to Deactivate Image\"],\"BZpsYm\":[\"Failed to load containers: \",[\"errorMessage\"]],\"BgMp/T\":[\"Invalid format combination for selected disk format\"],\"BoIAP6\":[\"The ID of the network associated with the floating IP.\"],\"BoPocW\":[\"MD5 checksum\"],\"BrrIs8\":[\"Storage\"],\"CA8ZeT\":[\"Image \\\"\",[\"imageName\"],\"\\\" visibility updated to \",[\"visibility\"]],\"CBFSfX\":[\"Please fix the validation errors below.\"],\"CFMxC8\":[\"Images Deleted\"],\"CMVP7y\":[\"This action cannot be undone. The rule will be permanently deleted.\"],\"CgZxr7\":[\"Min RAM (MB)\"],\"ChOuUj\":[\"Floating IP not found\"],\"Cj2Gtd\":[\"Size\"],\"ClGcRq\":[\"Containers\"],\"Cu6xuZ\":[\"This is a <0>dynamic large object (DLO)</0> manifest. Metadata changes apply to the manifest only — segment objects are not affected.\"],\"CunRry\":[\"Invalid project ID format. Must be 32 hexadecimal characters (e.g., b90f9c4bc76140e18540b2cec1299e2a) or UUID format (e.g., 12345678-1234-1234-1234-123456789abc)\"],\"Cxgv2U\":[\"Min. Disk\"],\"D+Se/Q\":[\"Empty container:\"],\"D/8vkD\":[\"It will appear in your image list.\"],\"D1F789\":[\"<0>Are you sure?</0> The container will be deleted. This cannot be undone.\"],\"D3IRXw\":[\"Detaching Floating IP...\"],\"D7qT9F\":[\"Why Choose Aurora?\"],\"DDRhQm\":[\"Your session has expired.\"],\"DJT9tB\":[\"Account quotas\"],\"DKkOPx\":[\"Extra Specs\"],\"DNVql8\":[\"Full lifecycle management of Floating IPs, including attachment, port association/disassociation, DNS settings, and deletion\"],\"DcMIiu\":[\"Could not update ACLs for container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"Df0YHr\":[\"Update Security Group\"],\"Dh1qvV\":[\"You are about to delete \",[\"deletableCount\"],\" image(s). This action cannot be undone.\"],\"Dia2Ue\":[\"There are no RBAC policies for this security group\"],\"Do5/uH\":[\"The flavor or tenant could not be found. It may have already been removed.\"],\"Dqnh7K\":[\"Specific referrer: \",[\"host\"]],\"Dt5W9T\":[\"Remove RBAC Policy\"],\"DvB4XF\":[\"Drop your file here\"],\"E/QGRL\":[\"Disabled\"],\"E4QYe7\":[\"Suggested Images\"],\"E6nRW7\":[\"Copy URL\"],\"EF2EU9\":[\"Deleting...\"],\"EPMHs9\":[\"You don't have permission to delete flavors in this project.\"],\"EQnVgi\":[\"Flavor service is not available for this project.\"],\"EdQY6l\":[\"None\"],\"Ef7StM\":[\"Unknown\"],\"Enpdmy\":[\"Type <0>remove</0> to confirm:\"],\"EoKe5U\":[\"Domain\"],\"Eq5PsT\":[\"Type \\\"detach\\\" to confirm\"],\"EqSPkP\":[\"Loading Flavors...\"],\"Erlvqg\":[\"Object name cannot have leading or trailing whitespace\"],\"ExLULX\":[\"Image Name\"],\"EztMB8\":[\"Failed to fetch flavors from server.\"],\"F02e8I\":[\"No custom metadata. Click \\\"Add Property\\\" to create one.\"],\"F6YIQe\":[\"Efficient bulk deletion\"],\"FKL6Jv\":[\"e.g. .r:*,.rlistings\"],\"FNcMGM\":[\"Creation Date\"],\"FOcBn3\":[\"Detach\"],\"FPsvA8\":[\"Got it!\"],\"FQBaXG\":[\"Activate\"],\"FRtmJJ\":[\"Storage container not found\"],\"FSbpS7\":[\"CPU\"],\"FjONW3\":[\"Error Loading Flavor\"],\"FjPnAE\":[\"Error loading security group\"],\"Flugry\":[[\"progressPct\"],\"%\"],\"FwSyEp\":[\"The specified project does not exist or you don't have permission to share with it.\"],\"Fzrzfe\":[\"Folder name is required\"],\"G6AP+o\":[\"Shared:\"],\"GEgjm+\":[\"Loading Objects...\"],\"GPuCEo\":[\"Leave empty for all types\"],\"GSIPwA\":[\"Temporary URL\"],\"GbKqnI\":[\"Activated \",[\"successCount\"],\" image(s), but \",[\"failedCount\"],\" image(s) could not be activated.\"],\"Gfx1qQ\":[\"Unable to Load Content\"],\"GxkJXS\":[\"Uploading...\"],\"Gyd3No\":[\"No specific tenant access configured for this private flavor. Click \\\"Add Tenant Access\\\" to grant access.\"],\"H+a5j6\":[\"Release\"],\"H4Qwmp\":[\"No objects match your search. Try adjusting your search term.\"],\"H7u085\":[\"No projects have access to this image yet. Click \\\"Add Project Access\\\" to grant access.\"],\"HAkrpK\":[\"At Aurora, our mission is to provide a centralized platform that unifies cloud management. We aim to simplify the complexities of provisioning, configuring, and scaling resources across diverse cloud environments while enabling seamless growth for your business.\"],\"HBI6nY\":[\"Need help?\"],\"HBpi4q\":[\"Loading Images...\"],\"HM56Bx\":[\"Creating...\"],\"HNlEFZ\":[\"delete\"],\"HQH8HM\":[\"Could not update \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"HVdrr1\":[\"ANY referrer\"],\"HivZR9\":[\"Create Credential\"],\"Hivb/4\":[\"Server is experiencing issues. Please try again later.\"],\"Hiw1Ha\":[\"No containers found\"],\"HlwgQN\":[\"Object \\\"\",[\"objectName\"],\"\\\" was permanently deleted.\"],\"HuA8iQ\":[\"Allocating Floating IP...\"],\"HxTYrE\":[\"The flavor could not be found. It may have been deleted.\"],\"I5kZVK\":[\"Remote Source\"],\"INUP6f\":[\"<0>Effortless Resource Provisioning:</0> Quickly provision, configure, and deploy resources like servers, networks, and volumes with just a few clicks.\"],\"IOkHLC\":[\"Failed to copy object: \",[\"errorMessage\"]],\"IUwGEM\":[\"Save Changes\"],\"IWF68U\":[\"Storage Overview\"],\"IZ6Mh2\":[\"Enter your domain\"],\"IbYr/u\":[\"Content type\"],\"J4DKSM\":[\"Container format is required\"],\"J6EOll\":[\"Move/Rename object:\"],\"J7+bZb\":[\"Folder Deleted\"],\"J9QcnV\":[\"Successfully activated \",[\"successCount\"],\" of \",[\"totalCount\"],\" image(s)\"],\"J9cmxx\":[\"Failed to update visibility to \",[\"newVisibility\"]],\"JB0bhm\":[\"Get Involved\"],\"JNGYAW\":[\"Container name is required\"],\"JT3I1g\":[\"Delete Flavor\"],\"JeRXll\":[\"This key is reserved and managed separately\"],\"JfWCsP\":[\"Partial Deactivation Success\"],\"Jh4rAZ\":[\"Error loading image\"],\"Jim5X9\":[\"Stateful\"],\"JoECY1\":[\"The extra spec data provided is invalid. Please check your input.\"],\"JpZn1L\":[\"Already deactivated (will be skipped)\"],\"JrmKyf\":[\"Failed: \",[\"errorDetails\"]],\"JtHgVz\":[\"Delete Images\"],\"K+e/0e\":[\"RAM (MiB)\"],\"K3bUTE\":[\"Minimum disk must be 0 or greater\"],\"K8Qnlj\":[\"Moving...\"],\"K9eC8x\":[\"This could be due to insufficient permissions or a temporary service issue. Please check your access rights or try refreshing the page.\"],\"KDw4GX\":[\"Try again\"],\"KJC+M7\":[\"Server error occurred while fetching flavor details. Please try again later.\"],\"KOpPMt\":[\"Total size quota\"],\"KSW/GC\":[\"There are no flavors available for this project with the current filters applied. Try adjusting your filter criteria or create a new flavor.\"],\"KZN4Lc\":[\"Delete All\"],\"Km4AGG\":[\"Creating security group...\"],\"KoQP4F\":[\"Server error occurred while adding tenant access. Please try again later.\"],\"KsIM0b\":[\"Boot RAM\"],\"KsnZ3m\":[\"Folder \\\"\",[\"folderName\"],\"\\\" was successfully created.\"],\"KzUd7m\":[\"new-folder-name\"],\"LI8Z2I\":[\"Download \",[\"rowDisplayName\"]],\"LK0pQN\":[\"Disk format is required\"],\"LMdsuJ\":[\"Port (from)\"],\"LQQCas\":[\"Folder \\\"\",[\"folderName\"],\"\\\" and \",[\"deletedCount\"],\" object was permanently deleted.\"],\"Llcakz\":[\"Updated At\"],\"LqMb+g\":[\"To confirm this action, type the word <0>\\\"release\\\"</0> in the field below.\"],\"LtI9AS\":[\"Owner\"],\"Lylr9Z\":[\"Object Copied\"],\"M470oJ\":[\"The flavor could not be found or has no extra specs.\"],\"M5Epeo\":[\"Edit Image Details\"],\"M5RhXF\":[\"Removing...\"],\"M5rEN5\":[\"Session Expired\"],\"M9H+/G\":[\"projects\"],\"MEIAzV\":[\"Unnamed\"],\"MJtNLd\":[\"Images to delete:\"],\"MOug+V\":[\"Enter a tag and press Enter or click Add\"],\"MRB7nI\":[\"Direction\"],\"MXoA/6\":[\"Upload Object\"],\"MXw7Fr\":[\"Server Name\"],\"MZGbkp\":[\"VCPUs\"],\"MbKJNP\":[\"You don't have permission to access flavor access information for this flavor.\"],\"MgZyuJ\":[\"You are about to activate <0>\",[\"deactivatedCount\"],\" image(s)</0>. Activated images will be available for launching new instances.\"],\"MmtQVF\":[\"Invalid value for public flavor setting.\"],\"Mt6sRo\":[\"You are not authorized to access flavor access information. Please log in again.\"],\"MtzSbv\":[\"Object name is required\"],\"MuKU9V\":[\"Failed to load objects: \",[\"errorMessage\"]],\"N2S1rs\":[\"Empty\"],\"N5I2RJ\":[\"Type \\\"release\\\" to confirm\"],\"N5vGcw\":[\"Enter your credentials to access your account\"],\"NH2fsP\":[\"Already deactivated (will be skipped):\"],\"NOdFZR\":[\"Generating...\"],\"NQU1Nn\":[\"Copy container name\"],\"NRMm0E\":[\"This tenant already has access to the flavor.\"],\"NRP2uq\":[\"Share object:\"],\"NRVSdy\":[\"Member ID\"],\"NW4PIb\":[\"Could not create folder \\\"\",[\"folderName\"],\"\\\": \",[\"errorMessage\"]],\"NZJhro\":[\"Object name cannot contain slashes\"],\"Nc7QKU\":[\"Fixed IP Address\"],\"NeUjqc\":[\"Enable file listing\"],\"NixRmA\":[\"Min Disk (GB)\"],\"NlcF/v\":[\"No flavor selected for deletion.\"],\"NopYGU\":[\"Disk Format\"],\"Np28ib\":[\"or drag and drop\"],\"Nu4oKW\":[\"Description\"],\"Nvfd2b\":[\"Versioning is enabled\"],\"O80bQY\":[\"Loading object properties...\"],\"O8tK4v\":[\"Add rule\"],\"ONWvwQ\":[\"Upload\"],\"OR475H\":[\"Network\"],\"OSlLnz\":[\"Image Visibility\"],\"OYHzN1\":[\"Tags\"],\"OZImTR\":[\"Container listing limit\"],\"OaSktR\":[\"Device Owner\"],\"OlmKCg\":[\"A flavor with this ID or name already exists. Please use different values.\"],\"OvEjsP\":[\"Copying...\"],\"Ovofy+\":[\"Release Floating IP \",[\"floating_ip_address\"]],\"OxDN2m\":[\"Failed to create flavor. Please try again.\"],\"OxaeYj\":[\"We are building Aurora Dashboard to serve you better. Your feedback is invaluable in shaping a tool that meets the unique needs of businesses like yours. Stay connected and join us as we redefine cloud management.\"],\"Oxl1UN\":[\"If there is no index file, the URL displays a list of objects in the container.\"],\"PAKSdy\":[\"Enter a floating IP or leave blank to auto-assign one\"],\"PEGvy+\":[\"Minimum RAM must be 0 or greater\"],\"PHt+EV\":[\"Type <0>delete</0> to confirm:\"],\"PIbPRX\":[\"RX/TX Factor must be an integer ≥ 1.\"],\"PLwzWR\":[\"All containers\"],\"PYQUjU\":[\"Failed to load metadata configuration.\"],\"PZnUbs\":[\"Please log in again to continue.\"],\"PgNNGl\":[\"More Actions\"],\"PiH3UR\":[\"Copied!\"],\"PiyQJ/\":[\"No flavors found\"],\"PkfPsB\":[\"Enter the ID of the project you want to share this security group with. You can find project IDs in the account/project switcher or in the Identity service.\"],\"Pkw7J9\":[\"This folder is empty.\"],\"PsEGri\":[\"Ubuntu 22.04 LTS\"],\"PtjzS+\":[\"Associates on the selected port. If the port has multiple IPs, select the desired fixed IP address.\"],\"PzgYM9\":[\"Checksum\"],\"Q1W//7\":[\"No services available for this project.\"],\"Q2xmVl\":[\"Symlinks\"],\"Q9f2QF\":[[\"numberDeleted\"],\" objects were deleted successfully, but some deletions failed.\"],\"QAUa4B\":[\"Enter a single port, or define a range by also filling \\\"Port (to)\\\". \\\"Port (to)\\\" is optional.\"],\"QEtDlS\":[\"Copying object...\"],\"QNHur0\":[\"Failed to load container ACLs: \",[\"errorMessage\"]],\"QQ8wUG\":[\"This action cannot be undone. The flavor will be permanently deleted.\"],\"QV1ZPO\":[\"Key is required\"],\"QWdKwH\":[\"Move\"],\"QYiqYb\":[\"Failed to update access status\"],\"Qb+14I\":[\"This action cannot be undone. The security group will be permanently deleted.\"],\"QetsXP\":[\"Upload failed: \",[\"uploadError\"]],\"Qg4EG6\":[\"Unable to connect to the compute service. Please check your connection and try again.\"],\"QuJSSl\":[\"Failed to create the flavor. Please try again.\"],\"QvqBQa\":[\"Target container\"],\"Qx7DM7\":[\"Capabilities\"],\"QxBGbh\":[\"Protected (will be skipped):\"],\"R0xWJL\":[\"Cannot delete. Container contains objects. Please empty it first.\"],\"R6kcsL\":[\"Must be a valid PQDN or FQDN (alphanumeric and hyphens only, cannot start or end with hyphen).\"],\"R6u5CR\":[\"Failed to activate \",[\"failedCount\"],\" of \",[\"totalCount\"],\" image(s). Some images may already be active or in an invalid state.\"],\"RByeNR\":[\"Your session expired. Please login again.\"],\"RCr0yv\":[\"Failed to load flavor details. Please try again.\"],\"RFDYCD\":[\"Minimum disk size required to boot this image\"],\"RGhYAo\":[\"RAM\"],\"RGrgxg\":[\"Could not delete container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"RGwfoL\":[\"Max meta count\"],\"RNBvdl\":[\"Max SLO segments\"],\"RSFkXF\":[\"Activate Image\"],\"RSMPjT\":[\"You are currently on the dashboard route.\"],\"RSg/pq\":[\"Failed to Delete Object\"],\"RTQFAw\":[\"You are not authorized to create extra specs. Please log in again.\"],\"Rih53k\":[\"Max container name length\"],\"Rlp5zj\":[\"Create Flavor\"],\"S0kLOH\":[\"ID\"],\"S1iTXO\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been deleted\"],\"S3olSf\":[\"No extra specs found. Click \\\"Add Metadata\\\" to create one.\"],\"S5CUKP\":[\"Member ID (project UUID) is required.\"],\"S63NbU\":[\"The image \\\"\",[\"imageId\"],\"\\\" could not be deactivated: \",[\"message\"]],\"S8/j2h\":[\"Failed to Empty Container\"],\"SBGiGm\":[\"Read ACLs\"],\"SCY5an\":[\"Failed to Move Object\"],\"SFo0kK\":[\"All Images\"],\"SIfYq6\":[\"Edit Metadata\"],\"SLEH7X\":[\"Enter DNS name\"],\"STc+7E\":[\"Max containers per extraction\"],\"SU0uxT\":[\"Upload object to:\"],\"SUSS9i\":[\"Container name\"],\"SVLToM\":[\"Type \\\"remove\\\" to confirm\"],\"SZw9tS\":[\"View Details\"],\"Sb/VT5\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully emptied. \",[\"deletedCount\"],\" objects deleted.\"],\"SfW/3r\":[\"There are no groups\"],\"Sgz1vJ\":[\"Member \\\"\",[\"trimmedMemberId\"],\"\\\" has been added successfully.\"],\"Smk7M2\":[\"Error loading floating IP\"],\"SuX2Ca\":[\"Basic Info\"],\"SysqAR\":[\"Flavor Details\"],\"T6Gm5y\":[\"Select an external network\"],\"T7mgdd\":[\"Successfully deleted \",[\"successCount\"],\" of \",[\"totalCount\"],\" image(s)\"],\"T8N6oi\":[\"Property Key\"],\"T9Mtpi\":[\"Tenant ID\"],\"TM93nK\":[\"Delete Security Group Rule\"],\"TPMaxo\":[\"Type “release” to confirm\"],\"TQn3hH\":[\"Failed to create image. Please try again.\"],\"TZJiVf\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully emptied. \",[\"deletedCount\"],\" object deleted.\"],\"TfC9O+\":[\"Last modified (UTC)\"],\"TfdeUd\":[\"Failed to delete the extra spec. Please try again.\"],\"TpGxnq\":[\"Enter member ID\"],\"Tx4Ym+\":[\"Enter a valid PQDN or FQDN (max 63 characters) to associate with the floating IP. A and PTR records are created automatically.\"],\"TyODHt\":[\"Save Metadata\"],\"U/oahm\":[\"URL Copied\"],\"U2wTy/\":[\"Note: The 'stateful' attribute cannot be changed if this security group is currently in use by one or more ports.\"],\"U4fmHG\":[\"The text must match “detach” in lowercase.\"],\"U6L+P/\":[\"Inactivity Timeout\"],\"U9q4M7\":[\"Back to Security Groups\"],\"UGhVPl\":[\"Object Type\"],\"UJVf0u\":[\"Loading Image...\"],\"UK2mpr\":[\"Generating temporary URL...\"],\"UKwOYH\":[\"Image File\"],\"UO3hJ2\":[\"Temporary URLs\"],\"UQ7Wyv\":[\"Manage Access for Image - \",[\"imageName\"]],\"URmyfc\":[\"Details\"],\"USiuNX\":[\"Container quotas\"],\"UVFHGY\":[\"e.g. PROJECT_ID:USER_ID\"],\"UVSFVV\":[\"Reject Shared Image\"],\"UYSopm\":[\"Minimum RAM (MB)\"],\"UbRKMZ\":[\"Pending\"],\"UdcGJu\":[\"Activate Images\"],\"UiNv/G\":[\"S3 Object Storage requires EC2 credentials (access key + secret key) to authenticate your requests. You need to create credentials before accessing S3 resources.\"],\"Uj+n/2\":[\"Failed to Delete Folder\"],\"UkVkoq\":[\"Leave empty for all codes\"],\"UmQ3/m\":[\"Deactivate Selected\"],\"Uwo8Xw\":[\"This image was shared with you by <0>\",[\"ownerProject\"],\"</0> on \",[\"sharedAt\"],\".\"],\"UztfYZ\":[\"Select port to associate\"],\"V/8B9A\":[\"I confirm that all existing versions will also be deleted\"],\"V/SINY\":[\"Update object\"],\"V1TzeS\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully deleted.\"],\"V66Jih\":[\"Access Status\"],\"V7fN5X\":[\"Copy object:\"],\"V804LY\":[\"Updating security group...\"],\"VCM3KS\":[\"Add Project Access\"],\"VKmlZ+\":[\"Containers to be emptied (\",[\"totalCount\"],\")\"],\"VLI9eO\":[\"Loading Floating IP Details...\"],\"VV1fdg\":[\"Any user has read access to objects. No token is required in the request.\"],\"VaA9mu\":[\"24 hours\"],\"VakxP/\":[\"Failed to Upload Object\"],\"Vg0k6h\":[\"Showing \",[\"filteredCount\"],\" of \",[\"totalCount\"],\" \",[\"itemName\"]],\"Vh/Uj5\":[\"Target path\"],\"Vj8XFg\":[\"Failed to Create Container\"],\"Vl4XTj\":[\"Folder name cannot contain slashes\"],\"Vmojta\":[\"Access status updated to \\\"\",[\"newStatus\"],\"\\\".\"],\"VoxR3s\":[\"Object was copied but could not be deleted from the source: \",[\"deleteErrorMessage\"]],\"Vz+7ZA\":[\"Could not create container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"Vzlopx\":[\"Delete container:\"],\"W0MCSG\":[\"Accept access to image\"],\"W5FkH9\":[\"Enter container name\"],\"W9PZE0\":[\"Objects Deleted\"],\"W9kfjU\":[\"QoS Policy ID\"],\"WCKEqI\":[\"This is a <0>static large object (SLO)</0> manifest. Metadata changes apply to the manifest only — segment objects are not affected.\"],\"WCLyHI\":[\"No Floating IPs found\"],\"WErCZy\":[\"Minimum RAM required to boot this image\"],\"WRZ3Mt\":[\"Loading container properties...\"],\"WYb0Td\":[\"<0>Are you sure?</0> All objects in the selected containers will be permanently deleted. This cannot be undone.\"],\"WYiUDa\":[\"Loading Containers...\"],\"Wbg1jv\":[\"Copy \",[\"text\"],\" to clipboard\"],\"Wca9WC\":[\"Failed to load security groups\"],\"WlpcJv\":[\"DNS Domain\"],\"WoSkGY\":[\"Remote IP\"],\"WrUky8\":[\"Share (Temporary URL)\"],\"WyKwnD\":[\"Store old object versions\"],\"WzVwU0\":[\"Target Project ID\"],\"X2OnDx\":[\"Ephemeral Disk (GiB)\"],\"X70LXS\":[[\"numberDeleted\"],\" objects were permanently deleted.\"],\"XLk16/\":[\"Together, we can unlock the true potential of your cloud infrastructure.\"],\"XYZLy9\":[\"Key contains invalid characters\"],\"XvjC4F\":[\"Saving...\"],\"XwxJJB\":[\"Container \\\"\",[\"containerName\"],\"\\\" was successfully created.\"],\"XxjLdW\":[[\"emptiedCount\",\"plural\",{\"one\":[\"#\",\" container was already empty.\"],\"other\":[\"#\",\" containers were already empty.\"]}]],\"Y+2SDm\":[\"Delete Security Group \\\"\",[\"securityGroupName\"],\"\\\"\"],\"Y1YKad\":[\"Edit Details\"],\"YIix5Y\":[\"Search...\"],\"YNgcgc\":[\"Loading Flavor Details...\"],\"YRexkb\":[\"Object Updated\"],\"YUU0QW\":[\"Flavor ID is required and cannot be empty.\"],\"YZmsaT\":[\"Partial Activation Success\"],\"YiMCKk\":[\"Show ACLs Preview\"],\"Yin3uB\":[\"Releasing Floating IP...\"],\"YjAOtb\":[\"Create Security Group\"],\"YrAy/S\":[\"You don't have permission to delete extra specs for this flavor.\"],\"YsOJlj\":[\"Server error occurred while fetching flavor access information. Please try again later.\"],\"YsrbQh\":[\"Owner Project ID\"],\"YuC9dj\":[\"Associate\"],\"YuGQWb\":[\"Rule Type\"],\"Z/eWPC\":[\"The object will be copied to this path. Navigate folders above to change the destination.\"],\"Z2fZGD\":[\"No project selected\"],\"Z3FXyt\":[\"Loading...\"],\"Z42tfY\":[\"Folders in object storage are virtual — they are created as zero-byte placeholder objects with a trailing slash. The folder will appear once created.\"],\"Z5r9vC\":[\"Partial Delete Success\"],\"Z8lGw6\":[\"Share\"],\"ZAx+d1\":[\"Max meta overall size\"],\"ZAy0zp\":[\"Successfully deactivated \",[\"successCount\"],\" of \",[\"totalCount\"],\" image(s)\"],\"ZUmOzn\":[\"Server returned unexpected data format for flavor details.\"],\"ZcWMT1\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been deactivated\"],\"Zgp2Sm\":[\"No projects available.\"],\"ZhVSpK\":[\"Failed to remove tenant access from flavor. Please try again.\"],\"Zq6Y5u\":[\"DNS name must be at most 63 characters.\"],\"ZvIpwi\":[\"Select a security group...\"],\"Zw49f9\":[\"folder-name\"],\"Zw8Q49\":[\"Security group not found\"],\"a/nTb8\":[\"Create Image\"],\"a12lSo\":[\"Port (to)\"],\"a13wDR\":[\"Type to search containers...\"],\"a3LDKx\":[\"Security\"],\"a4A2uB\":[\"Account listing limit\"],\"a4N/Bg\":[\"Load More\"],\"a7C4YS\":[\"Container Updated\"],\"a88X3d\":[\"<0>Are you sure?</0> Object <1>\\\"\",[\"displayName\"],\"\\\"</1> will be permanently deleted. This cannot be undone.\"],\"aG9OiI\":[\"Sharing Details\"],\"aI8Tgp\":[\"Owning Project ID\"],\"aL1w5Z\":[\"Used\"],\"aOeFR+\":[\"Empty Containers\"],\"aSsVD3\":[\"Public read access is not enabled. Before configuring static website serving, go to <0>Manage Access</0> and enable public read access.\"],\"aTqCTq\":[\"Image file is required\"],\"aV6KPH\":[\"Prevent accidental deletion\"],\"aiqFbS\":[\"<0>Are you sure?</0> The selected objects will be permanently deleted. This cannot be undone.\"],\"an5hVd\":[\"Images\"],\"ao/ZJi\":[\"Deleting folder and all its contents...\"],\"aqagJH\":[\"Unable to Update Image Visibility\"],\"arel2K\":[\"No objects found\"],\"azXlY+\":[\"Access Status:\"],\"b0uU1G\":[\"Store old object versions in container:\"],\"b2BLBa\":[\"Add Security Group Rule\"],\"bISG26\":[\"Failed to fetch flavor access information. Please try again.\"],\"bM1O3m\":[\"Image Instance\"],\"bQBMTH\":[\"This is a <0>static large object</0>. By default, all associated segment objects will also be permanently deleted.\"],\"bRgFkJ\":[\"Failed to upload file \\\"\",[\"fileName\"],\"\\\": \"],\"bYRFNi\":[\"Failed to Delete Objects\"],\"bc67JN\":[\"Custom Properties / Metadata\"],\"bmQLn5\":[\"Add Rule\"],\"bnql/K\":[\"Back to Images\"],\"boJ+Y1\":[\"Create Folder\"],\"boJlGf\":[\"Page Not Found\"],\"bpme7e\":[\"Flavor Not Found\"],\"bwRvnp\":[\"Action\"],\"bwhBhT\":[\"Security Group\"],\"byKna+\":[\"An unexpected error occurred. Please try again.\"],\"bzMKg7\":[\"Accepted\"],\"bzSI52\":[\"Discard\"],\"c+fUtV\":[\"Start typing to search for a container\"],\"c+xCSz\":[\"True\"],\"c1uL4p\":[\"The image \\\"\",[\"imageId\"],\"\\\" could not be deleted: \",[\"message\"]],\"c6b6fz\":[\"Delete Selected\"],\"cCfxH1\":[\"Downloading...\"],\"cJDQIO\":[\"Root Disk\"],\"cPKL6O\":[\"You don't have permission to create flavors in this project.\"],\"cWbW6w\":[\"Manage Access\"],\"cXuXkb\":[\"User \",[\"userId\"],\" from project \",[\"projectId\"]],\"chL5IG\":[\"Community\"],\"cj17eo\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been activated\"],\"cjEOmc\":[\"Share this security group with another project. The target project will be able to view and use this security group, but will not be able to modify or delete it.\"],\"cnGeoo\":[\"Delete\"],\"cpw++p\":[\"Static large object support\"],\"cqQyPB\":[\"Folder name\"],\"d+F6q9\":[\"Created\"],\"d+Ugpw\":[\"<0>Are you sure?</0> Folder <1>\\\"\",[\"folderDisplayName\"],\"\\\"</1> and all objects within it will be permanently deleted. This cannot be undone.\"],\"d/I0J3\":[\"Activate Selected\"],\"d0pLfy\":[\"Failed to delete security group\"],\"dEgA5A\":[\"Cancel\"],\"dFb5Nt\":[\"Id\"],\"dLFiER\":[\"Error Loading Containers: \",[\"errorMessage\"]],\"dOevLB\":[\"Expires in \",[\"selectedPresetLabel\"],\" — at \",[\"expiresAtFormatted\"]],\"dPBJAJ\":[\"Empty All (\",[\"selectedCount\"],\")\"],\"dPj4yB\":[\"Login to Your Account\"],\"dPoCVe\":[\"Type “detach” to confirm\"],\"dTNzBI\":[\"Key must contain at least one alphanumeric character\"],\"dVdc7N\":[\"You are not authorized to delete extra specs. Please log in again.\"],\"dd2ndz\":[\"Entries in ACLs are comma-separated. Examples:\"],\"diFNkW\":[\"Error loading component\"],\"e0NrBM\":[\"Project\"],\"eChIh7\":[\"Flavor \\\"\",[\"flavorName\"],\"\\\" has been successfully created.\"],\"eGEHJE\":[\"DNS Name\"],\"eKC+EC\":[\"-\"],\"ePK91l\":[\"Edit\"],\"eYlnXt\":[\"No images found\"],\"eh/k36\":[\"Select a rule type...\"],\"ekCRTP\":[\"Rejected\"],\"eks7oA\":[\"Port ID\"],\"eu70nA\":[\"Expires at (UTC)\"],\"eyRsaH\":[\"Root\"],\"ezT9KW\":[\"Container syncing\"],\"f+Uq1E\":[\"New object name\"],\"f0cwjH\":[\"Object properties\"],\"fCPhho\":[\"One or more objects could not be deleted:\"],\"fIvd7X\":[\"Failed to Delete Images\"],\"fJpv9x\":[\"Failed to Deactivate Images\"],\"fj5byd\":[\"N/A\"],\"fxnDd7\":[\"Failed to generate temporary URL: \",[\"generalError\"]],\"fzfAAa\":[\"Ingress\"],\"g+Jead\":[\"IPv6\"],\"g1IxCo\":[\"RAM must be an integer ≥ 128 MB.\"],\"g3BSCe\":[\"Swap Disk must be an integer ≥ 0.\"],\"g3UF2V\":[\"Accept\"],\"g8Yxlg\":[\"Temporary URL for \\\"\",[\"objectName\"],\"\\\" was copied to clipboard.\"],\"g9m7gK\":[\"ACL entries control who can read from or write to this container. Multiple entries are comma-separated. Changes take effect immediately after saving.\"],\"gFKJBP\":[\"Folder name cannot have leading or trailing whitespace\"],\"gGdfWx\":[\"The compute service is currently unavailable for this project. Please try again later.\"],\"gHTJc/\":[\"Object Storage\"],\"gMYsdZ\":[\"Set to \\\"Shared\\\"\"],\"gU7JFm\":[\"Creating security group rule...\"],\"gYe+hC\":[\"Click to upload\"],\"go0J2x\":[\"Failed to copy the temporary URL to the clipboard\"],\"grs4+e\":[\"Compute Overview\"],\"gztCjq\":[\"The tenant ID provided is invalid. Please check your input.\"],\"h3P8z+\":[\"Failed to delete the flavor. Please try again.\"],\"h47p9L\":[\"—\"],\"h8h6oz\":[\"Images Deactivated\"],\"h99+4y\":[\"Allocate Floating IP\"],\"hH3kDo\":[\"Loading Image Details...\"],\"hHL/wm\":[\"Image instance \\\"\",[\"imageName\"],\"\\\" has been updated\"],\"hLp49h\":[\"Type <0>\",[\"deleteWord\"],\"</0> to confirm:\"],\"hPz54a\":[\"Failed to Download\"],\"hQr1Cr\":[\"Deactivate Image\"],\"hXUWyd\":[\"Value is required\"],\"hYgDIe\":[\"Create\"],\"he3ygx\":[\"Copy\"],\"he4q+i\":[\"e.g., b90f9c4bc76140e18540b2cec1299e2a\"],\"hgpMHD\":[\"Total Size\"],\"hkjZ7P\":[\"Failed to update visibility for \\\"\",[\"imageName\"],\"\\\": \"],\"hrBow7\":[\"Network ID\"],\"i0qMbr\":[\"Home\"],\"i30J2U\":[\"No projects found\"],\"i41Xuw\":[\"Select a fixed IP address\"],\"i5MEDc\":[\"Failed to move object: \",[\"copyErrorMessage\"]],\"i6/ygf\":[\"A property with this key already exists\"],\"i9TIyi\":[\"Remote Security Group\"],\"i9qiyR\":[\"Expires in\"],\"iH8pgl\":[\"Back\"],\"igVDFt\":[\"Attach\"],\"iqUvrS\":[\"User C/D/I\"],\"izMhIO\":[\"User \",[\"userId\"],\" (any project)\"],\"j9hkgJ\":[\"rules\"],\"jBIkmi\":[\"QCOW2, Raw, VMDK, VHD, VHDX, VDI, AMI, ARI, AKI, ISO, PLOOP\"],\"jIPNJG\":[\"Basic Information\"],\"jK6wqe\":[\"Folder \\\"\",[\"folderName\"],\"\\\" and \",[\"deletedCount\"],\" objects were permanently deleted.\"],\"jKopCP\":[\"Network & Routing\"],\"jMc/mo\":[\"Server error occurred while creating extra specs. Please try again later.\"],\"jNzyQo\":[\"Object versioning\"],\"jPxavx\":[\"Failed to update security group\"],\"jS4B2+\":[\"Container name does not match\"],\"jSG7wx\":[\"Please enter a valid number of minutes greater than 0\"],\"jhU93c\":[\"The image \\\"\",[\"imageName\"],\"\\\" could not be created: \"],\"js24f6\":[\"Delete folder:\"],\"jtnAf8\":[\"Protected images (cannot be deleted)\"],\"jyqLKs\":[\"This member already has access to this image.\"],\"k0vAWv\":[\"Object count quota\"],\"k5nYwm\":[\"vCPU\"],\"k7ENJG\":[\"Preview \",[\"rowDisplayName\"]],\"k99j0U\":[\"Cancel upload\"],\"kA2lMP\":[\"External Network\"],\"kCLnJG\":[\"Empty All\"],\"kGmM/p\":[\"You don't have permission to create extra specs for this flavor.\"],\"kIuDMT\":[\"Configure the ingress and egress rules that control which traffic is allowed for this security group.\"],\"kKK8AH\":[\"Remaining Quota:\"],\"kQYfgO\":[\"One or more containers could not be emptied: \",[\"errorMessage\"]],\"kRfwU+\":[\"Nothing to do. Container is already empty.\"],\"kqJVBO\":[\"You don't have permission to share this security group.\"],\"kuYWaD\":[\"Reserved keys: web-index, web-listings, quota-count, quota-bytes\"],\"kzkYE6\":[\"Checking S3 credentials...\"],\"l75CjT\":[\"Yes\"],\"lAsm87\":[\"Protect this image from being deleted\"],\"lBVhQs\":[\"Something went wrong:\"],\"lN/Z9n\":[\"Security group name is required\"],\"lN3xvy\":[\"Delete Rule\"],\"lQ3EIe\":[\"Max deletes per request\"],\"lWTy+Y\":[\"Unable to Create Image\"],\"lWxDDh\":[\"Flavor Name\"],\"lZvIXd\":[\"Description must be at most 255 characters.\"],\"lhIa6x\":[\"Failed to load extra specs. Please try again.\"],\"lq/mBZ\":[\"Loading object info...\"],\"lw1412\":[\"You have been logged out due to inactivity.\"],\"lxentK\":[\"An unexpected error occurred\"],\"m16xKo\":[\"Add\"],\"m6X3ro\":[\"Group Name\"],\"mQSO1Y\":[\"Port Forwarding\"],\"mSLePW\":[\"You don't have permission to access flavors for this project.\"],\"mSfwLL\":[\"Project ID\"],\"mYnJeY\":[\"The text must match “release” in lowercase.\"],\"mpby9d\":[\"Contact support\"],\"mqljvE\":[\"Copy metadata\"],\"mvz5Eo\":[\"URL for public access\"],\"mxPfpY\":[\"Create new folder here\"],\"mzI/c+\":[\"Download\"],\"n0ZttO\":[\"Root Disk (GiB)\"],\"n1ekoW\":[\"Sign In\"],\"n1gB0L\":[\"Edit Floating IP \",[\"floating_ip_address\"]],\"n22YIM\":[\"Edit Description\"],\"n2IuBI\":[\"A temporary URL grants time-limited read access to this object without requiring authentication. Anyone with the link can download it until it expires.\"],\"n3eQzA\":[\"This property is reserved and cannot be modified\"],\"n46oLW\":[\"Failed to remove member\"],\"n9jJG6\":[\"Remove member access\"],\"nETBrc\":[\"Egress\"],\"nLvo6K\":[\"RBAC Policy Details:\"],\"nUuaq8\":[\"Failed to update container: \",[\"errorMessage\"]],\"nW/hX9\":[\"General Image Data\"],\"nZbdB+\":[\"Upload Cancelled\"],\"ne/GWZ\":[\"Inside a project, objects are stored in containers. Containers are where you define access permissions and quotas.\"],\"neiJm0\":[\"Flavors\"],\"nkpZyk\":[\"Container \\\"\",[\"containerName\"],\"\\\" was already empty.\"],\"nnxwBn\":[\"There are no rules for this security group\"],\"ntNlXu\":[\"Listing access\"],\"o/VDOG\":[\"Unable to Delete Image\"],\"o6M6l0\":[\"Failed to create security group\"],\"oDkgME\":[\"You are not authorized to create flavors. Please log in again.\"],\"oEGiW3\":[\"Uploading... \",[\"progressPct\"],\"%\"],\"ocUvR+\":[\"False\"],\"odVI9Y\":[\"Container Deleted\"],\"olfSYj\":[\"Access Control Updated\"],\"onHi/J\":[\"It will be removed from your image list.\"],\"p4nMut\":[\"Swap (MiB)\"],\"p6CSHM\":[\"Delete Objects\"],\"p7DzCB\":[\"Failed to Update Access Control\"],\"pFg+7w\":[\"Updated:\"],\"pOPvlj\":[\"Already active (will be skipped)\"],\"pU25+T\":[\"Upload of \\\"\",[\"objectName\"],\"\\\" was cancelled.\"],\"pbzA+s\":[\"Optional description\"],\"pebLmQ\":[\"Remove access for \",[\"memberIdDisplay\"]],\"plnnns\":[\"Deleted \",[\"successCount\"],\" image(s), but \",[\"failedCount\"],\" image(s) could not be deleted.\"],\"poCbZw\":[\"Loading ACLs...\"],\"podzPY\":[\"Project id\"],\"psPHye\":[\"Accept Shared Image\"],\"pubQie\":[\"Enter value\"],\"q0Rla3\":[\"Add Tenant Access\"],\"q44uUq\":[\"Containers Partially Emptied\"],\"q5sTNZ\":[\"<0>No Temp URL key configured.</0> A temporary URL key must be set at the account or container level before temporary URLs can be generated. Contact your administrator to configure <1>X-Account-Meta-Temp-URL-Key</1> or <2>X-Container-Meta-Temp-URL-Key</2>.\"],\"q6K46F\":[\"Key already exists\"],\"q88/6A\":[\"Failed to Create Folder\"],\"qAkkjP\":[\"Max object name length\"],\"qEDO1j\":[\"This is a <0>dynamic large object</0>. Only the manifest will be deleted — its segment objects (stored under the manifest prefix) are <1>not</1> automatically removed and must be deleted separately.\"],\"qFDA8L\":[\"Reject access to image\"],\"qJb6G2\":[\"Try Again\"],\"qQ1QBh\":[\"Hardware Specifications\"],\"qST5TS\":[\"Error - Image Details\"],\"qUlxA+\":[\"Folder \\\"\",[\"folderName\"],\"\\\" was permanently deleted.\"],\"qaAo9Y\":[\"Server error occurred while creating the flavor. Please try again later.\"],\"qh5W8q\":[\"Remove Policy\"],\"qs+BrU\":[\"You don't have permission to remove tenant access from this flavor.\"],\"qtoOYG\":[\"No limit\"],\"quU9wK\":[\"Failed to deactivate \",[\"failedCount\"],\" of \",[\"totalCount\"],\" image(s). Some images may already be deactivated or in an invalid state.\"],\"qvF2D8\":[\"There are no images available for this project with the current filters applied. Try adjusting your filter criteria or create a new image.\"],\"qxxo7y\":[\"No policies match your search\"],\"qyNaF7\":[\"Enter a timestamp like \\\"YYYY-MM-DD HH:mm:ss\\\" to schedule automatic deletion\"],\"qzIZOL\":[\"Invalid file format. Supported formats: \",[\"supportedFileFormats\"]],\"qzhUb9\":[\"Showing first \",[\"maxOptions\"],\" of \",[\"totalCount\"],\" — refine your search to narrow results\"],\"r5SQFW\":[\"Container name must be \",[\"maxContainerNameLength\"],\" characters or fewer\"],\"r9Aac8\":[\"Ephemeral Disk\"],\"rAtQcX\":[\"You can rename the object by changing the name here.\"],\"rD9yV1\":[\"Images to deactivate:\"],\"rIe0oV\":[\"Failed to add tenant access to flavor. Please try again.\"],\"rIi6x4\":[\"The flavor could not be found. It may have already been deleted.\"],\"rJe6vw\":[\"7 days\"],\"rbuO5A\":[\"This security group is already shared with the specified project.\"],\"rcBt6T\":[\"Failed to create credential: \",[\"errorMessage\"]],\"rdUucN\":[\"Preview\"],\"rhaNn7\":[\"Loading containers...\"],\"riR9oD\":[\"Note: for <0>static and dynamic large objects</0> only the manifests are deleted — their segments outside this folder prefix are not affected.\"],\"rlgAtt\":[\"The object will be moved to this path. Navigate folders above to change the destination.\"],\"rp0Bd0\":[\"Compute\"],\"rrjuul\":[\"For more details, have a look at the <0>documentation</0>.\"],\"rvXsSb\":[\"Tenant access for \\\"\",[\"tenantIdToRemove\"],\"\\\" has been removed successfully.\"],\"rwBVXS\":[\"Images to be deleted (\",[\"deletableCount\"],\")\"],\"ryf/ee\":[\"Images Activated\"],\"ryxYVo\":[\"Images to be deactivated (\",[\"activeCount\"],\")\"],\"s/s1lz\":[\"Any user can perform a HEAD or GET operation on the container provided the user also has read access on objects. No token is required.\"],\"s2ubkU\":[\"Flavor ID\"],\"s4Vnq2\":[\"Emptying...\"],\"sNVNmf\":[\"MAC Address\"],\"sPFHpI\":[\"Disk\"],\"sSNyf3\":[\"Welcome to <0>Aurora Dashboard</0>, your next-generation cloud management solution. We are dedicated to simplifying how you interact with and manage your cloud infrastructure. Designed with efficiency, scalability, and usability at its core, Aurora empowers you to streamline operations and unlock the full potential of your cloud resources.\"],\"sWBLli\":[\"Add Property\"],\"sXd+qS\":[\"Properties of \\\"\",[\"objectName\"],\"\\\" were successfully updated.\"],\"sa4CV6\":[\"All users from project \",[\"projectId\"]],\"shKIZu\":[\"Images to be activated (\",[\"deactivatedCount\"],\")\"],\"sheDTJ\":[\"Please note: for <0>dynamic</0> and <1>static large objects</1> only the manifests are deleted. The related segments are not deleted.\"],\"sihD20\":[\"Loading images...\"],\"sjMCOP\":[\"Last Modified\"],\"slWh5C\":[\"Associate Floating IP \",[\"floating_ip_address\"],\" with Port\"],\"sxbP3b\":[\"Object Count\"],\"t/YqKh\":[\"Remove\"],\"t0X9+8\":[\"Container Name\"],\"t1POAD\":[\"No custom metadata properties found. Click \\\"Add Property\\\" to create one.\"],\"t1fq6V\":[\"Server returned unexpected data format.\"],\"t7ff15\":[\"valid token required: false\"],\"t95VRV\":[\"About Aurora Dashboard\"],\"tASa/P\":[\"Server error occurred while deleting the flavor. Please try again later.\"],\"tIrNgH\":[\"Server error occurred while fetching extra specs. Please try again later.\"],\"tLerHy\":[\"Ephemeral Disk must be an integer ≥ 0.\"],\"tM5SEI\":[\"ACLs for container \\\"\",[\"containerName\"],\"\\\" were successfully updated.\"],\"tOkmLM\":[\"Failed to Copy Object\"],\"tV/Ozb\":[\"Port Range\"],\"tVSmFT\":[\"Loading more...\"],\"tX5yOZ\":[\"New Folder\"],\"tasfos\":[\"remove\"],\"tbwGSx\":[\"Minimum Disk (GB)\"],\"tejJLY\":[\"Associating Floating IP...\"],\"tfAKBU\":[\"Could not upload \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"tfDRzk\":[\"Save\"],\"tfxu04\":[\"Remove access for \",[\"tenantId\"]],\"thHAVL\":[\"Accepted Images\"],\"tiflqy\":[\"Unable to Re-activate Image\"],\"tlfxPP\":[\"Could not copy \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"tmpGvt\":[\"production, linux\"],\"u+VWhB\":[\"Copied to clipboard!\"],\"u2xIeO\":[\"Failed to update ACLs: \",[\"errorMessage\"]],\"u5HztT\":[\"RX/TX Factor\"],\"u77/s4\":[\"Floating IPs\"],\"u7En0V\":[\"Add Metadata\"],\"uAI0yI\":[\"Delete object:\"],\"uAQUqI\":[\"Status\"],\"uLtFAr\":[\"Could not update container \\\"\",[\"containerName\"],\"\\\": \",[\"errorMessage\"]],\"uSdnuQ\":[\"VCPUs must be an integer ≥ 1.\"],\"ujK/QN\":[\"Loading objects...\"],\"uly9ET\":[\"Rule Details:\"],\"up0ZSW\":[\"Fingerprint\"],\"uuKb0T\":[\"Description must be less than 65535 characters.\"],\"v0hPHE\":[\"Show Details\"],\"v3djpU\":[\"Move/Rename\"],\"v9Dn8m\":[\"Aurora Dashboard is more than just a tool—it's your partner in navigating the cloud. Whether you're a small startup or a global enterprise, Aurora provides the flexibility, power, and simplicity you need to achieve your goals.\"],\"vBUQNE\":[\"The extra spec could not be found. It may have already been deleted.\"],\"vEkTR9\":[\"Quota\"],\"vH2C/2\":[\"Swap\"],\"vLYu4Z\":[\"Nothing to do. Container appears empty — the object count may not have synced yet due to a recent operation.\"],\"vR4HmN\":[\"Loading Instances...\"],\"vTh35P\":[\"Create Container\"],\"vXmL4D\":[\"Drop your image file here\"],\"vZUKSz\":[\"Detach Floating IP \",[\"floating_ip_address\"]],\"vbajgL\":[\"Public Flavor\"],\"vcQSZh\":[\"This folder is empty — use New Folder to create one.\"],\"vcXmqy\":[\"Network Overview\"],\"vcvCXq\":[\"Error - Flavor Details\"],\"vg84cD\":[[\"allCount\"],\" items\"],\"vmRPFm\":[\"Share Security Group\"],\"vmYyLY\":[\"Remote IP Prefix\"],\"vp5vfW\":[\"1 hour\"],\"vpt8cE\":[\"Generate URL\"],\"vrPCbw\":[\"Image ID\"],\"w3bAcf\":[\"This action is permanent. The address will be removed from your project and returned to the public pool. This action cannot be undone.\"],\"w9+8d7\":[\"Remove tenant access\"],\"wEfZld\":[\"Create New Flavor\"],\"wFaT8w\":[\"Failed to Empty Containers\"],\"wMHvYH\":[\"Value\"],\"wPrtGF\":[\"Enter key\"],\"wTg+FY\":[\"Max file size\"],\"wXxPjv\":[\"S3 Object Storage — Setup Required\"],\"wa1Bcq\":[\"Enter tenant ID\"],\"wbqM4L\":[[\"customMinutes\"],\" minutes\"],\"wcUecy\":[\"You don't have permission to view extra specs for this flavor.\"],\"we28Pq\":[\"Hide ACLs Preview\"],\"wlQNTg\":[\"Members\"],\"wlUDbB\":[\"Last updated: \",[\"formattedDate\"]],\"wrXcuy\":[\"Object Name\"],\"wrk/xj\":[\"Image Details\"],\"wyIOMP\":[\"Image name is required\"],\"wzqqS+\":[\"Key Features\"],\"x/XQrD\":[\"Any file type\"],\"x1bK0h\":[\"There are no containers available with the current search criteria. Try adjusting your search term.\"],\"x3T4pq\":[\"The container metadata reports objects but none were listed. This may be a temporary synchronization delay — please wait a moment and try again.\"],\"x5l/TK\":[\"Already active (will be skipped):\"],\"x9AdZ8\":[\"property_key\"],\"xNG/3n\":[\"Floating IP Address\"],\"xNZKYy\":[\"Failed to delete \",[\"failedCount\"],\" of \",[\"totalCount\"],\" image(s). Some images may be protected or in use.\"],\"xqhyRT\":[\"Object Uploaded\"],\"xw2UtT\":[\"Create New Image\"],\"y+KBOY\":[\"e.g., production, linux, ubuntu\"],\"y02Bu1\":[\"Container:\"],\"y0u86k\":[\"The requested flavor could not be found. It may have been deleted or you may not have access to it.\"],\"y1GYnY\":[\"Could not move \\\"\",[\"objectName\"],\"\\\": \",[\"errorMessage\"]],\"yPWFWy\":[\"ICMP Type\"],\"yTtJTy\":[\"Edit Image Metadata\"],\"yYxB17\":[\"Clear all\"],\"ylfbpz\":[\"Extra spec key is required and cannot be empty.\"],\"yp0UjB\":[\"Ethertype\"],\"yqPflB\":[\"... and \",[\"hiddenCount\"],\" more\"],\"yu9G3x\":[\"Edit Security Group\"],\"yz7wBu\":[\"Close\"],\"z+zpLP\":[\"valid token required: true\"],\"z1JceR\":[\"Back to Floating IPs\"],\"z45o5B\":[\"Object count\"],\"z9NAjZ\":[\"Object Deleted\"],\"zCD96i\":[\"You are not authorized to view flavor details. Please log in again.\"],\"zDS0JC\":[\"Name must be 2-50 characters long.\"],\"zWb/Nn\":[\"Max header size\"],\"zc5dcw\":[\"Login failed. Please check your credentials and try again.\"],\"zga9sT\":[\"OK\"],\"zhM8FP\":[\"Grant access to a user from a different project.\"],\"zm7+/D\":[\"You are about to deactivate <0>\",[\"activeCount\"],\" image(s)</0>. Deactivated images cannot be used to launch new instances.\"],\"zwBp5t\":[\"Private\"]}")as Messages;
